### 2024-02-12 #6345 @editorialbot
[PRE REVIEW]: MIRP: A Python package for standardised radiomics
**Submitting author:** @alexzwanenburg (<a href="http://orcid.org/0000-0002-0342-9545">Alex Zwanenburg</a>)
**Repository:** https://github.com/oncoray/mirp
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.1.0
