### 2024-02-14 #6345 @alexzwanenburg
R and Tex labels can be removed from this issue. The software is fully in Python.
