### 2024-02-16 #6345 @emdupre
👋 Hi @alexzwanenburg, and thanks for your submission to JOSS !

I see that you have suggested several potential reviewers, thank you !

I'll put together a list of folks to reach out to as reviewers based on these suggestions and my own recommendations. I'll update this thread with that outreach.
