### 2024-02-16 #6345 @emdupre
👋  Hi @Matthew-Jennings, @surajpaib, @andreped,

Would you be willing to review _MIRP: A Python package for standardised radiomics_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/oncoray/mirp) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.06345/joss.06345/10.21105.joss.06345.pdf

If you're available to review this work within the next month—or six weeks, at most—please let me know and I'll add you as a reviewer.

Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
