### 2024-02-24 #6345 @theanega
Hi,

Happy to review too!
Thanks,
Olivia

On Fri, 23 Feb 2024 at 18:37, Elizabeth DuPre ***@***.***>
wrote:

> 👋 Hi @drcandacemakedamoore <https://github.com/drcandacemakedamoore>,
> @theanega <https://github.com/theanega>, @Matthew-Jennings
> <https://github.com/Matthew-Jennings>,
>
> Would you be willing to review *MIRP: A Python package for standardised
> radiomics* for JOSS (the Journal of Open Source Software)?
>
> You can learn more about reviewing for JOSS -- including our conflict of
> interest policy -- here.
>
> The review takes place on GitHub and focuses on the software and this
> short paper:
>
> https://github.com/openjournals/joss-papers/blob/joss.06345/joss.06345/10.21105.joss.06345.pdf
>
> If you're available to review this work within the next month—or six
> weeks, at most—please let me know and I'll add you as a reviewer.
>
> Once I have sufficient reviewers I'll open a dedicated review issue.
>
> Thank you for considering !
>
