### 2024-02-26 #6413 @editorialbot
[REVIEW]: MIRP: A Python package for standardised radiomics
**Submitting author:** @alexzwanenburg (<a href="http://orcid.org/0000-0002-0342-9545">Alex Zwanenburg</a>)
**Repository:** https://github.com/oncoray/mirp
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.2.4

@surajpaib & @Matthew-Jennings & @drcandacemakedamoore & @theanega, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
