### 2024-02-29 8770d2f
Create CONTRIBUTING.md
Contributing notes.
