### 2024-03-07 3e6527c
Early stopping for checks on orientation and positioning of RT structure sets.
