### 2024-03-07 f3c4aae
SOP instance UID for stacks of DICOM files are now correctly collected.
