### 2024-03-07 e328841
Added tests for reading images and masks from xml files.
