+++ test/read_image_and_mask_test.py
@@ -377,3 +377,36 @@ def test_read_generic_image_and_mask_modality_specific():
     assert len(roi_list) == 1
     assert all(isinstance(roi, BaseMask) for roi in roi_list)
     assert roi_list[0].roi_name == "region_1"
+
+
+def test_read_dicom_image_and_mask_data_xml():
+    # Read the data settings xml file, and update path to image and mask.
+    from xml.etree import ElementTree as ElemTree
+
+    # Remove temporary data xml file if it exists.
+    if os.path.exists(os.path.join(CURRENT_DIR, "data", "configuration_files", "temp_test_config_data.xml")):
+        os.remove(os.path.join(CURRENT_DIR, "data", "configuration_files", "temp_test_config_data.xml"))
+
+    # Load xml.
+    tree = ElemTree.parse(os.path.join(CURRENT_DIR, "data", "configuration_files", "test_config_data.xml"))
+    paths_branch = tree.getroot()
+
+    # Update paths in xml file.
+    for image in paths_branch.iter("image"):
+        image.text = str(os.path.join(CURRENT_DIR, "data", "sts_images"))
+    for mask in paths_branch.iter("mask"):
+        mask.text = str(os.path.join(CURRENT_DIR, "data", "sts_images"))
+
+    # Save as temporary xml file.
+    tree.write(os.path.join(CURRENT_DIR, "data", "configuration_files", "temp_test_config_data.xml"))
+
+    image_list = import_image_and_mask(
+        image=os.path.join(CURRENT_DIR, "data", "configuration_files", "temp_test_config_data.xml")
+    )
+
+    image, roi_list = read_image_and_masks(image=image_list[0])
+    assert isinstance(image, PETImage)
+    assert len(roi_list) == 1
+    assert all(isinstance(roi, BaseMask) for roi in roi_list)
+    assert roi_list[0].roi_name == "GTV_Mass_PET"
+
