+++ mirp/images/ctImage.py
@@ -1,4 +1,4 @@
-from typing import Optional, Tuple, Any
+from typing import Any
 import numpy as np
 from mirp.images.genericImage import GenericImage
 
@@ -21,10 +21,10 @@ class CTImage(GenericImage):
 
     def normalise_intensities(
             self,
-            normalisation_method: Optional[str] = "none",
-            intensity_range: Optional[Tuple[Any, Any]] = None,
-            saturation_range: Optional[Tuple[Any, Any]] = None,
-            mask: Optional[np.ndarray] = None
+            normalisation_method: None | str = "none",
+            intensity_range: None | tuple[Any, Any] = None,
+            saturation_range: None | tuple[Any, Any] = None,
+            mask: None | np.ndarray = None
     ):
         """
         Normalise intensities. NOTE: this changes the class of the object from CTImage to GenericImage as
