+++ mirp/images/rtdoseImage.py
@@ -1,5 +1,5 @@
 import numpy as np
-from typing import Optional, Tuple, Any
+from typing import Any
 from mirp.images.genericImage import GenericImage
 
 
@@ -12,23 +12,16 @@ class RTDoseImage(GenericImage):
     def get_default_lowest_intensity():
         return 0.0
 
-    def update_image_data(self):
-        if self.image_data is None:
-            return
-
-        # Ensure that dose values are not negative.
-        self.image_data[self.image_data < 0.0] = 0.0
-
     def normalise_intensities(
             self,
-            normalisation_method: Optional[str] = "none",
-            intensity_range: Optional[Tuple[Any, Any]] = None,
-            saturation_range: Optional[Tuple[Any, Any]] = None,
-            mask: Optional[np.ndarray] = None
+            normalisation_method: None | str = "none",
+            intensity_range: None | tuple[Any, Any] = None,
+            saturation_range: None | tuple[Any, Any] = None,
+            mask: None | np.ndarray = None
     ):
         """
-        Normalise intensities. NOTE: this changes the class of the object from RTDose to GenericImage as
-        normalisation breaks the one-to-one relationship between intensities and dose.
+        Normalise intensities. NOTE: this changes the class of the object from RTDoseImage to GenericImage as
+        normalisation breaks the one-to-one relationship between intensities and Hounsfield units.
         """
         image = super().normalise_intensities(
             normalisation_method=normalisation_method,
