### 2024-03-07 281d7c6
Check for early stopping.
