### 2024-03-07 b530af2
filter kernel names are now correctly parsed as strings
