### 2024-03-08 229ed73
Added unit test for incorrectly specified image and mask imports.
