### 2024-03-10 oncoray/mirp#54 @surajpaib
Support Python versions below 3.11
It would be useful to support highly used Python versions so that the end-user does not need to create a separate environment to try out MIRP.

PyPI stats indicate that most packages are downloaded for 3.8/3.9/3.10: https://pypistats.org/packages/__all__
However, 3.8 goes EOL very soon, but 3.9 and 3.10 have quite some time before EOL (1.5yrs and 2.5yrs) https://devguide.python.org/versions/
