### 2024-03-10 oncoray/mirp#55 @surajpaib
Software Paper Comments Thread
In this thread, please find comments on the software paper submitted to JOSS for this package.
Reference: https://github.com/openjournals/joss-reviews/issues/6413

Comments:
1. Could you reference the packages compared against with DOIs in addition to the Github links?
2. As a mention of user friendly API is made in the statement of need section, it would be helpful to list how other non-GUI packages provide their interfaces (high-level/low-level API, docker, etc)
