### 2024-03-11 oncoray/mirp#54 @alexzwanenburg
Thanks for suggestion. The supported Python versions are currently constrained by functionality in the `typing` library. Notably, MIRP makes use of the following, newer features:

- `Self` type introduced with [PEP 673](https://peps.python.org/pep-0673/) in 3.11.
- Short union types introduced with [PEP 604](https://peps.python.org/pep-0604/) in 3.10.
- Type hinting generics introduced with [PEP 585](https://peps.python.org/pep-0585/) in 3.9.

The use of `Self` could be made backward compatible by (conditionally) importing `Self` from [`typing_extensions`](https://typing-extensions.readthedocs.io).

After some searching, I don't think there is a suitable way to make PEP 604 and 585 backward compatible with older Python versions. `from __future__ import annotations` ([PEP 563](https://peps.python.org/pep-0563/) introduced in 3.7) does not seem to work, according to this [thread](https://stackoverflow.com/questions/62871524/is-pep-585-unusable-at-runtime-under-python-3-7-and-3-8), and using it may cause issues in upcoming Python versions, see [PEP 649](https://peps.python.org/pep-0649/).

Realistically -- that is without forgoing `typing` entirely -- MIRP can be updated to support Python version 3.10 as the earliest version.
