+++ test/import_image_and_mask_test.py
@@ -1,3 +1,4 @@
+import copy
 import os.path
 import pytest
 
@@ -676,6 +677,7 @@ def test_failure_multiple_image_and_mask_import():
     assert "not contain any supported mask files" in str(exception_info.value)
     assert "that contain the name pattern (false_mask)" in str(exception_info.value)
 
+
 def test_failure_multiple_image_and_mask_import_data_xml():
     # Read the data settings xml file, and update path to image and mask.
     from xml.etree import ElementTree as ElemTree
@@ -695,72 +697,77 @@ def test_failure_multiple_image_and_mask_import_data_xml():
     paths_branch = tree.getroot()
 
     # Set basic data in xml file.
-    for image in paths_branch.iter("image"):
-        image.text = str(os.path.join(CURRENT_DIR, "data", "sts_images"))
-    for image_sub_folder in paths_branch.iter("image_sub_folder"):
-        image_sub_folder.text = str(r"CT/dicom/image")
-    for mask in paths_branch.iter("mask"):
-        mask.text = str(os.path.join(CURRENT_DIR, "data", "sts_images"))
-    for mask_sub_folder in paths_branch.iter("mask_sub_folder"):
-        mask_sub_folder.text = str(r"CT/dicom/mask")
-
-    # Save as temporary xml file.
-    tree.write(target_file)
-
+    for element in paths_branch.iter("image"):
+        element.text = str(os.path.join(CURRENT_DIR, "data", "sts_images"))
+    for element in paths_branch.iter("image_sub_folder"):
+        element.text = str(r"CT/dicom/image")
+    for element in paths_branch.iter("mask"):
+        element.text = str(os.path.join(CURRENT_DIR, "data", "sts_images"))
+    for element in paths_branch.iter("mask_sub_folder"):
+        element.text = str(r"CT/dicom/mask")
 
     # DICOM stack and masks for all samples, but with incorrect instructions.
     # No matching file type.
+    false_file_type_tree = copy.deepcopy(tree)
+    for element in false_file_type_tree.getroot().iter("image_file_type"):
+        element.text = "nifti"
+    false_file_type_tree.write(target_file)
+
     with pytest.raises(ValueError) as exception_info:
         image_list = import_image_and_mask(
-            image=os.path.join(CURRENT_DIR, "data", "sts_images"),
-            image_file_type="nifti",
-            image_sub_folder=os.path.join("CT", "dicom", "image"),
-            mask_sub_folder=os.path.join("CT", "dicom", "mask")
+            image=target_file
         )
     assert "did not contain any supported image files" in str(exception_info.value)
 
     # DICOM stack and masks for all samples, but with incorrect instructions.
     # No matching image modality.
+    false_image_modality_tree = copy.deepcopy(tree)
+    for element in false_image_modality_tree.getroot().iter("image_modality"):
+        element.text = "pet"
+    false_image_modality_tree.write(target_file)
+
     with pytest.raises(ValueError) as exception_info:
         image_list = import_image_and_mask(
-            image=os.path.join(CURRENT_DIR, "data", "sts_images"),
-            image_modality="pet",
-            image_sub_folder=os.path.join("CT", "dicom", "image"),
-            mask_sub_folder=os.path.join("CT", "dicom", "mask")
+            image=target_file
         )
     assert "No images were found" in str(exception_info.value)
 
     # DICOM stack and masks for all samples, but with incorrect instructions.
     # No matching mask modality.
+    false_mask_modality_tree = copy.deepcopy(tree)
+    for element in false_mask_modality_tree.getroot().iter("mask_modality"):
+        element.text = "seg"
+    false_mask_modality_tree.write(target_file)
+
     with pytest.raises(ValueError) as exception_info:
         image_list = import_image_and_mask(
[…]
