+++ mirp/settings/settingsImageTransformation.py
@@ -1,5 +1,5 @@
 import copy
-from typing import Union, List, Any
+from typing import Any
 from dataclasses import dataclass
 
 import numpy as np
@@ -331,49 +331,49 @@ class ImageTransformationSettingsClass:
             self,
             by_slice: bool,
             response_map_feature_settings: FeatureExtractionSettingsClass | None = None,
-            response_map_feature_families: Union[None, str, List[str]] = "statistical",
-            response_map_discretisation_method: Union[None, str, List[str]] = "fixed_bin_number",
-            response_map_discretisation_n_bins: Union[None, int, List[int]] = 16,
-            response_map_discretisation_bin_width: Union[None, int, List[int]] = None,
-            filter_kernels: Union[None, str, List[str]] = None,
-            boundary_condition: Union[None, str] = "mirror",
-            separable_wavelet_families: Union[None, str, List[str]] = None,
-            separable_wavelet_set: Union[None, str, List[str]] = None,
+            response_map_feature_families: None | str | list[str] = "statistical",
+            response_map_discretisation_method: None | str | list[str] = "fixed_bin_number",
+            response_map_discretisation_n_bins: None | int | list[int] = 16,
+            response_map_discretisation_bin_width: None | int | list[float] = None,
+            filter_kernels: None | str | list[str] = None,
+            boundary_condition: None | str = "mirror",
+            separable_wavelet_families: None | str | list[str] = None,
+            separable_wavelet_set: None | str | list[str] = None,
             separable_wavelet_stationary: bool = True,
-            separable_wavelet_decomposition_level: Union[None, int, List[int]] = 1,
+            separable_wavelet_decomposition_level: None | int | list[int] = 1,
             separable_wavelet_rotation_invariance: bool = True,
             separable_wavelet_pooling_method: str = "max",
-            separable_wavelet_boundary_condition: Union[None, str] = None,
-            nonseparable_wavelet_families: Union[None, str, List[str]] = None,
-            nonseparable_wavelet_decomposition_level: Union[None, int, List[int]] = 1,
-            nonseparable_wavelet_response: Union[None, str] = "real",
-            nonseparable_wavelet_boundary_condition: Union[None, str] = None,
-            gaussian_sigma: Union[None, float, List[float]] = None,
-            gaussian_kernel_truncate: Union[None, float] = 4.0,
-            gaussian_kernel_boundary_condition: Union[None, str] = None,
-            laplacian_of_gaussian_sigma: Union[None, float, List[float]] = None,
-            laplacian_of_gaussian_kernel_truncate: Union[None, float] = 4.0,
+            separable_wavelet_boundary_condition: None | str = None,
+            nonseparable_wavelet_families: None | str | list[str] = None,
+            nonseparable_wavelet_decomposition_level: None | int | list[int] = 1,
+            nonseparable_wavelet_response: None | str = "real",
+            nonseparable_wavelet_boundary_condition: None | str = None,
+            gaussian_sigma: None | float | list[float] = None,
+            gaussian_kernel_truncate: None | float = 4.0,
+            gaussian_kernel_boundary_condition: None | str = None,
+            laplacian_of_gaussian_sigma: None | float | list[float] = None,
+            laplacian_of_gaussian_kernel_truncate: None | float = 4.0,
             laplacian_of_gaussian_pooling_method: str = "none",
-            laplacian_of_gaussian_boundary_condition: Union[None, str] = None,
-            laws_kernel: Union[None, str, List[str]] = None,
-            laws_delta: Union[int, List[int]] = 7,
[…]
