+++ mirp/settings/settingsFeatureExtraction.py
@@ -1,4 +1,4 @@
-from typing import Union, List, Any
+from typing import Any
 from dataclasses import dataclass
 from mirp.settings.utilities import setting_def
 
@@ -194,22 +194,22 @@ class FeatureExtractionSettingsClass:
             by_slice: bool = False,
             no_approximation: bool = False,
             ibsi_compliant: bool = True,
-            base_feature_families: Union[None, str, List[str]] = "none",
-            base_discretisation_method: Union[None, str, List[str]] = None,
-            base_discretisation_n_bins: Union[None, int, List[int]] = None,
-            base_discretisation_bin_width: Union[None, float, List[float]] = None,
+            base_feature_families: None | str | list[str] = "none",
+            base_discretisation_method: None | str | list[str] = None,
+            base_discretisation_n_bins: None | int | list[int] = None,
+            base_discretisation_bin_width: None | float | list[float] = None,
             ivh_discretisation_method: str = "none",
-            ivh_discretisation_n_bins: Union[None, int] = 1000,
-            ivh_discretisation_bin_width: Union[None, float] = None,
-            glcm_distance: Union[float, List[float]] = 1.0,
-            glcm_spatial_method: Union[None, str, List[str]] = None,
-            glrlm_spatial_method: Union[None, str, List[str]] = None,
-            glszm_spatial_method: Union[None, str, List[str]] = None,
-            gldzm_spatial_method: Union[None, str, List[str]] = None,
-            ngtdm_spatial_method: Union[None, str, List[str]] = None,
-            ngldm_distance: Union[float, List[float]] = 1.0,
-            ngldm_difference_level: Union[float, List[float]] = 0.0,
-            ngldm_spatial_method: Union[None, str, List[str]] = None,
+            ivh_discretisation_n_bins: None | int = 1000,
+            ivh_discretisation_bin_width: None | float = None,
+            glcm_distance: float | list[float] = 1.0,
+            glcm_spatial_method: None | str | list[str] = None,
+            glrlm_spatial_method: None | str | list[str] = None,
+            glszm_spatial_method: None | str | list[str] = None,
+            gldzm_spatial_method: None | str | list[str] = None,
+            ngtdm_spatial_method: None | str | list[str] = None,
+            ngldm_distance: float | list[float] = 1.0,
+            ngldm_difference_level: float | list[float] = 0.0,
+            ngldm_spatial_method: None | str | list[str] = None,
             **kwargs
     ):
         # Set by slice.
@@ -229,7 +229,7 @@ class FeatureExtractionSettingsClass:
             base_feature_families = [base_feature_families]
 
         # Check which entries are valid.
-        valid_families: List[bool] = [ii in self.get_available_families() for ii in base_feature_families]
+        valid_families: list[bool] = [ii in self.get_available_families() for ii in base_feature_families]
 
         if not all(valid_families):
             raise ValueError(
@@ -237,7 +237,7 @@ class FeatureExtractionSettingsClass:
                 f"{', '.join([base_feature_families[ii] for ii, is_valid in enumerate(valid_families) if not is_valid])}")
 
         # Set families.
-        self.families: List[str] = base_feature_families
+        self.families: list[str] = base_feature_families
 
         if not self.has_any_feature_family():
             self.families = ["none"]
@@ -308,9 +308,9 @@ class FeatureExtractionSettingsClass:
             base_discretisation_bin_width = None
 
         # Set discretisation method-related parameters.
[…]

[…]
