### 2024-03-12 68648dd
Fixed deprecated style for code hinting.
