### 2024-03-12 ab40f06
Added details to paper.
