+++ paper/paper.bib
@@ -253,3 +253,128 @@
   doi      = "10.1038/s41467-023-44591-3",
   pmc      = "PMC10770385"
 }
+
+
+@ARTICLE{Van_Griethuysen2017-qp,
+  title    = "Computational Radiomics System to Decode the Radiographic
+              Phenotype",
+  author   = "van Griethuysen, Joost J M and Fedorov, Andriy and Parmar,
+              Chintan and Hosny, Ahmed and Aucoin, Nicole and Narayan, Vivek
+              and Beets-Tan, Regina G H and Fillion-Robin, Jean-Christophe and
+              Pieper, Steve and Aerts, Hugo J W L",
+  journal  = "Cancer Res.",
+  volume   =  77,
+  number   =  21,
+  pages    = "e104--e107",
+  month    =  nov,
+  year     =  2017,
+  language = "en",
+  issn     = "0008-5472, 1538-7445",
+  pmid     = "29092951",
+  doi      = "10.1158/0008-5472.CAN-17-0339",
+  pmc      = "PMC5672828"
+}
+
+
+
+@ARTICLE{Apte2018-sf,
+  title    = "Technical Note: Extension of {CERR} for computational radiomics:
+              A comprehensive {MATLAB} platform for reproducible radiomics
+              research",
+  author   = "Apte, Aditya P and Iyer, Aditi and Crispin-Ortuzar, Mireia and
+              Pandya, Rutu and van Dijk, Lisanne V and Spezi, Emiliano and
+              Thor, Maria and Um, Hyemin and Veeraraghavan, Harini and Oh, Jung
+              Hun and Shukla-Dave, Amita and Deasy, Joseph O",
+  journal  = "Med. Phys.",
+  volume   =  45,
+  number   =  8,
+  pages    = "3713--3720",
+  month    =  aug,
+  year     =  2018,
+  language = "en",
+  issn     = "0094-2405, 2473-4209",
+  pmid     = "29896896",
+  doi      = "10.1002/mp.13046"
+}
+
+
+@ARTICLE{Nioche2018-jr,
+  title    = "{LIFEx}: A Freeware for Radiomic Feature Calculation in
+              Multimodality Imaging to Accelerate Advances in the
+              Characterization of Tumor Heterogeneity",
+  author   = "Nioche, Christophe and Orlhac, Fanny and Boughdad, Sarah and
+              Reuz{\'e}, Sylvain and Goya-Outi, Jessica and Robert, Charlotte
+              and Pellot-Barakat, Claire and Soussan, Michael and Frouin,
+              Fr{\'e}d{\'e}rique and Buvat, Ir{\`e}ne",
+  journal  = "Cancer Res.",
+  volume   =  78,
+  number   =  16,
+  pages    = "4786--4789",
+  month    =  aug,
+  year     =  2018,
+  language = "en",
+  issn     = "0008-5472, 1538-7445",
+  pmid     = "29959149",
+  doi      = "10.1158/0008-5472.CAN-18-0125"
+}
+
+
+@ARTICLE{Vallieres2015-ui,
+  title    = "A radiomics model from joint {FDG-PET} and {MRI} texture features
+              for the prediction of lung metastases in soft-tissue sarcomas of
+              the extremities",
+  author   = "Valli{\`e}res, M and Freeman, C R and Skamene, S R and El Naqa, I",
+  journal  = "Phys. Med. Biol.",
+  volume   =  60,
+  number   =  14,
+  pages    = "5471--5496",
+  month    =  jul,
+  year     =  2015,
+  language = "en",
+  issn     = "0031-9155, 1361-6560",
+  pmid     = "26119045",
[…]
