+++ paper/paper.bib
@@ -340,7 +340,7 @@
 @ARTICLE{Vallieres2017-ye,
   title    = "Radiomics strategies for risk assessment of tumour failure in
               head-and-neck cancer",
-  author   = "Valli{\`e}res, Martin and Kay-Rivest, Emily and Perrin, L{\'e}o
+  author   = "Valli{\`e}res, M and Kay-Rivest, Emily and Perrin, L{\'e}o
               Jean and Liem, Xavier and Furstoss, Christophe and Aerts, Hugo J
               W L and Khaouam, Nader and Nguyen-Tan, Phuc Felix and Wang,
               Chang-Shu and Sultanem, Khalil and Seuntjens, Jan and El Naqa,
