### 2024-03-12 oncoray/mirp#55 @alexzwanenburg
I added references to the publications (with DOI) related to the publications. I also refined the description of the interface in Table 1.
