### 2024-03-13 oncoray/mirp#54 @alexzwanenburg
For reference, also see Scientific Python [SPEC 0](https://scientific-python.org/specs/spec-0000/), which recommends time-based policies for dropping support for dependencies. In case of python versions, their recommendation is 3 years. This recommendation provides for deprecating support for 3.10 late 2024, with earlier versions already having been deprecated.
