+++ pyproject.toml
@@ -13,6 +13,7 @@ license = {file = "LICENSE.txt"}
 readme = "README.md"
 classifiers = [
     "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
@@ -21,7 +22,7 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Image Processing",
     "Topic :: Scientific/Engineering :: Medical Science Apps."
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.10"
 dependencies = [
     "itk>=5.3.0",
     "matplotlib>=3.7.0",
@@ -31,7 +32,8 @@ dependencies = [
     "pywavelets>=1.4.0",
     "ray>=2.7.0",
     "scikit-image>=0.20.0",
-    "scipy>=1.11"
+    "scipy>=1.11",
+    "typing-extensions>=4.10; python<3.11"
 ]
 
 [project.urls]
@@ -41,7 +43,7 @@ Changelog = "https://github.com/oncoray/mirp/blob/master/NEWS.md"
 
 [project.optional-dependencies]
 test = ["pytest>=7.4.0"]
-docs = ["sphinx>=5.0.0", "sphinx_rtd_theme>=1.3.0"]
+docs = ["sphinx>=5.0.0", "sphinx_rtd_theme>=2.0.0"]
 
 [tool.setuptools.packages.find]
 where = ["."]
