### 2024-03-13 oncoray/mirp#54 @alexzwanenburg
Added support for Python 3.10 in version 2.1.2.
