### 2024-03-14 c15861c
Added intensity_scaling parameter.
