### 2024-03-14 afe00c3
WIP on implementing intensity scaling.
