### 2024-03-15 c558216
Update NEWS.md
