### 2024-03-15 310af89
Implemented scale_intensities method.
