+++ mirp/images/petImage.py
@@ -51,3 +51,20 @@ class PETImage(GenericImage):
         new_image.update_from_template(template=image)
 
         return new_image
+
+    def scale_intensities(self, scale: float):
+
+        image = super().scale_intensities(scale=scale)
+
+        if image.image_data is None:
+            return self
+
+        if scale == 1.0:
+            return self
+
+        # Scaling intensities changes the object class from PETImage to GenericImage as this breaks the one-to-one
+        # relationship between intensities and SUV values or PET counts.
+        new_image = GenericImage(image_data=image.image_data)
+        new_image.update_from_template(template=image)
+
+        return new_image
