+++ mirp/images/rtdoseImage.py
@@ -40,3 +40,20 @@ class RTDoseImage(GenericImage):
         new_image.update_from_template(template=image)
 
         return new_image
+
+    def scale_intensities(self, scale: float):
+
+        image = super().scale_intensities(scale=scale)
+
+        if image.image_data is None:
+            return self
+
+        if scale == 1.0:
+            return self
+
+        # Scaling intensities changes the object class from RTDoseImage to GenericImage as this breaks the one-to-one
+        # relationship between intensities and dose.
+        new_image = GenericImage(image_data=image.image_data)
+        new_image.update_from_template(template=image)
+
+        return new_image
