### 2024-03-15 c6d5957
Add unit test to assess intensity scaling.
