### 2024-03-15 23280d4
Added square transformation filter.
