+++ mirp/settings/settingsImageTransformation.py
@@ -768,7 +768,8 @@ class ImageTransformationSettingsClass:
             "separable_wavelet", "nonseparable_wavelet", "riesz_nonseparable_wavelet",
             "riesz_steered_nonseparable_wavelet", "gaussian", "riesz_gaussian", "riesz_steered_gaussian",
             "laplacian_of_gaussian", "log", "riesz_laplacian_of_gaussian", "riesz_steered_laplacian_of_gaussian",
-            "riesz_log", "riesz_steered_log", "laws", "gabor", "riesz_gabor", "riesz_steered_gabor", "mean"
+            "riesz_log", "riesz_steered_log", "laws", "gabor", "riesz_gabor", "riesz_steered_gabor", "mean",
+            "pyradiomics_square"
         ]
 
     def check_boundary_condition(self, x, var_name):
@@ -1105,6 +1106,14 @@ class ImageTransformationSettingsClass:
             x = [x]
 
         return x is not None and any(filter_kernel.startswith("riesz_steered") for filter_kernel in x)
+
+    def has_square_transform_filter(self, x=None):
+        if x is None:
+            x = self.spatial_filters
+        elif not isinstance(x, list):
+            x = [x]
+
+        return x is not None and any(filter_kernel == "pyradiomics_square" for filter_kernel in x)
 
 
 def get_image_transformation_settings() -> list[dict[str, Any]]:
