### 2024-03-15 440a731
Add safeguard against alpha == 0.0.
