### 2024-03-15 4e79a60
Added square root transformation function.
