+++ mirp/workflows/standardWorkflow.py
@@ -284,6 +284,11 @@ class StandardWorkflow(BaseWorkflow):
                 from mirp.imageFilters.squareTransformFilter import SquareTransformFilter
                 filter_obj = SquareTransformFilter(settings=self.settings, name=current_filter)
 
+            elif self.settings.img_transform.has_square_root_transform_filter(x=current_filter):
+                # Square root transform filter
+                from mirp.imageFilters.squareRootTransformFilter import SquareRootTransformFilter
+                filter_obj = SquareRootTransformFilter(settings=self.settings, name=current_filter)
+
             else:
                 raise ValueError(
                     f"{current_filter} is not implemented as a spatial filter. Please use one of ",
