### 2024-03-15 705f68d
Added logarithm transformation filter.
