+++ test/non_ibsi_compliant_filter_test.py
@@ -49,3 +49,26 @@ def test_square_root_transformation_filter():
     assert feature_data["sqrt_stat_min"].values[0] < 0.0
     assert np.max(data[0][1][0].get_voxel_grid()) == np.max(data[0][1][1].get_voxel_grid())
     assert not np.array_equal(data[0][1][0].get_voxel_grid(), data[0][1][1].get_voxel_grid())
+
+
+def test_logarithm_transformation_filter():
+    data = extract_features_and_images(
+        write_features=False,
+        export_features=True,
+        write_images=False,
+        export_images=True,
+        image_export_format="native",
+        image=os.path.join(CURRENT_DIR, "data", "ibsi_1_ct_radiomics_phantom", "dicom", "image"),
+        mask=os.path.join(CURRENT_DIR, "data", "ibsi_1_ct_radiomics_phantom", "dicom", "mask"),
+        roi_name="GTV-1",
+        ibsi_compliant=False,
+        base_feature_families="statistics",
+        filter_kernels="pyradiomics_logarithm"
+    )
+
+    feature_data = data[0][0]
+    assert len(feature_data) == 1
+    assert feature_data["stat_min"].values[0] == -1000.0
+    assert feature_data["lgrthm_stat_min"].values[0] < 0.0
+    assert np.max(data[0][1][0].get_voxel_grid()) == np.max(data[0][1][1].get_voxel_grid())
+    assert not np.array_equal(data[0][1][0].get_voxel_grid(), data[0][1][1].get_voxel_grid())
