### 2024-03-15 f3cdd04
Added exponential transformation filter.
