+++ mirp/workflows/standardWorkflow.py
@@ -294,6 +294,11 @@ class StandardWorkflow(BaseWorkflow):
                 from mirp.imageFilters.logarithmTransformFilter import LogarithmTransformFilter
                 filter_obj = LogarithmTransformFilter(settings=self.settings, name=current_filter)
 
+            elif self.settings.img_transform.has_exponential_transform_filter(x=current_filter):
+                # Exponential transform filter
+                from mirp.imageFilters.exponentialTransformFilter import ExponentialTransformFilter
+                filter_obj = ExponentialTransformFilter(settings=self.settings, name=current_filter)
+
             else:
                 raise ValueError(
                     f"{current_filter} is not implemented as a spatial filter. Please use one of ",
