### 2024-03-15 f9beae5
Added documentation for function transformations.
