### 2024-03-15 0b90802
Update NEWS.md
