### 2024-03-15 a4cbea6
Added ibsi_compliant as general configuration parameter.
