### 2024-03-15 #6413 @alexzwanenburg
@drcandacemakedamoore Steffen Löck is my professor and advised on the paper and the package. Stefan Leger contributed to early in-house versions of MIRP, prior to moving GitHub. Sebastian Starke also made minor contributions to earlier version of MIRP.
