### 2024-03-17 #6413 @drcandacemakedamoore
@alexzwanenburg on a less superficial issue, I note there is no developer's documentation. Many people may want to tinker with what you have done, and hopefully even contribute to the package. I am looking for documentation somewhere that tells people how to run the testing, so they can test new stuff before sending it. I am also looking for this, because it is not clear if you have any automated testing that runs in CI (did I miss it?) so instead of figuring out how to run it from there I would need instructions to run your tests properly (my naieve attempts are not working).
