### 2024-03-18 #6413 @emdupre
Thank you, @drcandacemakedamoore  !

If you could please open subsequent review comments [as issues on the _MIRP_ repository](https://github.com/oncoray/mirp/issues), this will help to make sure there is sufficient space for follow-up discussion and that action items are trackable across reviewers. I know that @alexzwanenburg has started to respond in-thread, but we'll generally ask to keep only high-level discussions in the general review thread and re-direct all other comments to the project issue tracker.

If you have any other questions, of course, please don't hesitate to ask.
