### 2024-03-18 oncoray/mirp#62 @drcandacemakedamoore
testing
There is no obvious  developer's documentation. The contributing.md would be fine except does not cover certain issues i.e. testing and linting. Many people may want to tinker with what you have done, and hopefully even contribute to the package. I am looking for documentation somewhere that tells people how to run the testing, so they can test new stuff before sending it. I eventually say some notes on contributing, somewhere. This could be in a more obvious place i.e. linked off the README, but the contributing.md is also fine as a stand-alone. Just a few more details please.
