### 2024-03-18 oncoray/mirp#63 @drcandacemakedamoore
PEP-8 compliant naming of files and modules
Preferably module names should be all lower case (and super-best is single word or if no other choice with underscore). I notice you have modules that are camelcase. Different file systems have different case conventions. The real point is that it could be possible to import these modules in two different ways then cause problems on different files systems where case conventions are different.
