### 2024-03-18 oncoray/mirp#63 @alexzwanenburg
Thanks for raising this point. The naming of modules was done with mixed case for historical reasons. Though we have experienced no problems operating MIRP on Linux or Windows platforms, [PEP 8](https://peps.python.org/pep-0008/#package-and-module-names) does recommend lowercase names for modules.

This is moreover a good opportunity to appropriately make most modules [private](https://peps.python.org/pep-0008/#public-and-internal-interfaces), in the sense that modules are not intended to be public-facing.
