### 2024-03-18 oncoray/mirp#62 @alexzwanenburg
Thanks for the feedback!

I will add instructions on style and testing to `contributing.md`, and link that file in `README.md`.
