### 2024-03-18 99edc9a
Added note on style and tests to contributing.md
closes #62
