+++ mirp/settings/settingsFeatureExtraction.py
@@ -256,7 +256,8 @@ class FeatureExtractionSettingsClass:
                 raise ValueError(
                     "Available values for the base_discretisation_method parameter are "
                     "'fixed_bin_number', 'fixed_bin_size', 'fixed_bin_size_pyradiomics' and 'none'. "
-                    "One or more values were not recognised.")
+                    "One or more values were not recognised."
+                )
 
             # Check discretisation_n_bins
             if "fixed_bin_number" in base_discretisation_method:
@@ -298,6 +299,13 @@ class FeatureExtractionSettingsClass:
                     raise ValueError(
                         "The base_discretisation_bin_width parameter is expected to contain floating "
                         "point values greater than 0.0. Found one or more values that were 0.0 or less.")
+
+            # Check ibsi_compliance
+            if self.ibsi_compliant and "fixed_bin_size_pyradiomics" in base_discretisation_method:
+                raise ValueError(
+                    "The fixed_bin_size_pyradiomics method is not IBSI-compliant. If you are sure that you want to "
+                    "use this method, use ibsi_compliant = False."
+                )
 
             else:
                 base_discretisation_bin_width = None
