+++ mirp/_data_import/imageDicomFileRTSTRUCT.py
@@ -1,0 +1,728 @@
+import copy
+import warnings
+
+import numpy as np
+import pydicom
+
+from mirp._data_import.maskContour import ContourClass
+from mirp._data_import.importImage import ImageFile
+from mirp._data_import.imageDicomFile import MaskDicomFile
+from mirp._masks.baseMask import BaseMask
+from mirp._data_import.utilities import get_pydicom_meta_tag, has_pydicom_meta_tag
+
+
+class MaskDicomFileRTSTRUCT(MaskDicomFile):
+    def __init__(self, **kwargs):
+
+        super().__init__(**kwargs)
+
+        # RTSTRUCT does not have its own image information.
+        self.image_origin = None
+        self.image_orientation = None
+        self.image_spacing = None
+        self.image_dimension = None
+
+    def is_stackable(self, stack_images: str):
+        return False
+
+    def create(self):
+        return self
+
+    def _complete_image_origin(self, force=False):
+        return
+
+    def _complete_image_orientation(self, force=False):
+        return
+
+    def _complete_image_spacing(self, force=False):
+        return
+
+    def _complete_image_dimensions(self, force=False):
+        return
+
+    def _complete_frame_of_reference_uid(self):
+        if self.frame_of_reference_uid is None:
+            # Try to obtain a frame of reference UID
+            if has_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0020, 0x0052)):
+                if get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0020, 0x0052), tag_type="str") is not None:
+                    self.frame_of_reference_uid = get_pydicom_meta_tag(
+                        dcm_seq=self.image_metadata,
+                        tag=(0x0020, 0x0052),
+                        tag_type="str")
+
+            # For RT structure sets, the FOR UID may be tucked away in the Structure Set ROI Sequence
+            if self.frame_of_reference_uid is None and \
+                    has_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x3006, 0x0020)):
+                structure_set_roi_sequence = self.image_metadata[(0x3006, 0x0020)]
+
+                for structure_set_roi_element in structure_set_roi_sequence:
+                    if has_pydicom_meta_tag(dcm_seq=structure_set_roi_element, tag=(0x3006, 0x0024)):
+                        self.frame_of_reference_uid = get_pydicom_meta_tag(
+                            dcm_seq=structure_set_roi_element,
+                            tag=(0x3006, 0x0024),
+                            tag_type="str")
+                        break
+
+    def check_mask(self, raise_error=True):
+        if self.image_metadata is None:
+            raise TypeError("DEV: the image_meta_data attribute has not been set.")
+
+        if not get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x3006, 0x0020), test_tag=True):
+            if raise_error:
+                warnings.warn(
+                    f"The RT-structure set ({self.file_path}) did not contain any ROI sequences.",
+                )
+            return False
+
+        roi_name_present = [
+            get_pydicom_meta_tag(
+                dcm_seq=current_structure_set_roi_sequence, tag=(0x3006, 0x0026), tag_type="str", default=None)
+            for current_structure_set_roi_sequence in self.image_metadata[(0x3006, 0x0020)]
+        ]
+
+        if len(roi_name_present) == 0 and raise_error:
+            warnings.warn(
+                f"The current RT-structure set ({self.file_path}) does not contain any ROI contours."
+            )
+            return False
+
+        if any(x is None for x in roi_name_present) and raise_error:
+            warnings.warn(
+                f"The current RT-structure set ({self.file_path}) lacks one or more ROI names."
[…]
