+++ mirp/_featuresets/volumeMorphology.py
@@ -1,0 +1,816 @@
+import hashlib
+import numpy as np
+import pandas as pd
+
+from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
+from mirp._images.genericImage import GenericImage
+from mirp._masks.baseMask import BaseMask
+
+
+def get_volumetric_morphological_features(
+        image: GenericImage,
+        mask: BaseMask,
+        settings: FeatureExtractionSettingsClass):
+    """
+    Extract morphological features from the image volume.
+    """
+
+    # Import functions
+    from scipy.spatial import ConvexHull
+    from scipy.spatial.distance import pdist
+    from skimage.measure import mesh_surface_area
+
+    # Create feature table
+    feat_names = [
+        "morph_volume", "morph_vol_approx", "morph_area_mesh", "morph_av", "morph_comp_1", "morph_comp_2",
+        "morph_sph_dispr", "morph_sphericity", "morph_asphericity", "morph_com",
+        "morph_diam", "morph_pca_maj_axis", "morph_pca_min_axis", "morph_pca_least_axis",
+        "morph_pca_elongation", "morph_pca_flatness", "morph_vol_dens_aabb", "morph_area_dens_aabb",
+        "morph_vol_dens_aee", "morph_area_dens_aee", "morph_vol_dens_conv_hull",
+        "morph_area_dens_conv_hull", "morph_integ_int", "morph_moran_i", "morph_geary_c"
+    ]
+
+    if not settings.ibsi_compliant:
+        feat_names += ["morph_vol_dens_ombb", "morph_area_dens_ombb",  "morph_vol_dens_mvee", "morph_area_dens_mvee"]
+
+    df_feat = pd.DataFrame(np.full(shape=(1, len(feat_names)), fill_value=np.nan))
+    df_feat.columns = feat_names
+
+    # Skip calculations if input information is missing
+    if image.is_empty() or mask.roi_intensity is None or mask.roi_morphology is None:
+        return df_feat
+
+    # Number of voxels within _masks
+    n_v_morph = np.sum(mask.roi_morphology.get_voxel_grid())
+    n_v_int = np.sum(mask.roi_intensity.get_voxel_grid())
+
+    # Check if any voxels are within the morphological mask
+    if n_v_morph == 0:
+        return df_feat
+
+    ####################################################################################################################
+    # Geometric features
+    ####################################################################################################################
+
+    # Surface area and volume from surface mesh
+    mesh_verts, mesh_faces, mesh_norms = mesh_voxels(mask=mask)
+
+    # Volume
+    volume = mesh_volume(vertices=mesh_verts, faces=mesh_faces)
+    df_feat["morph_volume"] = volume
+
+    # Approximate volume
+    df_feat["morph_vol_approx"] = n_v_morph * np.prod(mask.roi_morphology.image_spacing)
+
+    # Surface area
+    area = mesh_surface_area(verts=mesh_verts, faces=mesh_faces)
+    df_feat["morph_area_mesh"] = area
+
+    # Surface to volume ratio
+    df_feat["morph_av"] = area / volume
+
+    # Compactness 1
+    sphericity_base_feature = 36 * np.pi * volume ** 2.0 / area ** 3.0
+    df_feat["morph_comp_1"] = 1.0 / (6.0 * np.pi) * sphericity_base_feature ** (1.0 / 2.0)
+
+    # Compactness 2
+    df_feat["morph_comp_2"] = sphericity_base_feature
+
+    # Spherical disproportion
+    df_feat["morph_sph_dispr"] = sphericity_base_feature ** (-1.0 / 3.0)
+
+    # Sphericity
[…]

[…]
