### 2024-03-18 3c48873
Renamed files in _image_processing.
