### 2024-03-18 7e8a735
Renamed files in _imagefilters
