+++ mirp/_imagefilters/square_root_transform.py
@@ -2,12 +2,12 @@ import copy
 import numpy as np
 
 from mirp._images.genericImage import GenericImage
-from mirp._images.transformedImage import ExponentialTransformedImage
-from mirp._imagefilters.genericFilter import GenericFilter
+from mirp._images.transformedImage import SquareRootTransformedImage
+from mirp._imagefilters.generic import GenericFilter
 from mirp.settings.settingsGeneric import SettingsClass
 
 
-class ExponentialTransformFilter(GenericFilter):
+class SquareRootTransformFilter(GenericFilter):
 
     def __init__(self, settings: SettingsClass, name: str):
 
@@ -19,9 +19,9 @@ class ExponentialTransformFilter(GenericFilter):
     def generate_object(self):
         yield copy.deepcopy(self)
 
-    def transform(self, image: GenericImage) -> ExponentialTransformedImage:
+    def transform(self, image: GenericImage) -> SquareRootTransformedImage:
         # Create placeholder response map.
-        response_map = ExponentialTransformedImage(
+        response_map = SquareRootTransformedImage(
             image_data=None,
             template=image
         )
@@ -30,20 +30,14 @@ class ExponentialTransformFilter(GenericFilter):
             return response_map
 
         image_data = image.get_voxel_grid()
-        max_value = np.max(np.abs(image_data))
-
-        # Prevent issues with alpha values that are not strictly positive.
-        if not np.isfinite(max_value) or max_value == 0.0:
-            max_value = 1.0
-
-        alpha = np.log(max_value) / max_value
+        alpha = np.max(np.abs(image_data))
 
         # Prevent issues with alpha values that are not strictly positive.
         if not np.isfinite(alpha) or alpha == 0.0:
             alpha = 1.0
 
         response_map.set_voxel_grid(
-            voxel_grid=np.exp(image_data * alpha)
+            voxel_grid=np.sign(image_data) * np.sqrt(np.abs(image_data) * alpha)
         )
 
         return response_map
