+++ mirp/_imagefilters/gaussian.py
@@ -3,7 +3,7 @@ import copy
 
 from mirp._images.genericImage import GenericImage
 from mirp._images.transformedImage import GaussianTransformedImage
-from mirp._imagefilters.genericFilter import GenericFilter
+from mirp._imagefilters.generic import GenericFilter
 from mirp._imagefilters.utilities import FilterSet2D, FilterSet3D
 from mirp.settings.settingsGeneric import SettingsClass
 
