+++ mirp/_imagefilters/separable_wavelet.py
@@ -4,7 +4,7 @@ import copy
 from mirp._images.genericImage import GenericImage
 from mirp._images.transformedImage import SeparableWaveletTransformedImage
 from mirp.settings.settingsGeneric import SettingsClass
-from mirp._imagefilters.genericFilter import GenericFilter
+from mirp._imagefilters.generic import GenericFilter
 from mirp._imagefilters.utilities import pool_voxel_grids, SeparableFilterSet
 
 

[…]
