### 2024-03-18 2f7f56c
Renamed files in _images module.
