+++ mirp/_image_processing/normalise_image.py
@@ -1,7 +1,7 @@
 from typing import Any
 import numpy as np
 
-from mirp._images.genericImage import GenericImage
+from mirp._images.generic_image import GenericImage
 
 
 def normalise_image(
