+++ mirp/_featuresets/morphology_3d.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 
 from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
-from mirp._images.genericImage import GenericImage
+from mirp._images.generic_image import GenericImage
 from mirp._masks.baseMask import BaseMask
 
 
