+++ mirp/_imagefilters/mean.py
@@ -1,8 +1,8 @@
 import copy
 import numpy as np
 
-from mirp._images.genericImage import GenericImage
-from mirp._images.transformedImage import MeanTransformedImage
+from mirp._images.generic_image import GenericImage
+from mirp._images.transformed_image import MeanTransformedImage
 from mirp._imagefilters.generic import GenericFilter
 from mirp._imagefilters.utilities import SeparableFilterSet
 from mirp.settings.settingsGeneric import SettingsClass
