+++ mirp/_images/rtdose_image.py
@@ -1,6 +1,6 @@
 import numpy as np
 from typing import Any
-from mirp._images.genericImage import GenericImage
+from mirp._images.generic_image import GenericImage
 
 
 class RTDoseImage(GenericImage):
