### 2024-03-18 5df9c38
Renamed files in _images module.
