+++ mirp/_image_processing/randomise_mask.py
@@ -2,7 +2,7 @@ from mirp._image_processing.utilities import standard_image_process_checks
 from mirp._images.generic_image import GenericImage
 from mirp._images.mask_image import MaskImage
 from mirp._data_import.utilities import flatten_list
-from mirp._masks.baseMask import BaseMask
+from mirp._masks.base_mask import BaseMask
 
 
 def randomise_mask(
