+++ mirp/_image_processing/resegmentise_mask.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from mirp._image_processing.utilities import standard_image_process_checks
 from mirp._images.generic_image import GenericImage
-from mirp._masks.baseMask import BaseMask
+from mirp._masks.base_mask import BaseMask
 
 
 def resegmentise_mask(
