+++ test/read_image_and_mask_test.py
@@ -4,7 +4,7 @@ import numpy as np
 from mirp._data_import.import_image_and_mask import import_image_and_mask
 from mirp._data_import.read_data import read_image_and_masks
 from mirp._images.generic_image import GenericImage
-from mirp._masks.baseMask import BaseMask
+from mirp._masks.base_mask import BaseMask
 from mirp._images.ct_image import CTImage
 from mirp._images.pet_image import PETImage
 from mirp._images.mr_image import MRImage
