+++ test/dicom_rtdose_test.py
@@ -2,7 +2,7 @@ import os
 import numpy as np
 
 from mirp._images.rtdose_image import RTDoseImage
-from mirp._masks.baseMask import BaseMask
+from mirp._masks.base_mask import BaseMask
 from mirp.extractFeaturesAndImages import extract_features_and_images
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
