+++ mirp/_data_import/dicom_file_seg.py
@@ -6,7 +6,7 @@ from pydicom import dcmread
 
 from mirp._data_import.dicom_file import MaskDicomFile
 from mirp._data_import.utilities import get_pydicom_meta_tag, has_pydicom_meta_tag
-from mirp._masks.baseMask import BaseMask
+from mirp._masks.base_mask import BaseMask
 
 
 class MaskDicomFileSEG(MaskDicomFile):
