+++ mirp/_featuresets/dzm.py
@@ -6,7 +6,7 @@ import copy
 from mirp._featuresets.utilities import is_list_all_none
 from mirp._images.generic_image import GenericImage
 from mirp._masks.base_mask import BaseMask
-from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 from mirp.utilities.utilities import real_ndim
 
 
