+++ mirp/_featuresets/intensity_volume_histogram.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 from mirp._images.generic_image import GenericImage
 from mirp._images.transformed_image import TransformedImage
 from mirp._masks.base_mask import BaseMask
