+++ mirp/_featuresets/rlm.py
@@ -6,7 +6,7 @@ import pandas as pd
 from mirp._featuresets.utilities import get_neighbour_directions, is_list_all_none, coord2Index
 from mirp._images.generic_image import GenericImage
 from mirp._masks.base_mask import BaseMask
-from mirp.settings.settingsFeatureExtraction import FeatureExtractionSettingsClass
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 
 
 def get_rlm_features(
