+++ test/ibsi_1_test.py
@@ -2,7 +2,7 @@ import os
 import sys
 from math import log10, floor
 
-from mirp.extractFeaturesAndImages import extract_features
+from mirp.extract_features_and_images import extract_features
 from mirp.settings.generic import SettingsClass
 from mirp.settings.transformation_parameters import ImageTransformationSettingsClass
 from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
