+++ test/dl_preprocessing_test.py
@@ -1,7 +1,7 @@
 import os.path
 
 from mirp.settings.generic import SettingsClass
-from mirp.deepLearningPreprocessing import deep_learning_preprocessing
+from mirp.deep_learning_preprocessing import deep_learning_preprocessing
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
 
