+++ test/ibsi_2_feature_test.py
@@ -1,6 +1,6 @@
 import os
 
-from mirp.extractFeaturesAndImages import extract_features
+from mirp.extract_features_and_images import extract_features
 from mirp.settings.generic import SettingsClass
 from mirp.settings.transformation_parameters import ImageTransformationSettingsClass
 from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
