### 2024-03-18 a9ce93b
Fixes future deprecation warnings in pandas.
