+++ mirp/_data_import/itk_file_stack.py
@@ -104,7 +104,7 @@ class ImageITKFileStack(ImageFileStack):
             warnings.warn(
                 f"Inconsistent distance between slice origins of subsequent slices: {np.unique(image_slice_spacing)}. "
                 "Slices cannot be aligned correctly. This is likely due to missing slices. "
-                "MIRP will attempt to interpolate the missing slices and their ROI _masks for volumetric analysis.",
+                "MIRP will attempt to interpolate the missing slices and their ROI masks for volumetric analysis.",
                 UserWarning)
 
             # Update slice positions.
