+++ mirp/settings/image_processing_parameters.py
@@ -9,8 +9,8 @@ from mirp.settings.utilities import setting_def
 class ImagePostProcessingClass:
     """
     Parameters related to image processing. Note that parameters concerning image perturbation / augmentation and
-    resampling are set separately, see :class:`~mirp.settings.settingPerturbation.ImagePerturbationSettingsClass` and
-    :class:`~mirp.settings.settingsInterpolation.ImageInterpolationSettingsClass`.
+    resampling are set separately, see :class:`~mirp.settings.perturbation_parameters.ImagePerturbationSettingsClass` and
+    :class:`~mirp.settings.interpolation_parameters.ImageInterpolationSettingsClass`.
 
     Parameters
     ----------
@@ -75,7 +75,7 @@ class ImagePostProcessingClass:
         "range"` results in intensities being mapped to a [0.0, 1000.0] range instead of [0.0, 1.0].
 
     tissue_mask_type: {"none", "range", "relative_range"}, optional, default: "relative_range"
-        Type of algorithm used to produce an approximate tissue mask of the tissue. Such _masks can be used to select
+        Type of algorithm used to produce an approximate tissue mask of the tissue. Such masks can be used to select
         pixels for bias correction and intensity normalisation by excluding non-tissue voxels.
 
     tissue_mask_range: list of float, optional
