### 2024-03-18 74d6d19
Fixes a future deprecation warning in pandas.
