### 2024-03-18 37b50ed
Prevent future deprecation warnings in numpy.
