### 2024-03-18 ff7dac0
Made import_image, import_image_and_mask and import_mask part of the public API, as intended.
