+++ mirp/data_import/import_image_and_mask.py
@@ -1,5 +1,5 @@
-from mirp._data_import.import_image import import_image
-from mirp._data_import.import_mask import import_mask
+from mirp.data_import.import_image import import_image
+from mirp.data_import.import_mask import import_mask
 from mirp._data_import.generic_file import ImageFile, MaskFile
 from mirp._data_import.dicom_file import ImageDicomFile, MaskDicomFile
 from mirp._data_import.dicom_file_stack import ImageDicomFileStack
