+++ mirp/extract_features_and_images.py
@@ -348,7 +348,7 @@ def _base_extract_features_and_images(
         write_dir: None | str = None,
         **kwargs
 ):
-    from mirp._data_import.import_image_and_mask import import_image_and_mask
+    from mirp.data_import.import_image_and_mask import import_image_and_mask
     from mirp.settings.import_config_parameters import import_configuration_settings
 
     # Infer write_images, export_images, write_features, export_features based on write_dir.
