### 2024-03-18 1a4d0f7
Update NEWS to more accurately reflect changes.
