+++ docs/_sources/mirp.importData.rst.txt
@@ -201,7 +201,7 @@ MIRP processes and analyses images and masks. There are multiple ways to provide
 
 * By specifying the configuration in a stand-alone data ``xml`` file. An empty copy of the ``xml`` file can be
   created using :func:`mirp.utilities.config_utilities.get_data_xml`. The tags of the``xml`` file are the same as the
-  arguments of :func:`~mirp.importData.importImageAndMask.import_image_and_mask`, that are listed below.
+  arguments of :func:`~mirp.data_import.import_image_and_mask.import_image_and_mask`, that are listed below.
 
 Selecting specific images and masks
 -----------------------------------
@@ -294,8 +294,8 @@ API documentation
 
 .. note:: The :func:`~mirp.importData.importImageAndMask.import_image_and_mask` function is called internally by other
   functions. These function pass through keyword arguments to
-  :func:`~mirp.importData.importImageAndMask.import_image_and_mask`.
+  :func:`~mirp.data_import.import_image_and_mask.import_image_and_mask`.
 
-.. autofunction:: mirp.importData.importImageAndMask.import_image_and_mask
+.. autofunction:: mirp.data_import.import_image_and_mask.import_image_and_mask
 
 .. autofunction:: mirp.utilities.config_utilities.get_data_xml
