+++ docs/_sources/mirp.settings.rst.txt
@@ -6,16 +6,16 @@ Standardization Initiative. Many aspects of this workflow can be configured. Thi
 
 * Using keyword arguments. The keyword arguments match the parameters used to initialise the various settings objects
   documented below.
-* By creating a :class:`~mirp.settings.settingsGeneric.SettingsClass` object. This object can be initialised using the
+* By creating a :class:`~mirp.settings.generic.SettingsClass` object. This object can be initialised using the
   same keyword arguments as above. Alternatively, the attributes of the
-  :class:`~mirp.settings.settingsGeneric.SettingsClass` can be filled with the specific objects documented below.
+  :class:`~mirp.settings.generic.SettingsClass` can be filled with the specific objects documented below.
 * By specifying the configuration in a stand-alone settings ``xml`` file. An empty copy of the ``xml`` file can be
   created using :func:`~mirp.utilities.config_utilities.get_settings_xml`.
 
 General settings
 ----------------
 
-.. automodule:: mirp.settings.settingsGeneral
+.. automodule:: mirp.settings.general_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -23,7 +23,7 @@ General settings
 Image processing settings
 -------------------------
 
-.. automodule:: mirp.settings.settingsImageProcessing
+.. automodule:: mirp.settings.image_processing_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -31,7 +31,7 @@ Image processing settings
 Image perturbation settings
 ---------------------------
 
-.. automodule:: mirp.settings.settingsPerturbation
+.. automodule:: mirp.settings.perturbation_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -39,7 +39,7 @@ Image perturbation settings
 Image interpolation settings
 ----------------------------
 
-.. automodule:: mirp.settings.settingsInterpolation
+.. automodule:: mirp.settings.interpolation_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -47,7 +47,7 @@ Image interpolation settings
 Mask resegmentation settings
 ----------------------------
 
-.. automodule:: mirp.settings.settingsMaskResegmentation
+.. automodule:: mirp.settings.resegmentation_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -55,7 +55,7 @@ Mask resegmentation settings
 Feature computation settings
 ----------------------------
 
-.. automodule:: mirp.settings.settingsFeatureExtraction
+.. automodule:: mirp.settings.feature_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -63,7 +63,7 @@ Feature computation settings
 Image transformation settings
 -----------------------------
 
-.. automodule:: mirp.settings.settingsImageTransformation
+.. automodule:: mirp.settings.transformation_parameters
    :members:
    :no-undoc-members:
    :show-inheritance:
@@ -71,7 +71,7 @@ Image transformation settings
 Generic settings object
 -----------------------
 
-.. automodule:: mirp.settings.settingsGeneric
+.. automodule:: mirp.settings.generic
    :members:
    :no-undoc-members:
    :show-inheritance:
