+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = [
     "ray>=2.7.0",
     "scikit-image>=0.20.0",
     "scipy>=1.11",
-    "typing-extensions>=4.10; python<3.11"
+    "typing-extensions>=4.10;  python_version<'3.11'"
 ]
 
 [project.urls]
