### 2024-03-19 oncoray/mirp#66 @alexzwanenburg
I think this may be an issue with current version of `ray` instead of `mirp`. `ray` requires (but does not install) the `chardet` package. The error may disappear after installing that package using pip.
