### 2024-03-24 oncoray/mirp#66 @drcandacemakedamoore
> I think this may be an issue with current version of `ray` instead of `mirp`. `ray` requires (but does not install) the `chardet` package. The error may disappear after installing that package using pip.

I installed chardet, and it put in the latest version (5.2.0), but this does not solve the problem. I see you use the ray package. I have not looked too deeply into it, but it could even be a bug there. Do all your tests pass?
