### 2024-03-24 oncoray/mirp#67 @drcandacemakedamoore
supported matrix
As support for ray on Windows is experimental, last I checked, it isn't clear if this package is meant to run on Windows or not. Can you include a supported matrix to awnser questions of conda python and pypi python on one axis and OS on the other?

below is an example

  | Linux | Win | OSX
-- | -- | -- | --
pyth  | Supported | Unknown | Unknown
 conda | Unknown | Unknown | Unknown
