### 2024-03-25 oncoray/mirp#67 @alexzwanenburg
I can confirm `ray` works on Windows. Would you like to see such a matrix for `ray` or for MIRP itself?
