### 2024-03-26 ed3430f
Added automated test for continous integration.
