+++ test_auto_ci/ci_import_image_test.py
@@ -1,0 +1,636 @@
+import os.path
+import shutil
+
+import itk
+import numpy as np
+import pytest
+
+from mirp.data_import.import_image import import_image
+from mirp._data_import.itk_file import ImageITKFile
+from mirp._data_import.dicom_file_stack import ImageDicomFileStack
+from mirp._data_import.numpy_file import ImageNumpyFile
+from mirp._data_import.numpy_file_stack import ImageNumpyFileStack
+
+# Find path to the test directory. This is because we need to read datafiles stored in subdirectories.
+CURRENT_DIR = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "test")
+
+
+def _convert_to_numpy(as_slice=False):
+    """
+    Helper script for converting NIfTI files to numpy for testing numpy-based imports.
+
+    :param as_slice:
+    :return:
+    """
+
+    main_folders = ["STS_001", "STS_002", "STS_003"]
+    sub_folders = ["CT", "MR_T1", "PET"]
+
+    for sample_name in main_folders:
+        for modality in sub_folders:
+
+            # Read and parse image
+            source_image_file = os.path.join(
+                CURRENT_DIR, "data", "sts_images", sample_name, modality, "nifti", "image", "image.nii.gz")
+
+            source_image = itk.imread(source_image_file)
+            source_image = itk.GetArrayFromImage(source_image).astype(np.float32)
+
+            if as_slice:
+                folder_name = "numpy_slice"
+                target_dir = os.path.join(
+                    CURRENT_DIR, "data", "sts_images", sample_name, modality, folder_name, "image")
+
+                if not os.path.exists(target_dir):
+                    os.makedirs(target_dir)
+
+                for ii in range(source_image.shape[0]):
+                    target_image_file = os.path.join(
+                        target_dir, "_".join([sample_name, "{:02d}".format(ii), "image.npy"]))
+
+                    np.save(target_image_file, arr=source_image[ii, :, :].squeeze())
+
+            else:
+                folder_name = "numpy"
+                target_dir = os.path.join(
+                    CURRENT_DIR, "data", "sts_images", sample_name, modality, folder_name, "image")
+                target_image_file = os.path.join(target_dir, "_".join([sample_name, "image.npy"]))
+
+                if not os.path.exists(target_dir):
+                    os.makedirs(target_dir)
+
+                np.save(target_image_file, arr=source_image)
+
+            # Read and parse mask
+            source_mask_file = os.path.join(
+                CURRENT_DIR, "data", "sts_images", sample_name, modality, "nifti", "mask", "mask.nii.gz")
+
+            source_mask = itk.imread(source_mask_file)
+            source_mask = itk.GetArrayFromImage(source_mask).astype(int)
+
+            if as_slice:
+                folder_name = "numpy_slice"
+                target_dir = os.path.join(CURRENT_DIR, "data", "sts_images", sample_name, modality, folder_name, "mask")
+
+                if not os.path.exists(target_dir):
+                    os.makedirs(target_dir)
+
+                for ii in range(source_mask.shape[0]):
+                    target_mask_file = os.path.join(
+                        target_dir, "_".join([sample_name, "{:02d}".format(ii), "mask.npy"]))
+
+                    np.save(target_mask_file, arr=source_mask[ii, :, :].squeeze())
+
+            else:
+                folder_name = "numpy"
[…]
