### 2024-03-26 oncoray/mirp#67 @drcandacemakedamoore
I meant just for your package. No one should expect you to have to keep track of ray; but if you see in testing it starts breaking stuff apart for a certain OS/distribution combination (could happen) then consider moving that combination to something other than supported. Maybe under development until it catches up to a fix.
