### 2024-03-26 oncoray/mirp#68 @drcandacemakedamoore
I think it's fine to not test everything, but there should be some critical group of tests that get run on auto- otherwise it's too easy, especially if someone is running away with a fork of their own, to just start slowly breaking everything without realizing it.
