### 2024-03-26 3bf519e
Don't run perturbation test in auto_ci.
