+++ test_auto_ci/ci_perturbation_test.py
@@ -1,365 +1,0 @@
-import os
-
-import numpy as np
-import pandas as pd
-
-from mirp.settings.generic import SettingsClass
-from mirp.settings.transformation_parameters import ImageTransformationSettingsClass
-from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
-from mirp.settings.resegmentation_parameters import ResegmentationSettingsClass
-from mirp.settings.perturbation_parameters import ImagePerturbationSettingsClass
-from mirp.settings.image_processing_parameters import ImagePostProcessingClass
-from mirp.settings.interpolation_parameters import ImageInterpolationSettingsClass, MaskInterpolationSettingsClass
-from mirp.settings.general_parameters import GeneralSettingsClass
-
-# Find path to the test directory. This is because we need to read datafiles stored in subdirectories.
-CURRENT_DIR = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "test")
-WRITE_TEMP_FILES = False
-
-
-def test_noise_perturbation():
-
-    perturbation_settings = ImagePerturbationSettingsClass(
-        crop_around_roi=False,
-        perturbation_noise_repetitions=1
-    )
-
-    # Set up experiment.
-    data = run_experiment(perturbation_settings=perturbation_settings)
-    feature_table = pd.concat([x[0] for x in data])
-    image = [x[1][0] for x in data]
-    mask = [x[2][0] for x in data]
-
-    # Assert that object location and origin have not changed.
-    assert np.allclose(image[0]["image_origin"], [-101.4000, -79.9255, -174.7290])
-    assert np.allclose(image[0]["image_orientation"], np.array([[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]]))
-    assert np.allclose(mask[0]["image_origin"], [-101.4000, -79.9255, -174.7290])
-    assert np.allclose(mask[0]["image_orientation"], np.array([[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]]))
-
-    # Check noise levels and ids.
-    assert image[0]["noise_level"] > 0.0
-    assert image[0]["noise_id"] == 0
-
-    # Check feature table.
-    assert np.allclose(feature_table["morph_volume"], 357750.3)
-    assert np.all(40.0 < feature_table["stat_mean"].values[0] < 50.0)
-    assert np.all(40.0 < feature_table["stat_mean"].values[1] < 50.0)
-    assert not np.allclose(feature_table["stat_mean"], 43.085083)
-    assert feature_table["image_noise_iteration_id"].values[0] == 0
-
-
-def test_translation_perturbation():
-    perturbation_settings = ImagePerturbationSettingsClass(
-        crop_around_roi=False,
-        perturbation_translation_fraction=0.1
-    )
-
-    # Run experiment.
-    data = run_experiment(perturbation_settings=perturbation_settings)
-    feature_table = pd.concat([x[0] for x in data])
-    image = [x[1][0] for x in data]
-    mask = [x[2][0] for x in data]
-
-    # Origin has changed.
-    assert not np.allclose(image[0]["image_origin"], [-101.4000, -79.9255, -174.7290])
-    assert not np.allclose(mask[0]["image_origin"], [-101.4000, -79.9255, -174.7290])
-    assert np.allclose(image[0]["image_origin"], [-101.300, -79.826, -174.629])
-    assert np.allclose(mask[0]["image_origin"], [-101.300, -79.826, -174.629])
-
-    # Assert that object orientation did not change.
[…]
