### 2024-03-26 11af7a4
Run workflow on push (for testing purposes).
