### 2024-03-26 oncoray/mirp#68 @alexzwanenburg
I took the liberty of writing a CI workflow based on a standard GitHub workflow: [.github/workflows/auto-test-package.yml](https://github.com/oncoray/mirp/blob/f5a20f3333d1905c9b4e1fbc4d35ff1d097b8294/.github/workflows/auto-test-package.yml).

This also checks against different OS and python versions (currently limited to 3.10 and 3.11 due to limitations in the `ray` package).

A Pull Request is no longer required.
