### 2024-03-26 oncoray/mirp#68 @drcandacemakedamoore
Great but if I recall Python 3.11 should throw you some problems, and Python 3.12 definitely did when I tested it. Look very carefully at each package, and the requirements before adding Python 3.11.
