### 2024-03-26 af63045
Differentiate between push and pull actions for automated CI.
