+++ .github/workflows/auto-test-package_pull.yml
@@ -1,13 +1,9 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Run automated tests on multiple platforms
+name: Run automated tests
 
 on: [push]
-#  push:
-#    branches: [ "master" ]
-#  pull_request:
-#    branches: [ "master" ]
 
 jobs:
   build:
@@ -16,8 +12,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.10", "3.11"]
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ["3.11"]
+        os: [ubuntu-latest]
 
     steps:
     - uses: actions/checkout@v4
