+++ .github/workflows/auto-test-package_pull.yml
@@ -3,7 +3,7 @@
 
 name: Run automated tests
 
-on: [push]
+on: [pull_request]
 
 jobs:
   build:
