+++ .github/workflows/auto-test-package_push.yml
@@ -12,8 +12,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.10", "3.11"]
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ["3.11"]
+        os: [ubuntu-latest]
 
     steps:
     - uses: actions/checkout@v4
