+++ .github/workflows/auto-test-package_push.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Run automated tests on multiple platforms
+name: Run automated tests
 
 on: [push]
 
