### 2024-03-26 oncoray/mirp#68 @alexzwanenburg
Python 3.12 likely throws errors because its not supported by `ray` yet. Other versions seem to build correctly and test correctly: https://github.com/oncoray/mirp/actions/runs/8433681661.

The auto-ci test suite also includes a test that uses parallel processing.
