### 2024-03-26 bf61675
Added fall-back options for missing ray module.
