### 2024-03-26 3a74c77
Made ray package conditional.
