### 2024-03-26 c4bffba
Removed stated support for 3.13, as support is untested.
