### 2024-03-26 c901acc
Added python version 3.12 for continuous integration.
