### 2024-03-26 c006d43
Remove stated support for 3.12.
