+++ .github/workflows/auto-test-package_pull.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [ "3.10", "3.11", "3.12" ]
+        python-version: [ "3.10", "3.11"]
         os: [ ubuntu-latest, macos-latest, windows-latest ]
 
     steps:
