### 2024-03-26 oncoray/mirp#67 @alexzwanenburg
I added matrix to the readme. The aim is to automate the matrix with GitHub workflow badges, once GitHub supports the creation of multiple badges from a matrix workflow.
