### 2024-03-26 oncoray/mirp#66 @alexzwanenburg
Tests on multiple platforms were not able to reproduce the error, see #70.
