### 2024-03-27 ce8396b
Images can now be processed without providing a mask.
