+++ mirp/data_import/import_image_and_mask.py
@@ -1,6 +1,6 @@
 from mirp.data_import.import_image import import_image
 from mirp.data_import.import_mask import import_mask
-from mirp._data_import.generic_file import ImageFile, MaskFile
+from mirp._data_import.generic_file import ImageFile, MaskFile, MaskFullImage
 from mirp._data_import.dicom_file import ImageDicomFile, MaskDicomFile
 from mirp._data_import.dicom_file_stack import ImageDicomFileStack
 from mirp.utilities.utilities import random_string
@@ -109,7 +109,16 @@ def import_image_and_mask(
     list[ImageFile]
         The functions returns a list of ImageFile objects, if any were found with the specified filters.
     """
-    if mask is None:
+
+    # If mask = None, this can mean several things. Here we check that mask should not be interpreted as having the
+    # same meaning as image.
+    if mask is None and (
+            mask_name is not None or mask_sub_folder is not None or mask_modality is not None or
+            mask_file_type is not None or roi_name is not None
+    ):
+        mask = image
+
+    elif mask is None and isinstance(image, str) and image.endswith(".xml"):
         mask = image
 
     # Generate list of images.
@@ -123,75 +132,82 @@ def import_image_and_mask(
         stack_images=stack_images
     )
 
-    # Generate list of images.
-    mask_list = import_mask(
-        mask,
-        sample_name=sample_name,
-        mask_name=mask_name,
-        mask_file_type=mask_file_type,
-        mask_modality=mask_modality,
-        mask_sub_folder=mask_sub_folder,
-        stack_masks=stack_masks,
-        roi_name=roi_name
-    )
-
     if len(image_list) == 0:
         raise ValueError(f"No images were found. Possible reasons are lack of images with the preferred modality.")
-    if len(mask_list) == 0:
-        raise ValueError(f"No masks were found. Possible reasons are lack of masks with the preferred modality.")
 
-    # Determine association strategy, if this is unset.
-    possible_association_strategy = set_association_strategy(
-        image_list=image_list,
-        mask_list=mask_list
-    )
-
-    if association_strategy is None:
-        association_strategy = possible_association_strategy
-    elif isinstance(association_strategy, str):
-        association_strategy = [association_strategy]
-
-    if not isinstance(association_strategy, set):
-        association_strategy = set(association_strategy)
-
-    # Test association strategy.
-    unavailable_strategy = association_strategy - possible_association_strategy
-    if len(unavailable_strategy) > 0:
-        raise ValueError(
-            f"One or more strategies for associating images and masks are not available for the provided image and "
-            f"mask set: {', '.join(list(unavailable_strategy))}. Only the following strategies are available: "
-            f"{'. '.join(list(possible_association_strategy))}"
+    if mask is not None:
+        # Generate list of masks from mask.
+        mask_list = import_mask(
+            mask,
+            sample_name=sample_name,
+            mask_name=mask_name,
+            mask_file_type=mask_file_type,
+            mask_modality=mask_modality,
+            mask_sub_folder=mask_sub_folder,
+            stack_masks=stack_masks,
+            roi_name=roi_name
         )
 
-    if len(possible_association_strategy) == 0:
-        raise ValueError(
-            f"No strategies for associating images and masks are available, indicating that there is no clear way to "
-            f"establish an association."
+        if len(mask_list) == 0:
+            raise ValueError(f"No masks were found. Possible reasons are lack of masks with the preferred modality.")
+
+        # Determine association strategy, if this is unset.
+        possible_association_strategy = set_association_strategy(
+            image_list=image_list,
+            mask_list=mask_list
         )
 
-    # Start association.
-    if association_strategy == {"list_order"}:
-        # If only the list_order strategy is available, use this.
[…]
