+++ NEWS.md
@@ -2,6 +2,9 @@
 
 # Minor changes
 
+- If mask-related parameters are not provided for computing features or processing of images for deep learning, a 
+  mask is generated that covers the entire image.
+
 - Add fall-back methods for missing installation of the `ray` package for parallel processing. This can happen when 
   a python version is not supported by the `ray` package. `ray` is now a conditional dependency, until that package 
   is released for python `3.12`.
