### 2024-03-27 oncoray/mirp#66 @alexzwanenburg
I cannot reproduce the issue on my Windows installation. The issue is likely with `ray`:

- https://github.com/ray-project/ray/issues/40867
- https://github.com/ray-project/ray/issues/21994

I have added an try-except block to capture OS errors that prevent ray instances from being to created. A warning is shown instead.
