### 2024-03-27 e11d6f6
Try-except for capturing failure to start ray instances.
