### 2024-03-27 e3ab0eb
WIP on long-form documentation
#69
