+++ docs_source/source/quick_start.rst
@@ -1,0 +1,5 @@
+Quick-start
+===========
+
+Before you begin, you need to:
+    1. Install MIRP (see :ref:`installation`).
