### 2024-03-28 7178289
Update to internal image plotter.
