### 2024-03-28 f88567d
WIP on long-form documentation.
#69
