### 2024-04-02 0fe6bc8
Made defaults for extract_images correspond to extract_features.
