+++ docs_source/source/index.rst
@@ -14,19 +14,17 @@ Welcome to MIRP
    deep_learning
    quantitative_image_analysis
 
-Quick-start
-===========
-Medical Image Radiomics Processor (MIRP) is an IBSI-compliant python package for medical image analysis.
-MIRP focuses on radiomics applications and supports computation of features for conventional radiomics
-and image processing for deep-learning applications.
+Medical Image Radiomics Processor (MIRP) is an IBSI-compliant [Zwanenburg2020]_, [Whybra2024]_ python package for
+medical image  analysis. MIRP focuses on radiomics applications and supports computation of features for conventional
+radiomics and image processing for deep-learning applications.
 
-Installing MIRP
----------------
-MIRP is available from PyPI and can be installed using ``pip``, or other installer tools:
+References
+----------
 
-.. code-block:: console
+.. [Zwanenburg2020] Zwanenburg A, Vallieres M, Abdalah MA, Aerts HJWL, Andrearczyk V, Apte A, et al. The Image
+  Biomarker Standardization Initiative: Standardized Quantitative Radiomics for High-Throughput Image-based
+  Phenotyping. Radiology. 2020;295: 328-338. doi:10.1148/radiol.2020191145
 
-    pip install mirp
-
-
-
+.. [Whybra2024] Whybra P, Zwanenburg A, Andrearczyk V, Schaer R, Apte AP, Ayotte A, et al. The Image Biomarker
+  Standardization Initiative: Standardized Convolutional Filters for Reproducible Radiomics and Enhanced Clinical
+  Insights. Radiology. 2024;310: e231319. doi:10.1148/radiol.231319
