### 2024-04-02 #6413 @emdupre
👋  Hi everyone ! Thank you for your comments on _MIRP_ to date !

I just wanted to note that we have now passed the four week review window. If you could please work on finalizing your initial reviews as soon as possible, I would appreciate it.

Once you have finalized your initial reviews, you can let me know by responding directly in this thread. Of course, if you have any questions or blockers, please don't hesitate to let me know !

🙏 @surajpaib @Matthew-Jennings @drcandacemakedamoore @theanega
