### 2024-04-03 0f1765f
Changed default image_export_format to "dict"
