### 2024-04-03 f8d0d8e
Updated documentation to address #69.
