+++ docs_source/source/quantitative_image_analysis.rst
@@ -2,8 +2,8 @@ Process image and compute quantitative image features
 =====================================================
 
 Two of the main uses for MIRP are to process images and compute quantitative features from images. Both use the same
-standardized, IBSI 1 and IBSI 2 compliant, workflow. Two versions of the image processing and feature computation
-function exist:
+standardized, IBSI 1 and IBSI 2 compliant, workflow [Zwanenburg2020]_, [Whybra2024]_. Two versions of the image
+processing and feature computation function exist:
 
 * :func:`~mirp.extract_features_and_images.extract_features_and_images`: conventional function that processes images and
   computes features.
@@ -15,13 +15,18 @@ images) and only processing images (without computing features):
 
 * :func:`~mirp.extract_features_and_images.extract_features`: conventional function that only computes features.
 * :func:`~mirp.extract_features_and_images.extract_features_generator`: generator that only yields feature values.
-* :func:`~mirp.extract_features_and_images.extract_features_and_images`: conventional function that only processes images.
-* :func:`~mirp.extract_features_and_images.extract_features_and_images_generator`: generator that yields processed images.
+* :func:`~mirp.extract_features_and_images.extract_images`: conventional function that only processes images.
+* :func:`~mirp.extract_features_and_images.extract_images_generator`: generator that yields processed images.
 
 Examples
 --------
 
-MIRP can compute features from regions of interest in images. The simplest example is:
+MIRP can compute features from regions of interest in images. The features are described in [Zwanenburg2016]_.
+
+Minimal example
+^^^^^^^^^^^^^^^
+
+The following computes features from a single image and mask:
 
 .. code-block:: python
 
@@ -36,6 +41,9 @@ MIRP can compute features from regions of interest in images. The simplest examp
 
 The ``base_discretisation_method`` and its corresponding parameters are required as long as any texture or
 intensity-histogram features are involved.
+
+Interpolation example
+^^^^^^^^^^^^^^^^^^^^^
 
 A more realistic example involves interpolation to ensure that voxel spacing is the same for all images in a dataset.
 For example, a positron emission tomography (PET) dataset may be resampled to 3 by 3 by 3 mm isotropic voxels. This
@@ -56,6 +64,9 @@ is achieved by providing the ``new_spacing`` argument, i.e. ``new_spacing=3.0`` 
 
 Here, ``image_modality="PET"`` is used to declare that the image is a PET image. If this is a DICOM image, this
 argument is not necessary -- the modality can be inferred from the metadata.
+
+Slice-wise example
+^^^^^^^^^^^^^^^^^^
 
 Sometimes, in-plane resolution is much higher than axial resolution. For example, in (older) computed tomography (CT)
 images, in-plane resolution may be 1 by 1 mm, but the distance between slices can be 7 mm or greater.
@@ -78,6 +89,9 @@ This is achieved by providing the ``by_slice`` argument, i.e. ``by_slice=True``.
     )
 
 In the above example ``new_spacing=1.0`` causes all images to be resampled in-plane to a 1 mm resolution.
+
+Fixed Bin Number discretisation example
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 The previous examples used the *Fixed Bin Number* to discretise intensities within the mask into a fixed number of bins.
 For some imaging modalities, intensities carry a physical (or at least calibrated) meaning, such as Hounsfield units in
@@ -102,6 +116,9 @@ Because the resegmentation range is not set, the lower bound of the initial bin 
[…]
