### 2024-04-03 af7accb
Added contributing section to long-form documentation.
