### 2024-04-04 oncoray/mirp#69 @Matthew-Jennings
@theanega and @alexzwanenburg, just on point 2:

I believe that the four citations @theanega refers to are "narrative" in-text citations whereas the others are parenthetical. It is common to include both formats in a paper. See [this APA Style and Grammar Guidelines webpage](https://apastyle.apa.org/style-grammar-guidelines/citations/basic-principles/parenthetical-versus-narrative). I don't think any correction is needed here.
