### 2024-04-04 8b63576
Implement tmp_path Delete unused variables.
