+++ test/import_image_and_mask_test.py
@@ -588,7 +588,7 @@ def test_failure_multiple_image_and_mask_import():
     # DICOM stack and _masks for all samples, but with incorrect instructions.
     # No matching file type.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             image_file_type="nifti",
             image_sub_folder=os.path.join("CT", "dicom", "image"),
@@ -599,7 +599,7 @@ def test_failure_multiple_image_and_mask_import():
     # DICOM stack and _masks for all samples, but with incorrect instructions.
     # No matching image modality.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             image_modality="pet",
             image_sub_folder=os.path.join("CT", "dicom", "image"),
@@ -610,7 +610,7 @@ def test_failure_multiple_image_and_mask_import():
     # DICOM stack and _masks for all samples, but with incorrect instructions.
     # No matching mask modality.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             mask_modality="seg",
             image_sub_folder=os.path.join("CT", "dicom", "image"),
@@ -621,7 +621,7 @@ def test_failure_multiple_image_and_mask_import():
     # DICOM stack and _masks for all samples, but with incorrect instructions.
     # Wrong image_name.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             image_name="false_image",
             image_sub_folder=os.path.join("CT", "dicom", "image"),
@@ -633,7 +633,7 @@ def test_failure_multiple_image_and_mask_import():
     # DICOM stack and _masks for all samples, but with incorrect instructions.
     # Wrong mask_name.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             mask_name="false_mask",
             image_sub_folder=os.path.join("CT", "dicom", "image"),
@@ -645,7 +645,7 @@ def test_failure_multiple_image_and_mask_import():
     # Read Nifti image and _masks for all samples, but with incorrect instructions.
     # No matching sample name.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             image_sub_folder=os.path.join("CT", "nifti", "image"),
             sample_name="false_sample_name",
@@ -656,7 +656,7 @@ def test_failure_multiple_image_and_mask_import():
     # Read Nifti image and _masks for all samples, but with incorrect instructions.
     # Wrong image_name.
     with pytest.raises(ValueError) as exception_info:
-        image_list = import_image_and_mask(
+        import_image_and_mask(
             image=os.path.join(CURRENT_DIR, "data", "sts_images"),
             image_sub_folder=os.path.join("CT", "nifti", "image"),
             image_name="false_image",
@@ -668,7 +668,7 @@ def test_failure_multiple_image_and_mask_import():
     # Read Nifti image and _masks for all samples, but with incorrect instructions.
     # Wrong mask_name.
     with pytest.raises(ValueError) as exception_info:
[…]
