### 2024-04-04 oncoray/mirp#75 @Matthew-Jennings
JOSS review item: CI
1. FWIW, my investigations show ~75% test coverage. This isn't too bad, but it might be worth dedicating some spare time to lifting it a little, especially for the files with little-to-no coverage. See [mirp_coverage.xlsx](https://github.com/oncoray/mirp/files/14865970/mirp_coverage.xlsx) for per-file coverage details.

2. Improvements to temporary test directory handling are in #74. Not important for JOSS publication.

3. Comments related to implementation of GitHub CI can be found [here](https://github.com/oncoray/mirp/pull/70#discussion_r1550846352) and [here](https://github.com/oncoray/mirp/pull/70#discussion_r1550850394). Not important for JOSS publication.
