### 2024-04-04 219df87
Removed unused code.
Unused code identified in #75.
