### 2024-04-04 bf4bae5
Added test for gaussian filter.
Filter was missing coverage in #75.
