### 2024-04-04 6988ee5
Add type hints for pathlib.Path.
