+++ test/ibsi_2_feature_test.py
@@ -13,9 +13,6 @@ from mirp.settings.general_parameters import GeneralSettingsClass
 # Find path to the test directory. This is because we need to read datafiles stored in subdirectories.
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
 
-PERTURB_IMAGES = False
-WRITE_TEMP_FILES = True
-
 
 def within_tolerance(ref, tol, x):
     # Read from pandas Series
@@ -57,19 +54,9 @@ def _get_default_settings(
         resegmentation_intensity_range=[-1000.0, 400.0]
     )
 
-    if PERTURB_IMAGES:
-        perturbation_settings = ImagePerturbationSettingsClass(
-            crop_around_roi=False,
-            perturbation_rotation_angles=[-15.0, -10.0, 5.0, 0.0, 5.0, 10.0, 15.0],
-            perturbation_translation_fraction=[0.00, 0.25, 0.50, 0.75],
-            perturbation_roi_adapt_size=[-2.0, 0.0, 2.0],
-            perturbation_roi_adapt_type="distance"
-        )
-
-    else:
-        perturbation_settings = ImagePerturbationSettingsClass(
-            crop_around_roi=False
-        )
+    perturbation_settings = ImagePerturbationSettingsClass(
+        crop_around_roi=False
+    )
 
     feature_computation_parameters = FeatureExtractionSettingsClass(
         by_slice=by_slice,
@@ -84,8 +71,8 @@ def _process_experiment(
         configuration_id: str,
         by_slice: bool,
         image_transformation_settings: ImageTransformationSettingsClass,
-        tmp_path: str | os.PathLike,
-        base_feature_families: str = "none"):
+        base_feature_families: str = "none"
+):
 
     # Get default settings.
     general_settings, image_interpolation_settings, feature_computation_parameters, resegmentation_settings, \
@@ -116,20 +103,10 @@ def _process_experiment(
 
     data = data[0]
 
-    if WRITE_TEMP_FILES:
-        file_name = "_".join([configuration_id, "perturb", "features.csv"] if PERTURB_IMAGES else [configuration_id, "features.csv"])
-
-        data.to_csv(
-            tmp_path / file_name,
-            sep=";",
-            decimal=".",
-            index=False
-        )
-
     return data
 
 
-def test_ibsi_2_config_none(tmp_path):
+def test_ibsi_2_config_none():
     """
     Compare computed feature values with reference values for configurations 1A and 1B of IBSI 2 phase 2.
     """
@@ -147,7 +124,6 @@ def test_ibsi_2_config_none(tmp_path):
         configuration_id="1.A",
         by_slice=True,
         image_transformation_settings=image_transformation_settings,
-        tmp_path=tmp_path,
         base_feature_families="statistics"
     )
 
@@ -183,7 +159,6 @@ def test_ibsi_2_config_none(tmp_path):
         configuration_id="1.B",
         by_slice=False,
         image_transformation_settings=image_transformation_settings,
-        tmp_path=tmp_path,
         base_feature_families="statistics"
     )
 
@@ -207,7 +182,7 @@ def test_ibsi_2_config_none(tmp_path):
     assert (within_tolerance(52600, 2800, data["stat_var"]))
 
 
-def test_ibsi_2_config_mean_filter(tmp_path):
+def test_ibsi_2_config_mean_filter():
     """
     Compare computed feature values with reference values for configurations 2A and 2B of IBSI 2 phase 2.
     """
@@ -226,7 +201,6 @@ def test_ibsi_2_config_mean_filter(tmp_path):
         configuration_id="2.A",
         by_slice=True,
[…]
