### 2024-04-04 330d7f9
Removed unused global variable.
