+++ test/perturbation_test.py
@@ -15,8 +15,6 @@ from mirp.settings.general_parameters import GeneralSettingsClass
 # Find path to the test directory. This is because we need to read datafiles stored in subdirectories.
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
 
-WRITE_TEMP_FILES = False
-
 
 def test_noise_perturbation():
 
@@ -475,8 +473,6 @@ def run_experiment(perturbation_settings, by_slice=False):
         by_slice=by_slice,
         perturbation_settings=perturbation_settings
     )
-
-    # TODO: Is support for writing a tempdir needed? `write_dir` wasn't used...
 
     data = extract_features_and_images(
         write_features=False,
