### 2024-04-04 oncoray/mirp#68 @alexzwanenburg
To do:
- [ ] Mark automated tests using custom pytest.mark decorators: https://docs.pytest.org/en/latest/example/markers.html#mark-examples
