### 2024-04-04 #6413 @emdupre
👋  Hello again,

I just wanted to follow up on the previous message as I know if created some concern and confusion :

- I should have noted that the "initial review" refers to each reviewer working through their checklist and opening up associated issues. I did not mean to imply that those issues should have also been resolved !
- The "four week review window" was in reference to our original request to finish reviews within four weeks if possible or six weeks at latest. We have just passed five weeks on this review ; however, if you require additional time beyond the six week window, please let me know. I will try to accommodate specific timelines as possible.

Apologies for being unclear on these points in my previous message. If there's anything else I can clarify, please let me know. And thank you again for your work in reviewing _MIRP_ !
