### 2024-04-05 8edb5a5
Introduce custom ci marker to mark CI tests.
