### 2024-04-05 def40bc
Update GitHub workflow to select tests with the CI marker.
