+++ test_auto_ci/ci_import_mask_test.py
@@ -1,419 +1,0 @@
-import os.path
-
-import pytest
-
-# Find path to the test directory. This is because we need to read datafiles stored in subdirectories.
-from mirp.data_import.import_mask import import_mask
-from mirp._data_import.itk_file import MaskITKFile
-from mirp._data_import.dicom_file_rtstruct import MaskDicomFileRTSTRUCT
-from mirp._data_import.numpy_file import MaskNumpyFile
-from mirp._data_import.numpy_file_stack import MaskNumpyFileStack
-
-CURRENT_DIR = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "test")
-
-
-def test_single_mask_import():
-
-    # Read a Nifti image directly.
-    mask_list = import_mask(
-        os.path.join(CURRENT_DIR, "data", "sts_images", "STS_001", "CT", "nifti", "mask", "mask.nii.gz"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskITKFile)
-    assert mask_list[0].modality == "generic_mask"
-
-    # Read a DICOM RTSTRUCT file directly.
-    mask_list = import_mask(
-        os.path.join(CURRENT_DIR, "data", "sts_images", "STS_001", "CT", "dicom", "mask", "RS.dcm"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskDicomFileRTSTRUCT)
-    assert mask_list[0].sample_name == "STS_001"
-    assert mask_list[0].modality == "rtstruct"
-
-    # Read a numpy image directly.
-    mask_list = import_mask(
-        os.path.join(CURRENT_DIR, "data", "sts_images", "STS_001", "CT", "numpy", "mask", "STS_001_mask.npy"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskNumpyFile)
-    assert mask_list[0].modality == "generic_mask"
-
-    # Read a numpy stack.
-    mask_list = import_mask(
-        mask=os.path.join(CURRENT_DIR, "data", "sts_images", "STS_001", "CT", "numpy_slice", "mask"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskNumpyFileStack)
-    assert mask_list[0].modality == "generic_mask"
-
-    # Read a Nifti image for a specific sample.
-    mask_list = import_mask(
-        mask=os.path.join(CURRENT_DIR, "data", "sts_images"),
-        sample_name="STS_001",
-        mask_sub_folder=os.path.join("CT", "nifti", "mask"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskITKFile)
-    assert mask_list[0].sample_name == "STS_001"
-    assert mask_list[0].modality == "generic_mask"
-
-    # Read a DICOM RTSTRUCT file for a specific sample.
-    mask_list = import_mask(
-        mask=os.path.join(CURRENT_DIR, "data", "sts_images"),
-        sample_name="STS_001",
-        mask_sub_folder=os.path.join("CT", "dicom", "mask"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskDicomFileRTSTRUCT)
-    assert mask_list[0].sample_name == "STS_001"
-    assert mask_list[0].modality == "rtstruct"
-
-    # Read a numpy image for a specific sample.
-    mask_list = import_mask(
-        mask=os.path.join(CURRENT_DIR, "data", "sts_images"),
-        sample_name="STS_001",
-        mask_sub_folder=os.path.join("CT", "numpy", "mask"))
-    assert len(mask_list) == 1
-    assert isinstance(mask_list[0], MaskNumpyFile)
-    assert mask_list[0].sample_name == "STS_001"
[…]

[…]
