### 2024-04-05 9566f95
Added missing DICOM SEG test.
