### 2024-04-05 d9c18a5
Fixed improper indexing for slice-wise approaches.
