### 2024-04-05 bfff685
Added missing boundary split unit test.
