### 2024-04-05 f2a3289
Added test for non-ibsi-compliant features.
