### 2024-04-05 ff534cf
Add test for noise estimation methods.
