### 2024-04-05 221aa6a
Added test to assess correct image metadata export.
