+++ test/extract_image_parameters_test.py
@@ -1,4 +1,6 @@
 import os
+
+import pandas as pd
 import pytest
 from mirp.extract_image_parameters import extract_image_parameters
 
@@ -15,7 +17,7 @@ def test_extract_image_parameters_default():
     assert all(x in image_parameters.columns for x in ["modality", "spacing_z", "spacing_y", "spacing_x"])
     assert len(image_parameters) == 1
 
-    # Read multiple _images.
+    # Read multiple images.
     image_parameters = extract_image_parameters(
         image=os.path.join(CURRENT_DIR, "data", "sts_images"),
         image_sub_folder=os.path.join("CT", "nifti", "image")
@@ -32,7 +34,7 @@ def test_extract_image_parameters_dicom():
     )
     assert len(image_parameters) == 1
 
-    # Read multiple CT _images.
+    # Read multiple CT images.
     image_parameters = extract_image_parameters(
         image=os.path.join(CURRENT_DIR, "data", "sts_images"),
         image_sub_folder=os.path.join("CT", "dicom", "image")
@@ -45,7 +47,7 @@ def test_extract_image_parameters_dicom():
     )
     assert len(image_parameters) == 1
 
-    # Read multiple PET _images.
+    # Read multiple PET images.
     image_parameters = extract_image_parameters(
         image=os.path.join(CURRENT_DIR, "data", "sts_images"),
         image_sub_folder=os.path.join("PET", "dicom", "image")
@@ -58,14 +60,14 @@ def test_extract_image_parameters_dicom():
     )
     assert len(image_parameters) == 1
 
-    # Read multiple T1-weighted MR _images.
+    # Read multiple T1-weighted MR images.
     image_parameters = extract_image_parameters(
         image=os.path.join(CURRENT_DIR, "data", "sts_images"),
         image_sub_folder=os.path.join("MR_T1", "dicom", "image")
     )
     assert len(image_parameters) == 3
 
-    # Read multiple DICOM _images.
+    # Read multiple DICOM images.
     image_parameters = extract_image_parameters(
         image=os.path.join(CURRENT_DIR, "data", "sts_images"),
         image_file_type="dicom"
@@ -77,3 +79,16 @@ def test_extract_image_parameters_dicom():
         image=os.path.join(CURRENT_DIR, "data", "rtdose_images", "Pancreas-CT-CB_001", "rtdose")
     )
     assert len(image_parameters) == 1
+
+
+def test_extract_image_parameters_dicom_to_file(tmp_path):
+
+    extract_image_parameters(
+        image=os.path.join(CURRENT_DIR, "data", "sts_images"),
+        image_sub_folder=os.path.join("CT", "dicom", "image"),
+        write_file=True,
+        write_dir=tmp_path
+    )
+
+    image_parameters = pd.read_csv(os.path.join(tmp_path, "image_metadata.csv"))
+    assert len(image_parameters) == 3
