### 2024-04-05 888fa9c
Removed redundant write_file argument.
