### 2024-04-05 562da80
Added write test for extract_mask_labels.
