### 2024-04-05 oncoray/mirp#75 @alexzwanenburg
I have increased test coverage to 89% by writing additional tests. Some tests remain.
