### 2024-04-05 7948b81
Skeleton for tutorial notebook #73 #72 #69
Aim is to transition quick_start.rst into the notebook and create a working example.
