### 2024-04-06 oncoray/mirp#76 @drcandacemakedamoore
minor issues before some JOSS checkmarks
The ways to seek help could be a little clearer given the typical level of all the kinds of people interested in radiomics (physicians, graduate students not from computer science or  software engineering.) Maybe add a sentence telling people to open issues, and make a general code and use issue instead of the default issue being about a version release. Also, given the variety of people who may use this, I would write out the acronym IBSI on both the read-me and documentation (some people don't know about these standards yet).
