### 2024-04-07 oncoray/mirp#75 @Matthew-Jennings
I get 84%. See [mirp_coverage_24-04-07.xlsx](https://github.com/oncoray/mirp/files/14897197/mirp_coverage_24-04-07.xlsx). Not quite 89%! But probably enough 😄

For fair comparison, I'm using `pytest-cov` with the following command:

```shell
pytest -v --color=yes --cov=mirp --cov-report html --cov-report term --cov-report xml:cov.xml -n=14
```

Very pleased to find that I was able to run `pytest-xdist` (the `-n=14`) with no issue!
