### 2024-04-07 oncoray/mirp#75 @alexzwanenburg
Thanks again for the pull request that made pytest-xdist possible. One difference may be that my coverage results included the tests themselves.
