### 2024-04-08 eeb334f
Add options to set slice_id to show.
