### 2024-04-09 83a7873
Enabled building of html from jupyter notebook.
