+++ docs_source/source/index.rst
@@ -1,19 +1,40 @@
-Welcome to MIRP
-===============
+Introduction
+============
 
 .. toctree::
+   :hidden:
    :maxdepth: 1
-   :caption: Contents:
+   :caption: Getting Started
 
+   index
    installation
-   quick_start
+
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+   :caption: Tutorials
+
+   tutorial_compute_radiomics_features_mr
+
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+   :caption: Deep Dive
+
    image_mask_import
    configuration
    image_metadata
    mask_labels
    deep_learning
    quantitative_image_analysis
+
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+   :caption: Contributing
+
    contributing
+
 
 What is radiomics?
 ------------------
