+++ pyproject.toml
@@ -41,7 +41,7 @@ Changelog = "https://github.com/oncoray/mirp/blob/master/NEWS.md"
 
 [project.optional-dependencies]
 test = ["pytest>=7.4.0"]
-docs = ["sphinx>=5.0.0", "sphinx_rtd_theme>=2.0.0"]
+docs = ["sphinx>=5.0.0", "sphinx_rtd_theme>=2.0.0", "nbsphinx"]
 
 [tool.pytest.ini_options]
 addopts = ["--strict-markers"]
