+++ README.md
@@ -1,19 +1,37 @@
 <img src="https://raw.githubusercontent.com/oncoray/mirp/master/icon/mirp.svg" align="right" width="120"/>
 
-| python | Linux     | Win       | OSX       |
+# Medical Image Radiomics Processor
+
+Medical Image Radiomics Processor (MIRP) is a python package for medical image analysis that is compliant with the 
+reference standards of the Image Biomarker Standardisation Initiative (IBSI). MIRP focuses on radiomics applications 
+and supports computation of features for conventional radiomics and image processing for deep-learning applications.
+
+## Documentation and tutorials
+
+Documentation and tutorials can be found here: https://oncoray.github.io/mirp/
+
+## Supported Python and OS
+
+MIRP currently supports the following Python versions and operating systems: 
+
+| Python | Linux     | Win       | OSX       |
 |--------|-----------|-----------|-----------|
 | 3.10   | Supported | Supported | Supported |
 | 3.11   | Supported | Supported | Supported |
 
-# Medical Image Radiomics Processor
+## Supported imaging and mask modalities
 
-Medical Image Radiomics Processor (MIRP) is an IBSI-compliant python package for medical image analysis.
-MIRP focuses on radiomics applications and supports computation of features for conventional radiomics
-and image processing for deep-learning applications.
+MIRP currently supports the following image modalities:
 
-## Documentation
+| File format | File type | Supported modality |
+|-------------|-----------|--------------------|
+| DICOM       | image     | CT, MR, PT, RTDOSE |
+| DICOM       | mask      | RTSTRUCT, SEG      |
+| NIfTI       | any       | any                |
+| NRRD        | any       | any                |
+| numpy       | any       | any                |
 
-Documentation can be found here: https://oncoray.github.io/mirp/
+NIfTI, NRRD, and numpy files support any kind of (single-channel) image. MIRP cannot process RGB or 4D images.
 
 ## Installing MIRP
 MIRP is available from PyPI and can be installed using `pip`, or other installer tools:
@@ -21,39 +39,6 @@ MIRP is available from PyPI and can be installed using `pip`, or other installer
 ```commandline
 pip install mirp
 ```
-
-## Transitioning to version 2
-
-Version 2 is a major refactoring of the previous code base. For users this brings the following noticeable changes:
-
-- MIRP was previously configured using two `xml` files: [`config_data.xml`](mirp/config_data.xml) for configuring
-  directories, data to be read, etc., and [`config_settings.xml`](mirp/config_settings.xml) for configuring experiments.
-  While these two files can still be used, MIRP can now be configured directly, without using these files.
-- The main functions of MIRP (`mainFunctions.py`) have all been re-implemented.
-  - `mainFunctions.extract_features` is now `extract_features` (functional form) or
-    `extract_features_generator` (generator). The replacements allow for both writing
-    feature values to a directory and returning them as function output. 
-  - `mainFunctions.extract_images_to_nifti` is now `extract_images` (functional form) or
-     `extract_images_generator` (generator). The replacements allow for both writing 
-     images to a directory (e.g., in NIfTI or numpy format) and returning them as function output.
-  - `mainFunctions.extract_images_for_deep_learning` has been replaced by 
-    `deep_learning_preprocessing` (functional form) and 
-    `deep_learning_preprocessing_generator` (generator).
-  - `mainFunctions.get_file_structure_parameters` and `mainFunctions.parse_file_structure` are deprecated, as the
-    the file import system used in version 2 no longer requires a rigid directory structure.
-  - `mainFunctions.get_roi_labels` is now `extract_mask_labels`.
-  - `mainFunctions.get_image_acquisition_parameters` is now `extract_image_parameters`.
-
[…]
