### 2024-04-09 00db439
Added bug template.
#76
