+++ .github/ISSUE_TEMPLATE/general-issue.md
@@ -1,0 +1,29 @@
+---
+name: Bugs and problems
+about: General issue template
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+Please use the below template to submit your issue:
+
+# Expected behaviour
+
+What did you expect MIRP to do?
+
+# Observed behaviour
+
+What did MIRP actually do? What error (if any) was produced?
+
+# Steps to reproduce the problem
+
+What steps are required to reproduce the problem? If possible, provide code and data that led to the problem.
+
+# Specifications
+
+- MIRP version: 
+- Python version:
+- Operating system:
+- Other package versions (*if relevant*):
