+++ docs_source/source/index.rst
@@ -1,12 +1,12 @@
-Introduction
-============
+MIRP
+====
 
 .. toctree::
    :hidden:
    :maxdepth: 1
    :caption: Getting Started
 
-   index
+   introduction
    installation
 
 .. toctree::
@@ -35,25 +35,4 @@ Introduction
 
    contributing
 
-
-What is radiomics?
-------------------
-
-
-What is MIRP?
--------------
-
-Medical Image Radiomics Processor (MIRP) is an IBSI-compliant [Zwanenburg2020]_, [Whybra2024]_ python package for
-medical image  analysis. MIRP focuses on radiomics applications and supports computation of features for conventional
-radiomics and image processing for deep-learning applications.
-
-References
-----------
-
-.. [Zwanenburg2020] Zwanenburg A, Vallieres M, Abdalah MA, Aerts HJWL, Andrearczyk V, Apte A, et al. The Image
-  Biomarker Standardization Initiative: Standardized Quantitative Radiomics for High-Throughput Image-based
-  Phenotyping. Radiology. 2020;295: 328-338. doi:10.1148/radiol.2020191145
-
-.. [Whybra2024] Whybra P, Zwanenburg A, Andrearczyk V, Schaer R, Apte AP, Ayotte A, et al. The Image Biomarker
-  Standardization Initiative: Standardized Convolutional Filters for Reproducible Radiomics and Enhanced Clinical
-  Insights. Radiology. 2024;310: e231319. doi:10.1148/radiol.231319
+Welcome to MIRP!
