+++ docs_source/source/introduction.rst
@@ -1,0 +1,30 @@
+Introduction
+============
+
+What is MIRP?
+-------------
+
+Medical Image Radiomics Processor (MIRP) is a python package for medical image analysis that is compliant with the
+reference standards of the Image Biomarker Standardisation Initiative (IBSI) [Zwanenburg2020]_,
+[Whybra2024]_. MIRP focuses on radiomics applications  and supports computation of features for conventional
+radiomics and image processing for deep-learning applications.
+
+What is radiomics?
+------------------
+
+Contributing
+------------
+
+Contact
+-------
+
+References
+----------
+
+.. [Zwanenburg2020] Zwanenburg A, Vallieres M, Abdalah MA, Aerts HJWL, Andrearczyk V, Apte A, et al. The Image
+  Biomarker Standardization Initiative: Standardized Quantitative Radiomics for High-Throughput Image-based
+  Phenotyping. Radiology. 2020;295: 328-338. doi:10.1148/radiol.2020191145
+
+.. [Whybra2024] Whybra P, Zwanenburg A, Andrearczyk V, Schaer R, Apte AP, Ayotte A, et al. The Image Biomarker
+  Standardization Initiative: Standardized Convolutional Filters for Reproducible Radiomics and Enhanced Clinical
+  Insights. Radiology. 2024;310: e231319. doi:10.1148/radiol.231319
