### 2024-04-09 8df065c
Added details on supported modalities, Python versions and OS
