### 2024-04-10 5de43e4
Fixed compute_diagnostic_features method.
