### 2024-04-10 9097ca9
Added type hint to write methods.
