### 2024-04-10 0fb030e
Removed append_name method as it was not used.
