### 2024-04-10 d15c186
Added tests for several tests for BaseMask methods that lacked coverage.
