### 2024-04-10 5c7b994
Added test for BaseMask interpolate method.
