### 2024-04-10 cceed11
Added test for the BaseMask get_slices method.
