### 2024-04-10 1869727
Added coverage exclusion for code that is currently not used or reachable.
