### 2024-04-10 1e653f8
Added unit tests for BaseMask methods related to mask operations.
