### 2024-04-11 72efd75
Add piping to pass arguments to load_data methods.
#75
