+++ mirp/_workflows/standardWorkflow.py
@@ -84,7 +84,10 @@ class StandardWorkflow(BaseWorkflow):
         logging.info(self._message_start())
 
         # Read image and masks.
-        image, masks = read_image_and_masks(self.image_file, to_numpy=False)
+        image, masks = read_image_and_masks(
+            self.image_file,
+            to_numpy=False
+        )
 
         if masks is None or len(masks) == 0:
             warnings.warn("No segmentation masks were read.")
