### 2024-04-11 oncoray/mirp#75 @alexzwanenburg
To expose (and test) SUV conversion we need to do the following:

- [ ] Add piping for passing arguments from `StandardWorkflow.standard_image_processing` to `ImageDicomFilePT.load_data`.
- [ ] Expose the following parameters by adding them to `ImagePostProcessingClass`:
  - [ ] `suv_type` with default `body_weight`.
  - [ ] `decay_correction_point` with default `administration`.
- [ ] Add documentation for new parameters.
- [ ] Add parameters to xml.
- [ ] Integrate `SUVScalingObj` into `ImageDicomFilePT` as private methods, since the required metadata for SUV conversion is not present in different file formats.
- [ ] Add unit tests for parameters and parameter combinations.
