### 2024-04-11 21df6b7
Add documentation and specification for pet_suv_conversion parameter
