+++ mirp/settings/image_processing_parameters.py
@@ -16,8 +16,10 @@ class ImagePostProcessingClass:
     ----------
     bias_field_correction: bool, optional, default: False
         Determines whether N4 bias field correction should be performed. When a tissue mask is present, bias field
-        correction is conducted using the information contained within the mask. Bias-field correction can only be
-        applied to MR imaging.
+        correction is conducted using the information contained within the mask.
+
+        .. note::
+            Bias-field correction can only be applied to MR imaging.
 
     bias_field_correction_n_fitting_levels: int, optional, default: 1
         The number of fitting levels for the N4 bias field correction algorithm.
@@ -28,6 +30,21 @@ class ImagePostProcessingClass:
 
     bias_field_convergence_threshold: float, optional, default: 0.001
         Convergence threshold for N4 bias field correction algorithm.
+
+    pet_suv_conversion: {"body_weight", "none"}, default: "body_weight"
+        Intensities in PET imaging are often stored as detected radiotracer activity. To make detected activity more
+        comparable between patients, these are converted to standardised uptake values. The following are possible:
+
+        * "body_weight": activity is normalised by body weight.
+        * "none": activity is not normalised.
+
+        .. note::
+            Conversion of activity to standardised uptake values can only be performed for PET data.
+
+        .. warning::
+            Conversion of activity to standardised uptake values requires metadata related to image acquisition and to
+            the patient. These metadata are only present in DICOM files. MIRP cannot convert activity to standardised
+            uptake values for images in different formats, and this parameter will have no effect.
 
     intensity_normalisation: {"none", "range", "relative_range", "quantile_range", "standardisation"}, default: "none"
         Specifies the algorithm used to normalise intensities in the image. Will use only intensities in voxels
@@ -94,6 +111,7 @@ class ImagePostProcessingClass:
             bias_field_correction_n_fitting_levels: int = 1,
             bias_field_correction_n_max_iterations: int | list[int] | None = None,
             bias_field_convergence_threshold: float = 0.001,
+            pet_suv_conversion: str = "body_weight",
             intensity_normalisation: str = "none",
             intensity_normalisation_range: list[float] | None = None,
             intensity_normalisation_saturation: list[float] | None = None,
@@ -183,6 +201,16 @@ class ImagePostProcessingClass:
 
         # Set convergence_threshold attribute.
         self.convergence_threshold: None | float = bias_field_convergence_threshold
+
+        # Check that pet_suv_conversion has the correct values.
+        if pet_suv_conversion not in ["body_weight", "none"]:
+            raise ValueError(
+                f"The pet_suv_conversion parameter is expected to have one of the following values: ",
+                f"'body_weight', 'none'. Found: {pet_suv_conversion}"
+            )
+
+        # Set suv_conversion_type parameter.
+        self.suv_conversion_type = pet_suv_conversion
 
         # Check that intensity_normalisation has the correct values.
         if intensity_normalisation not in ["none", "range", "relative_range", "quantile_range", "standardisation"]:
