### 2024-04-11 79bf860
Added settings definition for SUV conversion.
