### 2024-04-11 bbae64f
WIP on introducing SUV computation into ImageDicomFilePT
