+++ mirp/_data_import/dicom_file_pet.py
@@ -1,3 +1,5 @@
+import warnings
+
 import numpy as np
 import datetime
 
@@ -47,9 +49,15 @@ class ImageDicomFilePT(ImageDicomFile):
     def create(self):
         return self
 
-    def load_data(self, **kwargs):
+    def load_data(
+            self,
+            pet_suv_conversion: str = "body_weight",
+            **kwargs
+    ):
         image_data = self.load_data_generic()
 
+        # Perform decay correction
+        image_data *= self._get_administration_decay_factor()
         # TODO: integrate SUV computations locally.
         suv_conversion_object = SUVscalingObj(dcm=self.image_metadata)
         scale_factor = suv_conversion_object.get_scale_factor(suv_normalisation="bw")
@@ -386,3 +394,172 @@ class ImageDicomFilePT(ImageDicomFile):
 
         metadata.update(dict(dcm_meta_data))
         return metadata
+
+    def _get_administration_decay_factor(self):
+        self.load_metadata()
+
+        # TODO: Determine if any decay correction took place from (0018,9758), which has either YES or NO, if present.
+        #  If YES, (0018,9701) should be present as well.
+        # TODO: Use Decay Correction DateTime (0018,9701) as alternative for determining the correction time.
+
+        # TODO: cross-check with https://niftypet.readthedocs.io/en/latest/tutorials/corrqnt/#id1
+
+        # Type of decay correction that is used
+        decay_correction = get_pydicom_meta_tag(
+            dcm_seq=self.image_metadata,
+            tag=(0x0054, 0x1102),
+            tag_type="str",
+            default="NONE"
+        )
+        if decay_correction == "ADMIN":
+            return 1.0
+
+        elif decay_correction not in ["NONE", "START"]:
+            raise ValueError(
+                f"Decay correction DICOM tag in {self.file_path} was not recognised: {decay_correction}. One of ",
+                f"NONE, START or ADMIN was expected."
+            )
+
+        # Start of image acquisition for the current position
+        acquisition_start_date = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0008, 0x0022), tag_type="str")
+        acquisition_start_time = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0008, 0x0032), tag_type="str")
+        acquisition_ref_time = convert_dicom_time(date_str=acquisition_start_date, time_str=acquisition_start_time)
+
+        # Frame reference time frame (ms)
+        frame_duration = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0018, 0x1242), tag_type="float")
+        frame_reference_time = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0054, 0x1300), tag_type="float")
+
+        # Set radiotracer administration reference time.
+        radio_admin_ref_time = None
+
+        # Set radiotracer administration reference time from radionuclide administration
+        if get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0054, 0x0016), test_tag=True):
+            radio_admin_start_time = get_pydicom_meta_tag(
+                dcm_seq=self.image_metadata[0x0054, 0x0016][0],
+                tag=(0x0018, 0x1078),
+                tag_type="str"
+            )
+            radio_admin_ref_time = convert_dicom_time(datetime_str=radio_admin_start_time)
+
+            if radio_admin_ref_time is None:
+                # If unsuccessful, attempt determining administration time from (0x0018, 0x1072)
+                radio_admin_start_time = get_pydicom_meta_tag(
+                    dcm_seq=self.image_metadata[0x0054, 0x0016][0],
+                    tag=(0x0018, 0x1072),
+                    tag_type="str"
[…]
