### 2024-04-11 2d160f7
Do not allow WIP decay correction to run.
#75
