### 2024-04-12 979e4ca
WIP on integrating SUV computations into ImageDicomFilePT
