+++ mirp/_data_import/dicom_file_pet.py
@@ -56,7 +56,28 @@ class ImageDicomFilePT(ImageDicomFile):
     ):
         image_data = self.load_data_generic()
 
-        # Perform decay correction
+        # Get decay correction factor
+        try:
+            decay_factor = self._get_administration_decay_factor()
+        except ValueError as err:
+            warnings.warn(
+                f"Decay correction factor could not be determined. {str(err)}",
+                UserWarning
+            )
+            decay_factor = 1.0
+
+        # Get conversion factor to BQML
+        try:
+            bqml_factor = self._get_pet_unit_conversion_factor()
+        except ValueError as err:
+            warnings.warn(
+                f"BQML conversion factor could not be determined. {str(err)}",
+                UserWarning
+            )
+
+        # Get SUV conversion factor.
+        suv_factor = self._get_suv_conversion_factor(new_suv_type=pet_suv_conversion)
+
         # image_data *= self._get_administration_decay_factor()
         # TODO: integrate SUV computations locally.
         suv_conversion_object = SUVscalingObj(dcm=self.image_metadata)
@@ -395,14 +416,68 @@ class ImageDicomFilePT(ImageDicomFile):
         metadata.update(dict(dcm_meta_data))
         return metadata
 
-    def _get_administration_decay_factor(self):
+    def _get_tracer_administration_time(self) -> datetime.datetime:
         self.load_metadata()
 
-        # TODO: Determine if any decay correction took place from (0018,9758), which has either YES or NO, if present.
-        #  If YES, (0018,9701) should be present as well.
-        # TODO: Use Decay Correction DateTime (0018,9701) as alternative for determining the correction time.
+        # Set initial value of tracer administration reference time.
+        admin_ref_time = None
 
-        # TODO: cross-check with https://niftypet.readthedocs.io/en/latest/tutorials/corrqnt/#id1
+        # Administration time should come from the Radiopharmaceutical Information Sequence (0x0054, 0x0016).
+        has_sequence = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0054, 0x0016), test_tag=True)
+
+        # Prefer Radiopharmaceutical Start DateTime (0x0018, 0x1078)
+        if has_sequence and admin_ref_time is None:
+            admin_ref_time = get_pydicom_meta_tag(
+                dcm_seq=self.image_metadata[0x0054, 0x0016][0],
+                tag=(0x0018, 0x1078),
+                tag_type="str"
+            )
+            admin_ref_time = convert_dicom_time(datetime_str=admin_ref_time)
+
+        # Fallback to Radiopharmaceutical Start Time (0x0018, 0x1072)
+        if has_sequence and admin_ref_time is None:
+            admin_ref_time = get_pydicom_meta_tag(
+                dcm_seq=self.image_metadata[0x0054, 0x0016][0],
+                tag=(0x0018, 0x1072),
+                tag_type="str"
+            )
+
+            if admin_ref_time is not None:
+                # Infer start date.
+                acquisition_start_time = self._get_acquisition_start_time()
+                admin_ref_time = datetime.datetime(
+                    year=acquisition_start_time.year,
+                    month=acquisition_start_time.month,
+                    day=acquisition_start_time.day,
+                    hour=int(admin_ref_time[0:2]),
+                    minute=int(admin_ref_time[2:4]),
+                    second=int(admin_ref_time[4:6]),
+                    microsecond=0 if len(admin_ref_time) <= 6 else int(round(float(admin_ref_time[6:]) * 1000))
+                )
+
+                # Correct for overnight recordings.
+                if admin_ref_time > acquisition_start_time:
+                    admin_ref_time -= datetime.timedelta(days=(acquisition_start_time - admin_ref_time).days)
+
[…]
