+++ mirp/_data_import/dicom_file_pet.py
@@ -1,11 +1,8 @@
 import warnings
-
 import numpy as np
 import datetime
-
 from typing import Any
 
-from mirp._data_import.suv import SUVscalingObj
 from mirp._data_import.dicom_file import ImageDicomFile
 from mirp._data_import.utilities import parse_image_correction, convert_dicom_time, get_pydicom_meta_tag
 
@@ -74,20 +71,13 @@ class ImageDicomFilePT(ImageDicomFile):
                 f"BQML conversion factor could not be determined. {str(err)}",
                 UserWarning
             )
+            bqml_factor = 1.0
 
         # Get SUV conversion factor.
         suv_factor = self._get_suv_conversion_factor(new_suv_type=pet_suv_conversion)
 
-        # image_data *= self._get_administration_decay_factor()
-        # TODO: integrate SUV computations locally.
-        suv_conversion_object = SUVscalingObj(dcm=self.image_metadata)
-        scale_factor = suv_conversion_object.get_scale_factor(suv_normalisation="bw")
-
-        # Convert to SUV
-        image_data *= scale_factor
-
-        # Update relevant tags in the metadata
-        self.image_metadata = suv_conversion_object.update_dicom_header(dcm=self.image_metadata)
+        # Update image_data
+        image_data *= decay_factor * bqml_factor * suv_factor
 
         # Set image data.
         self.image_data = image_data
@@ -665,6 +655,7 @@ class ImageDicomFilePT(ImageDicomFile):
         translation_table = dict([
             ("none", "none"),
             ("BW", "body_weight"),
+            ("BSA", "body_surface_area"),
             ("LBM", "lean_body_mass_error"),
             ("LBMJAMES128", "lean_body_mass"),
             ("LBMJANMA", "lean_body_mass_bmi"),
@@ -687,6 +678,148 @@ class ImageDicomFilePT(ImageDicomFile):
         return revert_suv_factor * suv_factor
 
     def _compute_suv_factor(self, suv_type: str) -> float:
-        # TODO: update.
 
-        return 1.0
+        # No SUV -------------------------------------------------------------------------------------------------------
+        if suv_type == "none":
+            return 1.0
+
+        # Require body weight and administered dose.
+        patient_weight = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0010, 0x1030), tag_type="float")
+        if patient_weight is None:
+            raise ValueError(
+                f"Patient weight (0x0010, 0x1030) was missing in {self.file_path}. SUV normalisation is not possible."
+            )
+        elif patient_weight <= 0.0:
+            raise ValueError(
+                f"Patient weight (0x0010, 0x1030) was not positive ({patient_weight}) in {self.file_path}. SUV "
+                f"normalisation is not possible."
+            )
+        elif patient_weight >= 1000.0:
+            # Weight is likely provide in grams, not kilograms. Convert to kg.
+            patient_weight /= 1000.0
+
+        # Administered dose should come from the Radiopharmaceutical Information Sequence (0x0054, 0x0016).
+        administered_dose = None
+        has_sequence = get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0054, 0x0016), test_tag=True)
+        if has_sequence and administered_dose is None:
+            administered_dose = get_pydicom_meta_tag(
+                dcm_seq=self.image_metadata[0x0054, 0x0016][0],
+                tag=(0x0018, 0x1074),
+                tag_type="float"
+            )
+
+        if administered_dose is None:
+            raise ValueError(
+                f"Radionuclide Total Dose (0x0018, 0x1074) was missing in {self.file_path}. SUV normalisation is not "
+                f"possible."
+            )
+        elif administered_dose <= 0.0:
[…]
