+++ mirp/_workflows/standardWorkflow.py
@@ -86,7 +86,8 @@ class StandardWorkflow(BaseWorkflow):
         # Read image and masks.
         image, masks = read_image_and_masks(
             self.image_file,
-            to_numpy=False
+            to_numpy=False,
+            pet_suv_conversion=self.settings.post_process.suv_conversion_type
         )
 
         if masks is None or len(masks) == 0:
