### 2024-04-17 dd7b064
SUV values after BSA-correction are now closer to the expected value
According to Sugawara et al. (https://doi.org/10.1148/radiology.213.2.r99nv37521) SUVbsa is around 25-50 times less than typical SUVbw. That seems to fit the current values after conversion from kg to gram.
