+++ test/pet_suv_conversion_test.py
@@ -53,3 +53,54 @@ def test_pet_suv_conversion_body_surface_area():
     assert len(feature_data) == 1
     assert 0.195 < feature_data["stat_max"].values[0] < 0.200
     assert 0.004 < feature_data["stat_min"].values[0] < 0.005
+
+
+def test_pet_suv_conversion_lean_body_mass():
+
+    data = extract_features(
+        image=os.path.join(CURRENT_DIR, "data", "sts_images_raw", "STS_001", "FDG_PET", "image"),
+        mask=os.path.join(CURRENT_DIR, "data", "sts_images_raw", "STS_001", "FDG_PET", "mask"),
+        pet_suv_conversion="lean_body_mass",
+        roi_name="GTV_Mass",
+        base_feature_families="statistics"
+    )
+
+    feature_data = data[0]
+
+    assert len(feature_data) == 1
+    assert 6.0 < feature_data["stat_max"].values[0] < 6.1
+    assert 0.1 < feature_data["stat_min"].values[0] < 0.2
+
+
+def test_pet_suv_conversion_lean_body_mass_bmi():
+
+    data = extract_features(
+        image=os.path.join(CURRENT_DIR, "data", "sts_images_raw", "STS_001", "FDG_PET", "image"),
+        mask=os.path.join(CURRENT_DIR, "data", "sts_images_raw", "STS_001", "FDG_PET", "mask"),
+        pet_suv_conversion="lean_body_mass_bmi",
+        roi_name="GTV_Mass",
+        base_feature_families="statistics"
+    )
+
+    feature_data = data[0]
+
+    assert len(feature_data) == 1
+    assert 5.9 < feature_data["stat_max"].values[0] < 6.0
+    assert 0.1 < feature_data["stat_min"].values[0] < 0.2
+
+
+def test_pet_suv_conversion_ideal_body_weight():
+
+    data = extract_features(
+        image=os.path.join(CURRENT_DIR, "data", "sts_images_raw", "STS_001", "FDG_PET", "image"),
+        mask=os.path.join(CURRENT_DIR, "data", "sts_images_raw", "STS_001", "FDG_PET", "mask"),
+        pet_suv_conversion="ideal_body_weight",
+        roi_name="GTV_Mass",
+        base_feature_families="statistics"
+    )
+
+    feature_data = data[0]
+
+    assert len(feature_data) == 1
+    assert 6.7 < feature_data["stat_max"].values[0] < 6.8
+    assert 0.1 < feature_data["stat_min"].values[0] < 0.2
