### 2024-04-17 72add43
Prevent unnecessary copying if images and masks are on the file system anyway.
