### 2024-04-17 da9915e
Remove metadata after importing in importing images and masks.
