### 2024-04-17 cfede90
Add remove_metadata override method for file stacks.
