+++ mirp/_data_import/generic_file_stack.py
@@ -187,6 +187,10 @@ class ImageFileStack(ImageFile):
         for image_file_object in self.image_file_objects:
             image_file_object.load_metadata()
 
+    def remove_metadata(self):
+        for image_file_object in self.image_file_objects:
+            image_file_object.remove_metadata()
+
     def load_data(self, **kwargs):
         # Load data for underlying files in the order indicated by self.image_file_objects.
         for image_file_object in self.image_file_objects:
