### 2024-04-17 2a40bf5
Do not copy unnecessary data in dicom_file_rtstruct.to_object
