### 2024-04-17 256e09c
Add limited and include_image arguments to load_metadata.
