+++ mirp/_data_import/numpy_file.py
@@ -115,7 +115,7 @@ class ImageNumpyFile(ImageFile):
     def create(self):
         return self
 
-    def load_metadata(self):
+    def load_metadata(self, **kwargs):
         if self.image_metadata is not None:
             return
 
