### 2024-04-17 58deb0b
Load metadata when a tag is required for completing the file.
