### 2024-04-17 c932bed
Complete method no longer forces loading metadata, but delays this until required by one of its _complete_* methods.
