### 2024-04-17 7aca535
DICOM files are no longer read twice in load_data_generic.
