+++ mirp/_data_import/dicom_file.py
@@ -331,8 +331,8 @@ class ImageDicomFile(ImageFile):
 
             dimensions = tuple([
                 1,
-                get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0028, 0x010), tag_type="int"),
-                get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0028, 0x011), tag_type="int")
+                get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0028, 0x0010), tag_type="int"),
+                get_pydicom_meta_tag(dcm_seq=self.image_metadata, tag=(0x0028, 0x0011), tag_type="int")
             ])
 
             self.image_dimension = dimensions
