### 2024-04-17 8d03a50
Populate _get_limited_metadata_tags
