+++ mirp/_data_import/dicom_file_rtstruct.py
@@ -99,6 +99,14 @@ class MaskDicomFileRTSTRUCT(MaskDicomFile):
     def load_data(self, **kwargs):
         pass
 
+    @staticmethod
+    def _get_limited_metadata_tags():
+        tags = super()._get_limited_metadata_tags()
+
+        tags += [
+            (0x3006, 0x0020)  # Structure set roi sequence
+        ]
+
     def to_object(
             self,
             image: None | ImageFile,
