### 2024-04-17 d3e667a
Add self because of reference to super()
