### 2024-04-17 ee4ed65
Read dicom depending on limited.
Limited reads only specific tags.
