### 2024-04-18 fb8908f
Ensure that metadata are loaded for itk and numpy files.
