+++ mirp/_data_import/numpy_file.py
@@ -95,21 +95,25 @@ class ImageNumpyFile(ImageFile):
 
     def _complete_image_origin(self, force=False):
         if self.image_origin is None:
+            self.load_metadata()
             n_dim = len(self.image_metadata.shape)
             self.image_origin = tuple([0.0] * n_dim)
 
     def _complete_image_orientation(self, force=False):
         if self.image_orientation is None:
+            self.load_metadata()
             n_dim = len(self.image_metadata.shape)
             self.image_orientation = np.identity(n_dim, dtype=float)
 
     def _complete_image_spacing(self, force=False):
         if self.image_spacing is None:
+            self.load_metadata()
             n_dim = len(self.image_metadata.shape)
             self.image_spacing = tuple([1.0] * n_dim)
 
     def _complete_image_dimensions(self, force=False):
         if self.image_dimension is None:
+            self.load_metadata()
             self.image_dimension = tuple(self.image_metadata.shape)
 
     def create(self):
