### 2024-04-18 #6413 @emdupre
👋  Hi everyone, thanks for the updates !

I just wanted to summarize status:

@theanega : I see that you've completed your checklist, but you've also created issue https://github.com/oncoray/mirp/issues/69 which is still open. Could you please confirm the status of your review ?

@Matthew-Jennings : I see that you've not yet completed your checklist and that you've created the following open issues https://github.com/oncoray/mirp/issues/72, https://github.com/oncoray/mirp/issues/73, and https://github.com/oncoray/mirp/issues/75. I assume you've currently finished the initial review and are now waiting on the resolution of these issues, but if you're instead still working on the initial review,please let me know.

@drcandacemakedamoore  and @surajpaib : I know that you were not expecting to finish the initial review until this week. Please let us know when you are able to do so !

I'll also cross-link all of the other associated issues to date, in case this helps in finalizing initial reviews:
- https://github.com/oncoray/mirp/issues/54
- https://github.com/oncoray/mirp/issues/55
- https://github.com/oncoray/mirp/issues/62
- https://github.com/oncoray/mirp/issues/63
- https://github.com/oncoray/mirp/issues/66
- https://github.com/oncoray/mirp/issues/67
- https://github.com/oncoray/mirp/issues/68
- https://github.com/oncoray/mirp/issues/76
