### 2024-04-19 3ff94f3
Merge _get_limited_metada_tags into the main DICOM file classes.
pydicom will read the header entirely anyway, so its cheaper to read it as few times as possible.
