### 2024-04-19 f1000d3
Reduce the number of times DICOM metadata need to be read from file.
This now happens only twice:
1. a small metadata subset for selecting images and masks
2. full metadata and image during the main workflow.
