### 2024-04-19 28a4315
Prevent unnecessary loading of image metadata in to_object methods of mask files.
