### 2024-04-19 4befadb
Added metadata tags required to process RTDose files.
