+++ docs/_sources/tutorial_compute_radiomics_features_mr.ipynb.txt
@@ -1,0 +1,642 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "1e447b05-08e1-4326-9709-531acf639a69",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "# Tutorial: Computing radiomics features\n",
+    "\n",
+    "This tutorial describes step-by-step how you can compute radiomics features from medical imaging using MIRP.\n",
+    "\n",
+    "## Download example data\n",
+    "\n",
+    "We have prepared an (anonymised) example dataset of T1-weighted images of patients with soft-tissue sarcoma. This is a zip-file containing data for three patients."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "id": "3d8f91a9-e514-47bc-8c9b-5e37fb8662e5",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from urllib.request import urlopen\n",
+    "from io import BytesIO\n",
+    "from zipfile import ZipFile\n",
+    "\n",
+    "url = r\"https://github.com/oncoray/mirp/raw/acd0b11521acd07aa9597e8f55d480bc54af4e04/data/tutorial_radiomics_mr_t1_data.zip\"\n",
+    "\n",
+    "# Specify location where the data is stored.\n",
+    "save_dir = \".\"\n",
+    "\n",
+    "with urlopen(url) as zip_url_pointer:\n",
+    "    with ZipFile(BytesIO(zip_url_pointer.read())) as example_data:\n",
+    "        example_data.extractall(save_dir)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "1ac63cb7-a935-4e43-a6ce-0c109d9bc10b",
+   "metadata": {},
+   "source": [
+    "This creates a folder with the following structure:\n",
+    "\n",
+    "    sts_images\n",
+    "    ├─ STS_001\n",
+    "    │   └─ ...\n",
+    "    ├─ ...\n",
+    "    └─ STS_003\n",
+    "        └─ mr_t1\n",
+    "            ├─ image\n",
+    "            |  ├─ 1-01.dcm\n",
+    "            |  ├─ ...\n",
+    "            |  └─ 1-36.dcm\n",
+    "            └─ mask\n",
+    "                └─ 1-1.dcm\n",
+    "\n",
+    "In this example dataset, the T1-weighted images are stored in DICOM format in the `mr_t1/image` subdirectory. A segmentation mask is also stored in DICOM format, but in the `mr_t1/mask` subdirectory.\n",
+    "\n",
+    "## Finding mask labels\n",
+    "\n",
+    "Radiomics features are typically computed from regions of interest, such as a tumour. These regions are delineated by experts or auto-segmentation AI, and stored as segmentation masks. MIRP needs to know which mask label (region of interest) should be used for computing features. A first step is to identify which mask labels exist. This can be done using the `extract_mask_labels` function. We need to define the main directory (`.../sts_images`), and optionally the directory structure relative to the patient directory (`mr_t1/mask`), where the masks can be found."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "id": "a89c4b13-6fc5-4c72-8f67-2048891b23fa",
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "<div>\n",
+       "<style scoped>\n",
[…]
