+++ docs/_sources/image_mask_import.rst.txt
@@ -1,0 +1,301 @@
+Configure image and mask import
+===============================
+
+Many relevant MIRP functions require images, masks or both as input. This section provides details on how image and
+mask import is configured.
+
+Specifying input
+----------------
+
+MIRP processes and analyses images and masks. There are multiple ways to provide images and masks:
+
+* By specifying the directory where images and/or masks are found:
+    * **Nested flat layout**: In a nested flat layout all images and masks are separated for each sample. For
+      example, an image dataset of 128 samples may be organised as follows::
+
+        image_root_directory
+        ├─ sample_001
+        │   └─ ...
+        ├─ ...
+        └─ sample_127
+            └─ image_sub_folder
+            ├─ CT_dicom_000.dcm
+            ├─ ...
+            └─ CT_dicom_255.dcm
+            └─ mask.dcm
+
+      Images and mask files are directly under the sample directory. Only one keyword argument is required:
+
+      .. code-block:: python
+
+          some_function(
+              ...,
+              image = "image_root_directory",
+              ...
+          )
+
+      MIRP is generally able to determine which files are images and which files are masks. However, there may be
+      cases where MIRP is unable to determine if a file is an image or a mask. In those cases, additional keyword
+      arguments may be provided:
+
+      .. code-block:: python
+
+          some_function(
+              ...,
+              image = "image_root_directory",
+              image_name = "CT_dicom_*",
+              mask_name = "mask"
+              ...
+          )
+
+      Here, ``image_name`` and ``mask_name`` contain patterns for image and mask files, respectively. ``"CT_dicom_*"``
+      contains a wildcard character (``*``) that matches any pattern starting with ``"CT_dicom_"``. File extensions are
+      never of the pattern.
+
+    * **Fully nested structure**: In a nested structure all images and masks are separated for each. Unlike the above
+      example, image and mask data may be organised into different subdirectory structures::
+
+        image_root_directory
+        ├─ sample_001
+        │   └─ image_sub_folder
+        │   │   └─ ...
+        │   └─ mask_sub_folder
+        │       └─ ...
+        ├─ ...
+        └─ sample_127
+            └─ image_sub_folder
+            │   ├─ CT_dicom_000.dcm
+            │   ├─ ...
+            │   └─ CT_dicom_255.dcm
+            └─ mask_sub_folder
+                └─ mask.dcm
+
+      Here the directory for each sample contains consistently named subdirectory structures (``image_sub_folder``
+      and ``mask_sub_folder``), that contains the set of DICOM images and a mask, respectively. Then the following
+      keyword arguments may be specified:
+
+      .. code-block:: python
+
+          some_function(
+              ...,
+              image = "image_root_directory",
+              image_sub_folder = "image_sub_folder",
+              mask_sub_folder = "mask_sub_folder",
+              ...
+          )
+
+      The ``mask`` keyword argument is automatically assumed to be equal to ``image``, i.e. images and masks are
+      under the same root directory. If this is not the case, ``mask`` should be specified as well.
+
+      .. note::
+        MIRP will interpret the name of the directory that is neither part of the root directory or the subdirectory
+        structures as the sample name, unless the sample name can be determined from metadata (i.e. DICOM files). In
+        the example above, sample names based on the directory structure would be ``"sample_001"`` to ``"sample_127"``.
+
+    * **Flat layout**: In a flat layout, all image and mask files are contained in the same directory::
+
+        image_root_directory
+            ├─ sample_001_CT_dicom_000.dcm
+            ├─ ...
+            ├─ sample_001_CT_dicom_319.dcm
+            ├─ sample_127_CT_dicom_000.dcm
+            ├─ ...
+            ├─ sample_127_CT_dicom_255.dcm
[…]
