### 2024-04-19 oncoray/mirp#73 @alexzwanenburg
In version 2.2.1 I was able to address some of the suggestions. Some things still need to be addressed:

- [ ] Add statement of need (or at least make it clearer).
- [ ] Add additional tutorials, e.g. on using filters.
- [ ] Document design concepts for contributors:
  - [ ] Importing images and files.
  - [ ] Internal image and mask formats.
  - [ ] Defining workflows.
  - [ ] Feature classes.
  - [ ] Filter classes.
  - [ ] Designing tests.

@drcandacemakedamoore I hope we will be able to grow the community around MIRP 😃
