### 2024-04-19 oncoray/mirp#75 @alexzwanenburg
I made some progress in coverage with version 2.2.1. I am leaving this issue open as I need to create new test data to test remaining parts of the code.
