### 2024-04-19 #6413 @alexzwanenburg
I released version 2.2.1. This also includes updates to the documentation and a new tutorial: https://oncoray.github.io/mirp/tutorial_compute_radiomics_features_mr.html

I am looking forward to your feedback on these updates.
