### 2024-04-22 ade2fd2
Update NEWS.md
