### 2024-04-22 #6413 @surajpaib
Hi @emdupre I've opened an issue on the MIRP repo with more of my observations and with that I'm done with my initial review.

Much thanks for your patience
