### 2024-04-23 d30ee7f
Match xml branch names with configuration settings. #81
