### 2024-04-24 253bda1
Mask modality is now reported in extract_mask_labels. #81
