### 2024-04-24 bdb16f8
Deprecate the use of n_labels.
