+++ mirp/_data_import/dicom_file_seg.py
@@ -502,7 +502,7 @@ class MaskDicomFileSEG(MaskDicomFile):
             labels = [None]
 
         # Get general attributes.
-        parent_attributes = self._get_export_attributes(n_labels=n_labels)
+        parent_attributes = self._get_export_attributes()
 
         # Add roi labels as attribute.
         attributes = [("roi_label", labels)]
