### 2024-04-24 1f89d48
Define default DICOM attributes to export. #81
