### 2024-04-24 12cf61f
Fully document arguments in extract_features, extract_images and similar #81
