+++ mirp/extract_features_and_images.py
@@ -11,6 +11,7 @@ def extract_features(
         write_features: None | bool = None,
         export_features: None | bool = None,
         write_dir: None | str = None,
+        num_cpus: None | int = None,
         **kwargs
 ) -> None | list[Any]:
     """
@@ -30,8 +31,25 @@ def extract_features(
         Path to directory where feature tables should be written. If not set, feature tables are returned by this
         function. Required if ``write_features=True``.
 
+    num_cpus: int, optional, default: None
+        Number of CPU nodes that should be used for parallel processing. Image processing and feature computation can be
+        parallelized using the ``ray`` package. If a ray cluster is defined by the user, this cluster will be used
+        instead. By default, images are processed sequentially.
+
     **kwargs:
-        Keyword arguments passed to :func:`~mirp.extractFeaturesAndImages.extract_features_and_images`.
+        Keyword arguments passed for importing images and masks (
+        :func:`mirp.data_import.import_image_and_mask.import_image_and_mask`) and configuring settings:
+
+        * general settings (:class:`~mirp.settings.general_parameters.GeneralSettingsClass`)
+        * image post-processing (:class:`~mirp.settings.image_processing_parameters.ImagePostProcessingClass`)
+        * image perturbation / augmentation (:class:`~mirp.settings.perturbation_parameters.ImagePerturbationSettingsClass`)
+        * image interpolation / resampling (
+          :class:`~mirp.settings.interpolation_parameters.ImageInterpolationSettingsClass` and
+          :class:`~mirp.settings.interpolation_parameters.MaskInterpolationSettingsClass`)
+        * mask resegmentation (:class:`~mirp.settings.resegmentation_parameters.ResegmentationSettingsClass`)
+        * image transformation (:class:`~mirp.settings.transformation_parameters.ImageTransformationSettingsClass`)
+        * feature computation / extraction (
+          :class:`~mirp.settings.feature_parameters.FeatureExtractionSettingsClass`)
 
     Returns
     -------
@@ -49,6 +67,7 @@ def extract_features(
         write_images=False,
         export_images=False,
         write_dir=write_dir,
+        num_cpus=num_cpus,
         **kwargs
     )
 
@@ -72,7 +91,19 @@ def extract_features_generator(
         Determines whether features computed from images should be returned by the function.
 
     **kwargs:
-        Keyword arguments passed to :func:`~mirp.extractFeaturesAndImages.extract_features_and_images_generator`.
+        Keyword arguments passed for importing images and masks (
+        :func:`mirp.data_import.import_image_and_mask.import_image_and_mask`) and configuring settings:
+
+        * general settings (:class:`~mirp.settings.general_parameters.GeneralSettingsClass`)
+        * image post-processing (:class:`~mirp.settings.image_processing_parameters.ImagePostProcessingClass`)
+        * image perturbation / augmentation (:class:`~mirp.settings.perturbation_parameters.ImagePerturbationSettingsClass`)
+        * image interpolation / resampling (
+          :class:`~mirp.settings.interpolation_parameters.ImageInterpolationSettingsClass` and
+          :class:`~mirp.settings.interpolation_parameters.MaskInterpolationSettingsClass`)
+        * mask resegmentation (:class:`~mirp.settings.resegmentation_parameters.ResegmentationSettingsClass`)
+        * image transformation (:class:`~mirp.settings.transformation_parameters.ImageTransformationSettingsClass`)
+        * feature computation / extraction (
+          :class:`~mirp.settings.feature_parameters.FeatureExtractionSettingsClass`)
 
     Returns
     -------
@@ -97,6 +128,8 @@ def extract_images(
         write_images: None | bool = None,
[…]
