### 2024-04-24 9ab1912
Fixed code generating warnings due to numpy matrix subclass.
