### 2024-04-24 2cbc303
Added tests for modality-specific functions #75
