### 2024-04-24 1d37c06
Added initial tests for image file descriptors #75
