### 2024-04-25 6a96910
Added image descriptor to image and mask import errors and warnings.
#81
