### 2024-04-26 01c5363
Added hyperlink to paper references in documentation.
