+++ docs_source/source/introduction.rst
@@ -29,16 +29,18 @@ References
 ----------
 
 .. [Lambin2017] Lambin P, Leijenaar RTH, Deist TM, Peerlings J, de Jong EEC, van Timmeren J, et al. Radiomics: the
-  bridge between medical imaging and personalized medicine. Nat Rev Clin Oncol. 2017;14: 749–762.
-  doi:10.1038/nrclinonc.2017.141
+  bridge between medical imaging and personalized medicine. Nat Rev Clin Oncol. 2017;14: 749-762.
+  doi:`10.1038/nrclinonc.2017.141 <https://doi.org/10.1038/nrclinonc.2017.141>`_
 
 .. [vanTimmeren2020] van Timmeren JE, Cester D, Tanadini-Lang S, Alkadhi H, Baessler B. Radiomics in medical
-  imaging-“how-to” guide and critical reflection. Insights Imaging. 2020;11: 91. doi:10.1186/s13244-020-00887-2
+  imaging-“how-to” guide and critical reflection. Insights Imaging. 2020;11: 91.
+  doi:`10.1186/s13244-020-00887-2 <https://doi.org/10.1186/s13244-020-00887-2>`_
 
 .. [Zwanenburg2020] Zwanenburg A, Vallieres M, Abdalah MA, Aerts HJWL, Andrearczyk V, Apte A, et al. The Image
   Biomarker Standardization Initiative: Standardized Quantitative Radiomics for High-Throughput Image-based
-  Phenotyping. Radiology. 2020;295: 328-338. doi:10.1148/radiol.2020191145
+  Phenotyping. Radiology. 2020;295: 328-338.
+  doi:`10.1148/radiol.2020191145 <https://doi.org/10.1148/radiol.2020191145>`_
 
 .. [Whybra2024] Whybra P, Zwanenburg A, Andrearczyk V, Schaer R, Apte AP, Ayotte A, et al. The Image Biomarker
   Standardization Initiative: Standardized Convolutional Filters for Reproducible Radiomics and Enhanced Clinical
-  Insights. Radiology. 2024;310: e231319. doi:10.1148/radiol.231319
+  Insights. Radiology. 2024;310: e231319. doi:`10.1148/radiol.231319 <https://doi.org/10.1148/radiol.231319>`_
