### 2024-04-26 587007d
Update image_descriptors_test.py
#75
