+++ docs_source/source/design.rst
@@ -1,0 +1,4 @@
+General design
+==============
+
+Here we describe the overall design of MIRP.
