### 2024-05-02 46b3795
Added template for documentation issues.
