+++ .github/ISSUE_TEMPLATE/documentation.md
@@ -1,0 +1,19 @@
+---
+name: Documentation
+about: Template for suggesting documentation changes
+title: ''
+labels: documentation
+assignees: ''
+
+---
+
+Please use the below template to submit your issue:
+
+# Documentation changes
+
+What new addition or change to the documentation would you like to see?
+Where should this addition or change be made? If relevant, provide the name of the method, class or function.
+
+# Specifications
+
+- MIRP version: 
