### 2024-05-02 91ad773
Added hint for finding the installed MIRP version.
