### 2024-05-02 02ceb26
Add automated tests that run on the first and fifteenth day of each month
