+++ .github/workflows/auto-test-dependencies_timed.yml
@@ -1,0 +1,36 @@
+# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
+# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
+
+# Automated tests are run at 6:30 UTC on the first and fifteenth day of each month.
+
+name: Run automated tests on multiple platforms to detect dependency issues
+
+on:
+  workflow_dispatch:
+  schedule:
+    - cron: '30 6 1,15 * *'
+
+jobs:
+  build:
+
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: [ "3.10", "3.11"]
+        os: [ ubuntu-latest, macos-latest, windows-latest ]
+
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install .
+        python -m pip install .[test]
+    - name: Test with pytest
+      run: |
+        python -m pytest -m ci test
