### 2024-05-03 73b71ad
Work on design documentation.
