### 2024-05-05 32339af
Extended description of design.
