+++ docs_source/source/design.rst
@@ -73,15 +73,30 @@ Features
 Feature computation is called from `StandardWorkflow._compute_radiomics_features`. At the moment feature computation is
 mostly functional, and organised by feature family.
 
-Future Implementations
-^^^^^^^^^^^^^^^^^^^^^^
+Future Directions
+^^^^^^^^^^^^^^^^^
 Feature computation is currently family-based, i.e. an entire family of features is computed at once. It may be
 preferable to move to a more directive-based approach by treating feature as an object. There are no current direct
-benefits to doing so, and would make feature computation a bit harder to program to avoid doing unnecessary work.
-However, such a system would allow for exporting features with added metadata, instead of just their values. These
-metadata could include IBSI identifiers of the features, as well as processing-related metadata that are currently
-included in the feature name. This would make it easier to export structured reports of image data.
+benefits to doing so, and would make feature computation a bit harder to program to avoid doing unnecessary work. For
+example, computation of a grey level co-occurrence matrix feature relies on two prior steps: discretisation and
+computing the GLCM. This would mean that feature objects have to be sorted by discretisation required, and the set of
+parameters.
+
+A directive-based approach would have at least the following two advantages:
+
+* Features can be exported with metadata, instead of just their values. These metadata could include their IBSI
+  identifiers of the features, as well as processing-related metadata that are currently included in the feature name.
+  This would make it easier to export structured reports of image data.
+
+* Features can be defined that are not generated by default, e.g. different intensity-volume histogram features.
 
 Filters
 -------
-WIP
+All filters are implemented as objects, defined in the `_imagefilters` submodule. The filters themselves are accessed
+in the `StandardWorkflow.transform_images` method, yielding specific transformed image objects (defined in
+`_images.transformed_images`).
+
+Future Directions
+^^^^^^^^^^^^^^^^^
+We are generally happy with the current implementation of image filters. It is relatively straightforward to implement new
+filters should there be a need.
