### 2024-05-05 6e288c2
Details on the internal image representation #73
