### 2024-05-06 0708220
Added descriptions of image and mask representations.
