+++ docs_source/source/index.rst
@@ -36,6 +36,7 @@ MIRP
 
    contributing
    design
+   contributing_tests
 
 Welcome to MIRP! If you need to quantitatively analyse medical images using a standardised workflow, you have come to
 the right place!
