### 2024-05-06 3092322
Added initial feature tables to documentation.
#85
