### 2024-05-07 1de0527
Added remaining features and discretisation methods.
