+++ docs_source/source/features_names.rst
@@ -173,8 +173,21 @@ Features computed from grey level co-occurrence matrices (GLCM) are:
 * `cm_info_corr1`: First measure of information correlation (`R8DG`)
 * `cm_info_corr2`: Second measure of information correlation (`JN9H`)
 
-GLCM-features are computed with the following parameters:
-distance, spatial method .. TODO..
+GLCM-features are computed with the following parameters, in sequential order:
+
+* distance:
+    * `d#.#`: Chebyshev distance for considering the neighbourhood for determining co-occurrence (`PVMT`). Typically
+      `d1.0`.
+* feature and matrix aggregation:
+    * `2d_avg`: features computed by averaging feature values of each 2D directional matrix across all directions and
+      slices (`BTW3`)
+    * `2d_s_mrg`: features computed by averaging feature values for each slice after merging 2D directional matrices
+      within that slice (`SUJT`)
+    * `2.5d_d_mrg`: features computed by averaging feature values for each direction after merging 2D directional
+      matrices corresponding to that direction (`JJUI`)
+    * `2.5d_v_mrg`: feature computed from a single matrix after merging all 2D directional matrices (`ZW7Z`)
+    * `3d_avg`: features computed by averaging feature values of each 3D directional matrix (`ITBB`)
+    * `3d_v_mrg`: features computed from a single matrix after merging all 3D directional matrices (`IAZD`)
 
 Grey level run length matrix features (`TP0I`)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -198,14 +211,144 @@ Features computed from grey level run length matrices (GLRLM) are:
 * `rlm_rl_var`: Run length variance (`SXLW`)
 * `rlm_rl_entr`: Run entropy (`HJ9O`)
 
-GLRLM features are computed with the following parameters:
-distance, spatial method .. TODO..
+GLRLM features are computed with the following parameter:
+
+* feature and matrix aggregation:
+    * `2d_avg`: features computed by averaging feature values of each 2D directional matrix across all directions and
+      slices (`BTW3`)
+    * `2d_s_mrg`: features computed by averaging feature values for each slice after merging 2D directional matrices
+      within that slice (`SUJT`)
+    * `2.5d_d_mrg`: features computed by averaging feature values for each direction after merging 2D directional
+      matrices corresponding to that direction (`JJUI`)
+    * `2.5d_v_mrg`: feature computed from a single matrix after merging all 2D directional matrices (`ZW7Z`)
+    * `3d_avg`: features computed by averaging feature values of each 3D directional matrix (`ITBB`)
+    * `3d_v_mrg`: features computed from a single matrix after merging all 3D directional matrices (`IAZD`)
 
 Grey level size zone matrix features (`9SAK`)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Features computed from grey level size zone matrices (GLSZM) are:
 
+* `szm_sze`: Small zone emphasis (`5QRC`)
+* `szm_lze`: Large zone emphasis (`48P8`)
+* `szm_lgze`: Low grey level zone emphasis (`XMSY`)
+* `szm_hgze`: High grey level zone emphasis (`5GN9`)
[…]
