+++ docs_source/source/features_names.rst
@@ -124,19 +124,19 @@ Intensity-volume histogram features (`P88C`)
 Intensity volume histogram features are listed below. Note that the IBSI reference standard provides a general
 definition of these features, whereas MIRP computes these features for specific values:
 
-* `ivh_v10`: Volume fraction at 10% intensity (`BC2M`)
+* `ivh_v10`: Volume fraction at 10% intensity (`BC2M`; `NK6P`)
 * `ivh_v25`: Volume fraction at 25% intensity (`BC2M`)
 * `ivh_v50`: Volume fraction at 50% intensity (`BC2M`)
 * `ivh_v75`: Volume fraction at 75% intensity (`BC2M`)
-* `ivh_v90`: Volume fraction at 90% intensity (`BC2M`)
-* `ivh_i10`: Intensity at 10% volume (`GBPN`)
+* `ivh_v90`: Volume fraction at 90% intensity (`BC2M`; `4279`)
+* `ivh_i10`: Intensity at 10% volume (`GBPN`; `PWN1`)
 * `ivh_i25`: Intensity at 25% volume (`GBPN`)
 * `ivh_i50`: Intensity at 50% volume (`GBPN`)
 * `ivh_i75`: Intensity at 75% volume (`GBPN`)
-* `ivh_i90`: Intensity at 90% volume (`GBPN`)
-* `ivh_diff_v10_v90`: Difference in volume fraction between 10% and 90% intensity (`DDTU`)
+* `ivh_i90`: Intensity at 90% volume (`GBPN`; `BOHI`)
+* `ivh_diff_v10_v90`: Difference in volume fraction between 10% and 90% intensity (`DDTU`; `WITY`)
 * `ivh_diff_v25_v75`: Difference in volume fraction between 25% and 75% intensity (`DDTU`)
-* `ivh_diff_i10_i90`: Difference in intensity between 10% and 90% volume (`CNV2`)
+* `ivh_diff_i10_i90`: Difference in intensity between 10% and 90% volume (`CNV2`; `JXJA`)
 * `ivh_diff_i25_i75`: Difference in intensity between 25% and 75% volume (`CNV2`)
 * `ivh_auc`: Area under IVH curve (`9CMM`; reference values absent)
 
@@ -349,6 +349,13 @@ These are then followed by a parameter specifying the number of bins or bin size
 
 * `w#.#`: Width of each bin for fixed bin size discretisation methods.
 * `n#`: Number of bins for the fixed bin number discretisation method.
+
+Filters
+-------
+
+Features can not only be computed from the base image, but also from filtered images.
+
+
 
 Examples
 --------
