### 2024-05-10 2c0b91b
Added some filter descriptors #85
