### 2024-05-14 6cb2bed
Added final transfomations.
Closes #85
