### 2024-05-15 834fce2
Add test to compare between RTSTRUCT and SEG files.
