### 2024-05-15 45b991d
Update NEWS.md
