### 2024-05-16 #6413 @Matthew-Jennings
Hi @alexzwanenburg, would you mind keeping these Issues open until we've re-reviewed?

In my mind, they're not closed until either the reviewers have confirmed the recommendations/suggestions have been met, or where they're not met, agreement has been reached that publication can proceed anyway.
