### 2024-05-16 4415048
Added enhancement proposal template.
