+++ .github/ISSUE_TEMPLATE/general-enhancement.md
@@ -1,0 +1,18 @@
+---
+name: New features and enhancements
+about: Enhancement proposal template
+title: ''
+labels: enhancement
+assignees: ''
+
+---
+
+New ideas or suggestions for improvements? Great! Please use the below template to propose your ideas:
+
+# Idea description
+
+What would you like MIRP to do?
+
+# Prior work
+
+Is there any relevant work (e.g. scientific literature) that can be used as reference? Do you know of data that can be used for testing?
