### 2024-05-16 oncoray/mirp#72 @alexzwanenburg
A second tutorial may be found here: https://oncoray.github.io/mirp/tutorial_apply_image_filter.html
