### 2024-05-16 oncoray/mirp#73 @alexzwanenburg
I added a statement of need here: https://oncoray.github.io/mirp/introduction.html#why-mirp

There are now two tutorials:

- https://oncoray.github.io/mirp/tutorial_compute_radiomics_features_mr.html
- https://oncoray.github.io/mirp/tutorial_apply_image_filter.html

The contributing section was extended:

- I updated https://oncoray.github.io/mirp/contributing.html
- I wrote https://oncoray.github.io/mirp/design.html to discuss the overall design of MIRP.
- There is now a document on tests: https://oncoray.github.io/mirp/contributing_tests.html
