+++ .github/ISSUE_TEMPLATE/general-question.md
@@ -1,0 +1,31 @@
+---
+name: Questions
+about: General question template
+title: ''
+labels: question
+assignees: ''
+
+---
+
+Do you have a question? Please use the below template to submit your issue:
+
+# Question
+
+What do you want to ask?
+
+# What led to the question?
+
+If the question is related to using MIRP, what were you trying to do? If possible and relevant, provide code and data.
+
+# Specifications
+
+- MIRP version (*if relevant*): 
+- Python version (*if relevant*):
+- Operating system (*if relevant*):
+- Other package versions (*if relevant*):
+
+*Hint*: the installed version of MIRP can be found using `importlib_metadata`:
+```
+import importlib_metadata
+importlib_metadata.version("mirp")
+```
