### 2024-05-21 edbb3fd
Enable passing metadata to the MIRP-native image format. #87
