### 2024-05-22 4f726d1
BaseImage and derivative classes now have and pass object_metadata attribute #87
