+++ mirp/_data_import/dicom_file_stack.py
@@ -230,6 +230,14 @@ class ImageDicomFileStack(ImageFileStack):
             "position_x": image_position_x
         }).sort_values(by=["position_z", "position_y", "position_x"], ignore_index=True)
 
+    def set_object_metadata(self):
+        """
+        Updates the object metadata that is passed to native image and mask classes in to_object. In this case,
+        object metadata is copied from the underlying slices.
+        """
+        self.image_file_objects[0].set_object_metadata()
+        self.object_metadata = self.image_file_objects[0].object_metadata
+
     def export_metadata(self) -> None | dict[str, Any]:
         metadata = super().export_metadata()
         additional_metadata = self.image_file_objects[0].export_metadata(only_self=True)
