+++ mirp/_data_import/dicom_file.py
@@ -449,14 +449,60 @@ class ImageDicomFile(ImageFile):
 
         return image_data
 
+    def set_object_metadata(self):
+        """
+        Updates the object metadata that is passed to native image and mask classes in to_object.
+        """
+        metadata = []
+
+        # Ensure that metadata are present.
+        self.load_metadata(limited=False)
+
+        # Study date
+        study_date = get_pydicom_meta_tag(
+            dcm_seq=self.image_metadata,
+            tag=(0x0008, 0x0020),
+            tag_type="str"
+        )
+
+        if study_date is not None and not study_date == "":
+            metadata += [("study_date", study_date)]
+
+        # Study description
+        study_description = get_pydicom_meta_tag(
+            dcm_seq=self.image_metadata,
+            tag=(0x0008, 0x1030),
+            tag_type="str"
+        )
+
+        if study_description is not None and not study_description == "":
+            metadata += [("study_description", study_description)]
+
+        # Series description
+        series_description = get_pydicom_meta_tag(
+            dcm_seq=self.image_metadata,
+            tag=(0x0008, 0x103E),
+            tag_type="str"
+        )
+
+        if series_description is not None and not series_description == "":
+            metadata += [("series_description", series_description)]
+
+        # Series instance UID
+        if self.series_instance_uid is not None:
+            metadata += [("series_instance_uid", self.series_instance_uid)]
+
+        # Update object_metadata
+        self.object_metadata.update(dict(metadata))
+
     @staticmethod
     def _get_limited_metadata_tags():
         # Limited tags are read to populate basic
         return [
             (0x0008, 0x0018),  # SOP instance UID
             (0x0008, 0x0060),  # modality
-            (0x0008, 0x1030),  # series description
-            (0x0008, 0x103E),  # study description
+            (0x0008, 0x1030),  # study description
+            (0x0008, 0x103E),  # series description
             (0x0010, 0x0010),  # patient name
             (0x0010, 0x0020),  # patient id
             (0x0018, 0x0050),  # slice thickness
