### 2024-05-22 88a74dc
Object metadata is now parsed into feature dataframes.
