### 2024-05-29 #6413 @emdupre
👋 Hi everyone !

I just wanted to follow-up on this and check how this re-review of _MIRP_ is going. @Matthew-Jennings, @drcandacemakedamoore, @surajpaib, and @theanega, please let me know if you have any questions or concerns on process at this point ! And thank you again for your work in reviewing 🌻
