### 2024-05-31 d2e60f0
Add checks on associated masks.
