### 2024-05-31 6d42060
Add file_type as object metadata.
