### 2024-05-31 31ec343
Warn for masks that become empty after registering in the image space.
