### 2024-05-31 bc1a2c9
Add checks for positioning of associated masks.
