### 2024-05-31 de6f76e
Check whether image space characteristics were set correctly.
