### 2024-05-31 b5d9904
Update NEWS.md
