+++ NEWS.md
@@ -5,6 +5,12 @@
 - Tables with feature values now contain extra columns. These columns specify the file name (for non-DICOM input), 
   the directory path of the image and masks and several DICOM tags for identifying the input.
 
+- MIRP now checks whether there are potential problems between the frames of reference of image and mask files.
+
+# Fixes
+
+- Fixed an error that occurs when attempting to create a deep copy `ImageITKFile` objects.
+
 # Version 2.2.2
 
 ## Minor changes
