### 2024-06-05 cdc092d
Update test to read the correct column.
