+++ test/basic_test.py
@@ -153,7 +153,7 @@ def test_xml_configurations(tmp_path):
     assert len(data) == 2
     assert all(data["sample_name"].values == ["STS_002", "STS_003"])
     assert all(data["image_modality"].values == "pet")
-    assert all(data["image_mask_name"].values == "GTV_Mass_PET")
+    assert all(data["image_mask_label"].values == "GTV_Mass_PET")
     assert all(data["image_voxel_size_x"].values == 3.0)
     assert all(data["image_voxel_size_y"].values == 3.0)
     assert all(data["image_voxel_size_z"].values == 3.0)
