### 2024-06-07 oncoray/mirp#73 @Matthew-Jennings
Thanks, @alexzwanenburg!

1. Statement of need looks good :) - **close**.

2. What are your thoughts on this point? I do think that it's not immediately obvious to a prospective user (reading the README, or even the docs intro) what high-level functionality MIRP contains. Example/tutorials are great, but a user will assume this only shows a curated subset of functionality. You do state that it's an IBSI compliant medical image analysis toolkit, but this feels _too_ broad. Probably the closest docs content to what I'm talking about is buried in your Contributing section under [submodules](https://oncoray.github.io/mirp/design.html#submodules)

3. I like your tutes! Add more when you have the time/think it's suitable, but two are enough for the purposes of JOSS publication - **close**.

4. Nice to see the contrib section. Pretty well fleshed out. - **close (but still consider bus factor and @drcandacemakedamoore's comments)**

5. NEW: It took me (embarrassingly?) a little while to see that your technical reference (API documentation) was located in your DEEP DIVE section. This is probably fine, but maybe you could add "API Documentation" in parentheses next to the "DEEP DIVE" seciton title, or something. Not important for JOSS - just a thought.
