### 2024-06-07 oncoray/mirp#72 @Matthew-Jennings
Happy to consider the JOSS item related to this issue **closed**. Feel free to close the actual issue if you like :)
