### 2024-06-07 8542172
Update README to reflect wording of documentation. #73
