### 2024-06-07 oncoray/mirp#73 @alexzwanenburg
Thanks for the feedback!

> What are your thoughts on this point? I do think that it's not immediately obvious to a prospective user (reading the README, or even the docs intro) what high-level functionality MIRP contains. Example/tutorials are great, but a user will assume this only shows a curated subset of functionality. You do state that it's an IBSI compliant medical image analysis toolkit, but this feels too broad. Probably the closest docs content to what I'm talking about is buried in your Contributing section under [submodules](https://oncoray.github.io/mirp/design.html#submodules)

I have updated the [README](https://github.com/oncoray/mirp/tree/dev2.2.4) to provide a very short overview of what MIRP does, and to show the tutorials more clearly.

> NEW: It took me (embarrassingly?) a little while to see that your technical reference (API documentation) was located in your DEEP DIVE section. This is probably fine, but maybe you could add "API Documentation" in parentheses next to the "DEEP DIVE" seciton title, or something. Not important for JOSS - just a thought.

I renamed the section to *Documentation and API*.
