### 2024-06-15 #6413 @drcandacemakedamoore
I am happy that you addressed the open issue I had, and ready to close it. I would encourage you, while it is not strictly required, to put some badges on your repo (here is an example repo with badges: https://github.com/ReSurfEMG/ReSurfEMG) before JOSS publication . You, or whoever will maintain this software, will always have some technical issues to work on. I see some things that could be further developed, but what you really need to sustain the software is a community of users willing to pitch in and support that. Adding some badges helps people quickly see your software is citable, what version you are on, and so forth. The upcoming publication moment is the moment you may get the most eyeballs on your software. Take the opportunity to make it "look good". Anyways, from my perspective, I am happy to close my issue, and OK moving forward.
