### 2024-06-19 oncoray/mirp#73 @alexzwanenburg
Is it okay if I close this issue?
