### 2024-06-21 #6413 @emdupre
👋  Hi @alexzwanenburg, I've asked editorialbot to generate our post-review checklist, which provides action items for both the authors as well as the editor.

I'll work on getting paper comments to you as soon as possible. In the meantime, could you please address the "additional author tasks" listed [here](https://github.com/openjournals/joss-reviews/issues/6413#issuecomment-2183435920)?
