### 2024-06-22 #6413 @alexzwanenburg
Again, thanks for the excellent peer-reviews. I think it greatly helped making MIRP a better software in all aspects.

A to-do list for me:
- [ ] Double check authors and affiliations (including ORCIDs).
- [ ] Release version 2.2.4 to PyPI
- [ ] Archive the release on Zenodo.
- [ ] Post the DOI of the software release and the release version here.
