+++ paper/paper.md
@@ -9,7 +9,7 @@ authors:
     orcid: 0000-0002-0342-9545
     affiliation: "1, 2"
   - name: Steffen Löck
-    orcid: 0000-0002-0342-9545
+    orcid: 0000-0002-7017-3738
     affiliation: "2, 3, 4"
 affiliations:
   - name: "National Center for Tumor Diseases Dresden (NCT/UCC), Germany: German Cancer Research Center (DKFZ), Heidelberg, Germany; Faculty of Medicine and University Hospital Carl Gustav Carus, TUD Dresden University of Technology, Dresden, Germany; Helmholtz-Zentrum Dresden-Rossendorf (HZDR), Dresden, Germany"
