+++ paper/paper.md
@@ -89,8 +89,8 @@ used alternatives are listed in Table 1.
 
 |                      | `mirp`                                    | `pyradiomics`                                        | `CERR`                                 | `LIFEx`                               | `radiomics`                                          |
 |----------------------|-------------------------------------------|------------------------------------------------------|----------------------------------------|---------------------------------------|------------------------------------------------------| 
-| Version              | 2.1.1                                     | 3.1.0                                                | unknown                                | 7.4.0                                 | unknown                                              |
-| Last updated         | 3/2024                                    | 5/2023                                               | 2/2024                                 | 6/2023                                | 11/2019                                              |
+| Version              | 2.2.4                                     | 3.1.0                                                | unknown                                | 7.6.0                                 | unknown                                              |
+| Last updated         | 6/2024                                    | 5/2023                                               | 2/2024                                 | 4/2024                                | 11/2019                                              |
 | License              | EUPL-1.2                                  | BSD-3                                                | LGPL-2.1                               | custom                                | GPL-3.0                                              |
 | Programming language | Python                                    | Python                                               | MATLAB                                 | Java                                  | MATLAB                                               |
 | IBSI-1 compliant     | yes                                       | partial                                              | yes                                    | yes                                   | no claim                                             |
