### 2024-06-22 #6413 @alexzwanenburg
I fixed a DOI and updated a table with the most recent information.
