+++ docs/_sources/tutorial_compute_radiomics_features_mr.ipynb.txt
@@ -1,6 +1,21 @@
 {
  "cells": [
   {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "2830a672-16e3-4b55-a9b5-d404399e7991",
+   "metadata": {
+    "nbsphinx": "hidden"
+   },
+   "outputs": [],
+   "source": [
+    "import sys\n",
+    "import warnings\n",
+    "sys.path.append(r\"C:\\Users\\alexz\\Documents\\GitHub\\mirp\")\n",
+    "warnings.filterwarnings('ignore')"
+   ]
+  },
+  {
    "cell_type": "markdown",
    "id": "1e447b05-08e1-4326-9709-531acf639a69",
    "metadata": {
@@ -268,7 +283,7 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "INFO\t: MainProcess \t 2024-06-05 11:55:22,262 \t Initialising image extraction using mr images for STS_003.\n"
+      "INFO\t: MainProcess \t 2024-06-20 17:08:52,671 \t Initialising image extraction using mr images for STS_003.\n"
      ]
     }
    ],
@@ -299,7 +314,7 @@
    "outputs": [
     {
      "data": {
[…]
