### 2024-06-23 #6413 @alexzwanenburg
I uploaded the release on Zenodo, with the following DOI: 10.5281/zenodo.12493595
The release version is 2.2.4
