### 2024-07-04 47a1a6f
Use full journal names.
