+++ paper/paper.bib
@@ -4,7 +4,7 @@
   author   = "Huang, Erich P and O'Connor, James P B and McShane, Lisa M and
               Giger, Maryellen L and Lambin, Philippe and Kinahan, Paul E and
               Siegel, Eliot L and Shankar, Lalitha K",
-  journal  = "Nat. Rev. Clin. Oncol.",
+  journal  = "Nature Reviews Clinical Oncology",
   month    =  nov,
   year     =  2022,
   language = "en",
@@ -42,7 +42,7 @@
               Walker-Samuel, Simon and Wason, James and Williams, Kaye J and
               Workman, Paul and Yankeelov, Thomas E and Brindle, Kevin M and
               McShane, Lisa M and Jackson, Alan and Waterton, John C",
-  journal  = "Nat. Rev. Clin. Oncol.",
+  journal  = "Nature Reviews Clinical Oncology",
   volume   =  14,
   number   =  3,
   pages    = "169--186",
@@ -108,7 +108,7 @@
               M{\"o}nnich, David and Zips, Daniel and Krause, Mechthild and
               Baumann, Michael and Troost, Esther G C and L{\"o}ck, Steffen and
               Richter, Christian",
-  journal  = "Sci. Rep.",
+  journal  = "Scientific Reports",
   volume   =  7,
   number   =  1,
   pages    = "13206",
@@ -128,7 +128,7 @@
   author   = "Zwanenburg, Alex and Leger, Stefan and Agolli, Linda and Pilz,
               Karoline and Troost, Esther G C and Richter, Christian and
               L{\"o}ck, Steffen",
-  journal  = "Sci. Rep.",
+  journal  = "Scientific Reports",
   volume   =  9,
   number   =  1,
   pages    = "614",
@@ -150,7 +150,7 @@
                Stephanie E and Diefenhardt, Markus and R{\"o}del, Claus and
                Kirste, Simon and Grosu, Anca-Ligia and Baumann, Michael and
                Krause, Mechthild and Troost, Esther G C and L{\"o}ck, Steffen",
-  journal   = "Sci. Rep.",
+  journal   = "Scientific Reports",
   publisher = "Nature Publishing Group",
   volume    =  12,
   number    =  1,
@@ -224,7 +224,7 @@
               Henry and van Soest, Johan and Lustberg, Tim and Roelofs, Erik
               and van Elmpt, Wouter and Dekker, Andre and Mottaghy, Felix M and
               Wildberger, Joachim E and Walsh, Sean",
-  journal  = "Nat. Rev. Clin. Oncol.",
+  journal  = "Nature Reviews Clinical Oncology",
   volume   =  14,
   number   =  12,
   pages    = "749--762",
@@ -241,7 +241,7 @@
   title    = "Radiomic tractometry reveals tract-specific imaging biomarkers in
               white matter",
   author   = "Neher, Peter and Hirjak, Dusan and Maier-Hein, Klaus",
-  journal  = "Nat. Commun.",
+  journal  = "Nature Communications",
   volume   =  15,
   number   =  1,
   pages    = "303",
@@ -262,7 +262,7 @@
               Chintan and Hosny, Ahmed and Aucoin, Nicole and Narayan, Vivek
               and Beets-Tan, Regina G H and Fillion-Robin, Jean-Christophe and
               Pieper, Steve and Aerts, Hugo J W L",
-  journal  = "Cancer Res.",
+  journal  = "Cancer Research",
   volume   =  77,
   number   =  21,
   pages    = "e104--e107",
@@ -285,7 +285,7 @@
               Pandya, Rutu and van Dijk, Lisanne V and Spezi, Emiliano and
[…]
