+++ mirp/_features/rlm_features.py
@@ -1,21 +1,22 @@
 from functools import cache
 from typing import Generator
 
+import numpy as np
+
 from mirp._images.generic_image import GenericImage
 from mirp._masks.base_mask import BaseMask
-from mirp._features.base_feature import Feature
+from mirp._features.histogram import HistogramDerivedFeature, get_discretisation_parameters
+from mirp._features.texture_matrix import DirectionalMatrix
 from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 
 
 class MatrixRLM(DirectionalMatrix):
 
     def __init__(self):
-        pass
+        super().__init__()
 
 
-
-
-class FeatureRLM(Feature, HistogramBased):
+class FeatureRLM(HistogramDerivedFeature):
 
     def __init__(
             self,
@@ -25,14 +26,57 @@ class FeatureRLM(Feature, HistogramBased):
         super().__init__(**kwargs)
         self.spatial_method = spatial_method.lower()
 
+        # Perform close crop for RLM.
+        self.cropping_distance = 0.0
+
+    def get_matrix(
+            self,
+            image: GenericImage,
+            mask: BaseMask,
+    ) -> list[MatrixRLM]:
+        # First discretise the image.
+        image, mask = self.discretise_image(
+            image=image,
+            mask=mask,
+            discretisation_method=self.discretisation_method,
+            bin_width=self.bin_width,
+            bin_number=self.bin_number,
+            cropping_distance=self.cropping_distance
+        )
+
+        # Then get matrix or matrices
+        matrix_list = self._get_matrix(
+            image=image,
+            mask=mask,
+            spatial_method=self.spatial_method
+        )
+
+        return matrix_list
+
     @staticmethod
     @cache
-    def get_matrix(
+    def _get_matrix(
             image: GenericImage,
             mask: BaseMask,
             spatial_method: str
     ) -> list[MatrixRLM]:
         pass
+
+    def clear_cache(self):
+        super().clear_cache()
+        self._get_matrix.cache_clear()
+
+    def compute(self, image: GenericImage, mask: BaseMask):
+        # Compute or retrieve matrices from cache.
+        matrices = self.get_matrix(image=image, mask=mask)
+
+        # Compute feature value from matrices, and average over matrices.
+        values = [self._compute(matrix=matrix) for matrix in matrices]
+        self.value = np.nanmean(values)
+
+    @staticmethod
+    def _compute(matrix: MatrixRLM):
+        raise NotImplementedError("Implement _compute for feature-specific computation.")
 
 
 class FeatureRLMSRE(FeatureRLM):
@@ -44,8 +88,11 @@ class FeatureRLMSRE(FeatureRLM):
         self.ibsi_id = "22OV"
         self.ibsi_compliant = True
 
-    def compute(self, image: GenericImage, mask: BaseMask):
-        matrices = self.get_matrix()
+    @staticmethod
+    def _compute(matrix: MatrixRLM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.rj / matrix.j ** 2.0) / matrix.n_s
 
 
 def get_rlm_class_dict() -> dict[str, FeatureRLM]:
