### 2024-07-06 #6413 @emdupre
Thank you, @alexzwanenburg ! I'm now happy to recommend _mirp_ to the EiC team for publication in JOSS ✨🚀

Thank you, too, to @surajpaib, @Matthew-Jennings, @drcandacemakedamoore, and @theanega for your reviews and involvement throughout this process ! JOSS works because of your work 💐
