### 2024-07-08 cedcb1b
Move rep utility function from _featuresets.utilities to _features.utilities
