+++ mirp/_features/texture_matrix.py
@@ -1,16 +1,218 @@
+import copy
+
+import pandas as pd
 import numpy as np
+import sys
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
+
+from typing import Generator
+
+from mirp._images.generic_image import GenericImage
+from mirp._masks.base_mask import BaseMask
 
 
 class Matrix(object):
-    def __init__(self):
-        self.matrix: np.ndarray | None = None
+    def __init__(
+            self,
+            spatial_method: str | None,
+            matrix: pd.DataFrame | None = None,
+            slice_id: int | None = None,
+            n_voxels: int | None = None
+    ):
+        # Placeholder for the actual matrix.
+        self.matrix = matrix
+
+        # Spatial method for aggregating and merging matrices.
+        self.spatial_method = spatial_method
+
+        # Slice ID that is required for merging matrices within a slice.
+        self.slice_id = slice_id
+
+        # Number of voxels represented in the matrix.
+        self.n_voxels = n_voxels
+
+    def copy(self):
+        return copy.deepcopy(self)
 
     def is_empty(self):
         return self.matrix is None or len(self.matrix) == 0
 
+    def compute(self, image: GenericImage, mask: BaseMask, **kwargs):
+        raise NotImplementedError("Implement in subclasses.")
+
+    def set_values_from_matrix(self):
+        raise NotImplementedError("Implement in subclasses.")
+
+    def generate(self, prototype, n_slices: int, **kwargs) -> Generator[Self, None, None]:
+        if self.spatial_method in ["2d", "2.5d"]:
+            # Extract texture matrices from 2D slices.
+            for slice_id in np.arange(0, n_slices):
+                yield prototype(
+                    spatial_method=self.spatial_method,
+                    slice_id=slice_id,
+                    **kwargs
+                )
+
+        elif self.spatial_method == "3d":
+            # Extract texture matrix for the whole 3D volume.
+            yield prototype(
+                spatial_method=self.spatial_method,
+                **kwargs
+            )
+
+        else:
+            raise ValueError(
+                f"One of  \"2d\", \"2.5d\" or \"3d\" is expected as spatial method. Found: {self.spatial_method}."
+            )
+
+    def merge(
+            self,
+            matrix_list: list[Self],
+            prototype,
+            **kwargs
+    ) -> list[Self]:
+
+        if self.spatial_method == "2d":
+            # Average features over slices: maintain original 2D texture matrices.
+            return [matrix for matrix in matrix_list if not matrix.is_empty()]
+
+        elif self.spatial_method in ["2.5d", "3d"]:
+            # Merge 2D or 3D texture matrices into a single matrix.
+
+            # Remove all empty matrices.
+            updated_matrix_list = [matrix for matrix in matrix_list if not matrix.is_empty()]
+            if len(updated_matrix_list) == 0:
+                return []
+
+            collected_matrices = [matrix.matrix for matrix in updated_matrix_list]
+            n_voxels = np.sum([matrix.n_voxels for matrix in updated_matrix_list])
+
+            # Merge collected matrices.
+            merged_matrix = pd.concat(collected_matrices, axis=0)
+            merged_matrix = merged_matrix.groupby(by=self._get_grouping_columns()).sum().reset_index()
+
+            return [prototype(
+                spatial_method=self.spatial_method,
+                matrix=merged_matrix,
+                n_voxels=n_voxels,
+                **kwargs
+            )]
+
+        else:
+            raise ValueError(
+                f"One of  \"2d\", \"2.5d\" or \"3d\" is expected as spatial method. Found: {self.spatial_method}."
+            )
+
+    @staticmethod
[…]
