+++ mirp/_features/rlm_features.py
@@ -98,7 +98,6 @@ class FeatureRLM(HistogramDerivedFeature):
 
 
 class FeatureRLMSRE(FeatureRLM):
-
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "RLM - short runs emphasis"
@@ -114,7 +113,6 @@ class FeatureRLMSRE(FeatureRLM):
 
 
 class FeatureRLMLRE(FeatureRLM):
-
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "RLM - long runs emphasis"
@@ -130,7 +128,6 @@ class FeatureRLMLRE(FeatureRLM):
 
 
 class FeatureRLMLGRE(FeatureRLM):
-
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "RLM - low grey level run emphasis"
@@ -142,7 +139,204 @@ class FeatureRLMLGRE(FeatureRLM):
     def _compute(matrix: MatrixRLM) -> float:
         if matrix.is_empty():
             return np.nan
+        return np.sum(matrix.ri.ri / matrix.ri.i ** 2.0) / matrix.n_s
+
+
+class FeatureRLMHGRE(FeatureRLM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "RLM - high grey level run emphasis"
+        self.abbr_name = "rlm_hgre"
+        self.ibsi_id = "G3QZ"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixRLM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.ri.ri * matrix.ri.i ** 2.0) / matrix.n_s
+
+
+class FeatureRLMSRLGE(FeatureRLM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "RLM - short run low grey level emphasis"
+        self.abbr_name = "rlm_srlge"
+        self.ibsi_id = "HTZT"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixRLM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.rij.rij / (matrix.rij.i * matrix.rij.j) ** 2.0) / matrix.n_s
+
+
+class FeatureRLMSRHGE(FeatureRLM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "RLM - short run high grey level emphasis"
+        self.abbr_name = "rlm_srhge"
+        self.ibsi_id = "GD3A"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixRLM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.rij.rij * matrix.rij.i ** 2.0 / matrix.rij.j ** 2.0) / matrix.n_s
+
+
+class FeatureRLMLRLGE(FeatureRLM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "RLM - long run low grey level emphasis"
+        self.abbr_name = "rlm_lrlge"
+        self.ibsi_id = "IVPO"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixRLM) -> float:
+        if matrix.is_empty():
+            return np.nan
[…]
