+++ mirp/_features/texture_matrix.py
@@ -8,11 +8,7 @@ if sys.version_info >= (3, 11):
     from typing import Self
 else:
     from typing_extensions import Self
-
 from typing import Generator
-
-from mirp._images.generic_image import GenericImage
-from mirp._masks.base_mask import BaseMask
 
 
 class Matrix(object):
