### 2024-07-11 28c5715
Matrices are now correctly initialised prior to processing
