### 2024-07-11 2d352e7
Feature order in class_dict is now maintained.
