### 2024-07-11 e2d0a8f
Allow for passing additional values to set_values_from_matrix
