### 2024-07-11 e3ba710
Added _lookup_intensity helper method.
