+++ mirp/_features/cm_matrix.py
@@ -1,0 +1,135 @@
+import copy
+
+import numpy as np
+import pandas as pd
+
+from mirp._features.texture_matrix import DirectionalMatrix
+
+
+class MatrixCM(DirectionalMatrix):
+
+    def __init__(self, distance: int, **kwargs):
+        super().__init__(**kwargs)
+
+        # Set distance
+        self.distance = distance
+
+        # Placeholders for derivative values computed using set_values_from_matrix
+        # Co-occurrence matrix, expressed as probabilities.
+        self.pij: pd.DataFrame | None = None
+
+        # Marginal sum over grey levels of originating pixels i
+        self.pi: pd.DataFrame | None = None
+
+        # Marginal sum over grey levels of target pixels j
+        self.pj: pd.DataFrame | None = None
+
+        # Diagonal probabilities
+        self.pimj: pd.DataFrame | None = None
+
+        # Cross-diagonal probabilities
+        self.pipj: pd.DataFrame | None = None
+
+        # Number of intensities bins (grey levels)
+        self.n_g: int | None = None
+
+    def compute(
+            self,
+            data: pd.DataFrame | None,
+            image_dimension: tuple[int, int, int] | None = None,
+            **kwargs
+    ):
+        # Check if data actually exists
+        if data is None:
+            return
+
+        # Check if the roi contains any masked voxels. If this is not the case, don't construct the GLRLM.
+        if not np.any(data.roi_int_mask):
+            return
+
+        # Create local copies of the image table
+        if self.spatial_method in ["3d_average", "3d_volume_merge"]:
+            data = copy.deepcopy(data)
+
+        elif self.spatial_method in ["2d_average", "2d_slice_merge", "2.5d_direction_merge", "2.5d_volume_merge"]:
+            data = copy.deepcopy(data[data.z == self.slice_id])
+            data["index_id"] = np.arange(0, len(data))
+            data["z"] = 0
+            data = data.reset_index(drop=True)
+
+        else:
+            self._spatial_method_error()
+
+        # Set grey level of voxels outside ROI to NaN
+        data.loc[data.roi_int_mask == False, "g"] = np.nan
+
+        # Determine potential transitions. The distance parameter determines the lookup distance.
+        data["to_index"] = self.coord_to_index(
+            x=data.x.values + self.direction[2] * self.distance,
+            y=data.y.values + self.direction[1] * self.distance,
+            z=data.z.values + self.direction[0] * self.distance,
+            dims=image_dimension
+        )
+
+        # Look up intensity of neighbours.
+        data["to_g"] = self._lookup_intensity(
+            x=data.g.values,
+            index=data.to_index.values
+        )
+
+        # Check if any valid neighbours were found.
+        if np.all(np.isnan(data[["to_g"]])):
+            return
+
+        # Count occurrences intensity combinations.
+        data = data.groupby(by=["g", "to_g"]).size().reset_index(name="n")
+
+        # We assume symmetric GLCM. Hence, append grey level transitions in opposite direction.
+        additional_data = pd.DataFrame({
+            "g": data.to_g,
+            "to_g": data.g,
+            "n": data.n
+        })
+        data = pd.concat([data, additional_data], ignore_index=True)
+
+        # Sum occurrences of grey level transitions after making the matrix symmetric.
+        matrix = data.groupby(by=["g", "to_g"]).sum().reset_index()
+
+        # Rename columns
+        matrix.columns = ["i", "j", "n"]
+
[…]
