+++ mirp/_features/cm_features.py
@@ -1,0 +1,196 @@
+from functools import cache
+from typing import Generator
+
+import numpy as np
+
+from mirp._features.cm_matrix import MatrixCM
+from mirp._images.generic_image import GenericImage
+from mirp._masks.base_mask import BaseMask
+from mirp._features.histogram import get_discretisation_parameters
+from mirp._features.texture_features import FeatureTexture
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
+
+
+class FeatureCM(FeatureTexture):
+
+    def __init__(
+            self,
+            distance: int,
+            **kwargs
+    ):
+        super().__init__(**kwargs)
+
+        # Lookup distance (in voxels)
+        self.distance = distance
+
+        # Features are always computed from symmetric co-occurrence matrices.
+        self.symmetric_matrix = True
+
+        # Perform close crop for CM.
+        self.cropping_distance = 0.0
+
+    def get_matrix(
+            self,
+            image: GenericImage,
+            mask: BaseMask,
+    ) -> list[MatrixCM]:
+        # First discretise the image.
+        image, mask = self.discretise_image(
+            image=image,
+            mask=mask,
+            discretisation_method=self.discretisation_method,
+            bin_width=self.bin_width,
+            bin_number=self.bin_number,
+            cropping_distance=self.cropping_distance
+        )
+
+        # Then get matrix or matrices
+        matrix_list = self._get_matrix(
+            image=image,
+            mask=mask,
+            spatial_method=self.spatial_method,
+            distance=self.distance
+        )
+
+        return matrix_list
+
+    @staticmethod
+    @cache
+    def _get_matrix(
+            image: GenericImage,
+            mask: BaseMask,
+            spatial_method: str,
+            distance: int
+    ) -> list[MatrixCM]:
+        # Represent image and mask as a dataframe.
+        data = mask.as_pandas_dataframe(
+            image=image,
+            intensity_mask=True
+        )
+
+        # Instantiate a helper copy of the current class to be able to use class methods without tying the cache to the
+        # instance of the original object from which this method is called.
+        matrix_instance = MatrixCM(
+            spatial_method=spatial_method,
+            distance=distance
+        )
+
+        # Compute the required matrices.
+        matrix_list = list(matrix_instance.generate(prototype=MatrixCM, n_slices=image.image_dimension[0]))
+        for matrix in matrix_list:
+            matrix.compute(data=data, image_dimension=image.image_dimension)
+
+        # Merge according to the spatial method.
+        matrix_list = matrix_instance.merge(matrix_list, prototype=MatrixCM)
+
+        # Compute additional values from the individual matrices.
+        for matrix in matrix_list:
+            matrix.set_values_from_matrix()
+        print(f"CM Matrix being cached for: {spatial_method}.")
+        return matrix_list
+
+    def clear_cache(self):
+        super().clear_cache()
+        self._get_matrix.cache_clear()
+
+    def compute(self, image: GenericImage, mask: BaseMask):
+        # Compute or retrieve matrices from cache.
+        matrices = self.get_matrix(image=image, mask=mask)
+
+        # Compute feature value from matrices, and average over matrices.
+        values = [self._compute(matrix=matrix) for matrix in matrices]
+        self.value = np.nanmean(values)
+
+    @staticmethod
+    def _compute(matrix: MatrixCM):
+        raise NotImplementedError("Implement _compute for feature-specific computation.")
+
+    def create_table_name(self):
+        table_elements = (
+                self._get_base_table_name_element()
[…]
