### 2024-07-11 1c803ee
CM features can now be passed to the workflow.
