+++ mirp/_features/rlm_features.py
@@ -76,6 +76,7 @@ class FeatureRLM(FeatureTexture):
         # Compute additional values from the individual matrices.
         for matrix in matrix_list:
             matrix.set_values_from_matrix()
+        print(f"RLM Matrix being cached for: {spatial_method}.")
         return matrix_list
 
     def clear_cache(self):
