+++ mirp/_features/feature_generator.py
@@ -4,6 +4,7 @@ import pandas as pd
 
 from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 from mirp._features.base_feature import Feature
+from mirp._features.cm_features import generate_cm_features
 from mirp._features.rlm_features import generate_rlm_features
 
 
@@ -11,6 +12,9 @@ def generate_features(
         settings: FeatureExtractionSettingsClass,
         features: None | list[str] = None
 ) -> Generator[Feature, None, None]:
+
+    # Co-occurrence matrix features.
+    yield from generate_cm_features(settings=settings, features=features)
 
     # Run length matrix features.
     yield from generate_rlm_features(settings=settings, features=features)
