+++ test/ibsi_1_test.py
@@ -101,80 +101,80 @@ def test_ibsi_1_digital_phantom():
 
     data = data[0]
 
-    assert within_tolerance(556, 4, data["morph_volume"])
-    assert within_tolerance(592, 4, data["morph_vol_approx"])
-    assert within_tolerance(388, 3, data["morph_area_mesh"])
-    assert within_tolerance(0.698, 0.004, data["morph_av"])
-    assert within_tolerance(0.0411, 3e-04, data["morph_comp_1"])
-    assert within_tolerance(0.599, 0.004, data["morph_comp_2"])
-    assert within_tolerance(1.19, 0.01, data["morph_sph_dispr"])
-    assert within_tolerance(0.843, 0.005, data["morph_sphericity"])
-    assert within_tolerance(0.186, 0.001, data["morph_asphericity"])
-    assert within_tolerance(0.672, 0.004, data["morph_com"])
-    assert within_tolerance(13.1, 0.1, data["morph_diam"])
-    assert within_tolerance(11.4, 0.1, data["morph_pca_maj_axis"])
-    assert within_tolerance(9.31, 0.06, data["morph_pca_min_axis"])
-    assert within_tolerance(8.54, 0.05, data["morph_pca_least_axis"])
-    assert within_tolerance(0.816, 0.005, data["morph_pca_elongation"])
-    assert within_tolerance(0.749, 0.005, data["morph_pca_flatness"])
-    assert within_tolerance(0.869, 0.005, data["morph_vol_dens_aabb"])
-    assert within_tolerance(0.866, 0.005, data["morph_area_dens_aabb"])
-    assert within_tolerance(1.17, 0.01, data["morph_vol_dens_aee"])
-    assert within_tolerance(1.36, 0.01, data["morph_area_dens_aee"])
-    assert within_tolerance(0.961, 0.006, data["morph_vol_dens_conv_hull"])
-    assert within_tolerance(1.03, 0.01, data["morph_area_dens_conv_hull"])
-    assert within_tolerance(1200, 10, data["morph_integ_int"])
-    assert within_tolerance(0.0397, 3e-04, data["morph_moran_i"])
-    assert within_tolerance(0.974, 0.006, data["morph_geary_c"])
-    assert within_tolerance(2.6, 0, data["loc_peak_loc"])
-    assert within_tolerance(3.1, 0, data["loc_peak_glob"])
-    assert within_tolerance(2.15, 0, data["stat_mean"])
-    assert within_tolerance(3.05, 0, data["stat_var"])
-    assert within_tolerance(1.08, 0, data["stat_skew"])
-    assert within_tolerance(-0.355, 0, data["stat_kurt"])
-    assert within_tolerance(1, 0, data["stat_median"])
-    assert within_tolerance(1, 0, data["stat_min"])
-    assert within_tolerance(1, 0, data["stat_p10"])
-    assert within_tolerance(4, 0, data["stat_p90"])
-    assert within_tolerance(6, 0, data["stat_max"])
-    assert within_tolerance(3, 0, data["stat_iqr"])
-    assert within_tolerance(5, 0, data["stat_range"])
-    assert within_tolerance(1.55, 0, data["stat_mad"])
-    assert within_tolerance(1.11, 0, data["stat_rmad"])
-    assert within_tolerance(1.15, 0, data["stat_medad"])
-    assert within_tolerance(0.812, 0, data["stat_cov"])
[…]
