### 2024-07-11 #6413 @alexzwanenburg
Aside from an overly wide sentence on page 2 the proof looks fine from my perspective.
