### 2024-07-12 e46f9c3
Added remaining CM features.
