+++ mirp/_features/cm_features.py
@@ -236,6 +236,294 @@ class FeatureCMSumAverage(FeatureCM):
         return np.sum(matrix.pipj.k * matrix.pipj.pipj)
 
 
+class FeatureCMSumVariance(FeatureCM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "CM - sum variance"
+        self.abbr_name = "cm_sum_var"
+        self.ibsi_id = "OEEB"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixCM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        mu = np.sum(matrix.pipj.k * matrix.pipj.pipj)
+        return np.sum((matrix.pipj.k - mu) ** 2.0 * matrix.pipj.pipj)
+
+
+class FeatureCMSumEntropy(FeatureCM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "CM - sum entropy"
+        self.abbr_name = "cm_sum_entr"
+        self.ibsi_id = "P6QZ1"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixCM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return -np.sum(matrix.pipj.pipj * np.log2(matrix.pipj.pipj))
+
+
+class FeatureCMAngularSecondMoment(FeatureCM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "CM - angular second moment"
+        self.abbr_name = "cm_energy"
+        self.ibsi_id = "8ZQL"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixCM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.pij.pij ** 2.0)
+
+
+class FeatureCMContrast(FeatureCM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "CM - contrast"
+        self.abbr_name = "cm_contrast"
+        self.ibsi_id = "ACUI"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixCM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum((matrix.pij.i - matrix.pij.j) ** 2.0 * matrix.pij.pij)
+
+
+class FeatureCMDissimilarity(FeatureCM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "CM - dissimilarity"
+        self.abbr_name = "cm_dissimilarity`"
+        self.ibsi_id = "8S9J"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixCM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(np.abs(matrix.pij.i - matrix.pij.j) * matrix.pij.pij)
+
+
+class FeatureCMInverseDifference(FeatureCM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "CM - inverse difference"
+        self.abbr_name = "cm_inv_diff"
+        self.ibsi_id = "IB1Z"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixCM) -> float:
+        if matrix.is_empty():
+            return np.nan
[…]
