+++ mirp/_features/feature_generator.py
@@ -6,6 +6,7 @@ from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 from mirp._features.base_feature import Feature
 from mirp._features.cm_features import generate_cm_features
 from mirp._features.rlm_features import generate_rlm_features
+from mirp._features.szm_features import generate_szm_features
 
 
 def generate_features(
@@ -19,6 +20,9 @@ def generate_features(
     # Run length matrix features.
     yield from generate_rlm_features(settings=settings, features=features)
 
+    # Size zone matrix features.
+    yield from generate_szm_features(settings=settings, features=features)
+
 
 def feature_to_table(features: list[Feature]) -> pd.DataFrame | None:
     if features is None or len(features) == 0:
@@ -28,4 +32,4 @@ def feature_to_table(features: list[Feature]) -> pd.DataFrame | None:
     for feature in features:
         feature.create_table_name()
 
-    return pd.DataFrame(dict([(feature.table_name , [feature.value]) for feature in features]))
+    return pd.DataFrame(dict([(feature.table_name, [feature.value]) for feature in features]))
