### 2024-07-12 5b28413
Change signature of Matrix.compute to facilitate size zone matrices
