+++ mirp/_features/cm_matrix.py
@@ -41,7 +41,7 @@ class MatrixCM(DirectionalMatrix):
 
     def compute(
             self,
-            data: pd.DataFrame | None,
+            data: pd.DataFrame | None = None,
             image_dimension: tuple[int, int, int] | None = None,
             **kwargs
     ):
@@ -52,6 +52,9 @@ class MatrixCM(DirectionalMatrix):
         # Check if the roi contains any masked voxels. If this is not the case, don't construct the GLRLM.
         if not np.any(data.roi_int_mask):
             return
+
+        if image_dimension is None:
+            raise ValueError("image_dimension cannot be None, but may not have been provided in the calling function.")
 
         # Create local copies of the image table
         if self.spatial_method in ["3d_average", "3d_volume_merge"]:
