### 2024-07-12 09c5802
Implemented MatrixSZM.compute method.
