+++ mirp/_features/szm_matrix.py
@@ -2,8 +2,11 @@ import copy
 
 import numpy as np
 import pandas as pd
+import skimage.measure
 
 from mirp._features.texture_matrix import Matrix
+from mirp._images.generic_image import GenericImage
+from mirp._masks.base_mask import BaseMask
 
 
 class MatrixSZM(Matrix):
@@ -26,111 +29,70 @@ class MatrixSZM(Matrix):
 
     def compute(
             self,
-            data: pd.DataFrame | None,
-            image_dimension: tuple[int, int, int] | None = None,
+            image: GenericImage | None = None,
+            mask: BaseMask | None = None,
             **kwargs
     ):
+        from mirp.utilities.utilities import real_ndim
+
+        if image is None:
+            raise ValueError(
+                "image cannot be None, but may not have been provided in the calling function."
+            )
+        if mask is None:
+            raise ValueError(
+                "mask cannot be None, but may not have been provided in the calling function."
+            )
+
         # Check if data actually exists
-        if data is None:
+        if image.is_empty() or mask.roi_intensity.is_empty_mask():
             return
 
-        # Check if the roi contains any masked voxels. If this is not the case, don't construct the GLRLM.
-        if not np.any(data.roi_int_mask):
-            return
+        # Define neighbour directions
+        if self.spatial_method == "3d":
+            connectivity = 3
+            image = copy.deepcopy(image.get_voxel_grid())
+            mask = copy.deepcopy(mask.roi_intensity.get_voxel_grid())
 
-        # Create local copies of the image table
-        if self.spatial_method in ["3d_average", "3d_volume_merge"]:
-            data = copy.deepcopy(data)
-
-        elif self.spatial_method in ["2d_average", "2d_slice_merge", "2.5d_direction_merge", "2.5d_volume_merge"]:
-            data = copy.deepcopy(data[data.z == self.slice_id])
-            data["index_id"] = np.arange(0, len(data))
-            data["z"] = 0
-            data = data.reset_index(drop=True)
-
+        elif self.spatial_method in ["2d", "2.5d"]:
+            connectivity = 2
+            image = image.get_voxel_grid()[self.slice_id, :, :]
+            mask = mask.roi_intensity.get_voxel_grid()[self.slice_id, :, :]
         else:
+            connectivity = -1  # Does nothing, because _spatial_method_error throws an error.
             self._spatial_method_error()
 
-        # Set grey level of voxels outside ROI to NaN
-        data.loc[data.roi_int_mask == False, "g"] = np.nan
+        # Check dimensionality and update connectivity if necessary.
+        connectivity = min([connectivity, real_ndim(image)])
 
-        # Set the number of voxels
-        self.n_voxels = np.sum(data.roi_int_mask.values)
+        # Set voxels outside roi to 0.0
+        image[~mask] = 0.0
 
-        # Determine update index number for direction
-        if ((
-                self.direction[2]
-                + self.direction[1] * image_dimension[2]
-                + self.direction[0] * image_dimension[2] * image_dimension[1]
-        ) >= 0):
-            direction = self.direction
-        else:
-            direction = tuple(-x for x in self.direction)
+        # Count the number of voxels within the roi
+        self.n_voxels = np.sum(mask)
 
-        # Step size
-        index_update = (
-                direction[2]
-                + direction[1] * image_dimension[2]
-                + direction[0] * image_dimension[2] * image_dimension[1]
+        # Label all connected voxels with the same label.
+        labelled_image = skimage.measure.label(
+            label_image=image,
+            background=0.0,
+            connectivity=connectivity
         )
 
-        # Generate information concerning segments
[…]
