### 2024-07-12 8ee4c87
Update MatrixSZM.set_values_from_matrix
