### 2024-07-12 b3b3a74
Re-enable SZM tests.
