+++ test/ibsi_1_test.py
@@ -421,54 +421,54 @@ def test_ibsi_1_digital_phantom():
     assert within_tolerance(3.48, 0, data["rlm_gl_var_3d_v_mrg"])
     assert within_tolerance(0.598, 0, data["rlm_rl_var_3d_v_mrg"])
     assert within_tolerance(2.62, 0, data["rlm_rl_entr_3d_v_mrg"])
-    # assert within_tolerance(0.363, 0, data["szm_sze_2d"])
-    # assert within_tolerance(43.9, 0, data["szm_lze_2d"])
-    # assert within_tolerance(0.371, 0, data["szm_lgze_2d"])
-    # assert within_tolerance(16.4, 0, data["szm_hgze_2d"])
-    # assert within_tolerance(0.0259, 0, data["szm_szlge_2d"])
-    # assert within_tolerance(10.3, 0, data["szm_szhge_2d"])
-    # assert within_tolerance(40.4, 0, data["szm_lzlge_2d"])
-    # assert within_tolerance(113, 0, data["szm_lzhge_2d"])
-    # assert within_tolerance(1.41, 0, data["szm_glnu_2d"])
-    # assert within_tolerance(0.323, 0, data["szm_glnu_norm_2d"])
-    # assert within_tolerance(1.49, 0, data["szm_zsnu_2d"])
-    # assert within_tolerance(0.333, 0, data["szm_zsnu_norm_2d"])
-    # assert within_tolerance(0.24, 0, data["szm_z_perc_2d"])
-    # assert within_tolerance(3.97, 0, data["szm_gl_var_2d"])
-    # assert within_tolerance(21, 0, data["szm_zs_var_2d"])
-    # assert within_tolerance(1.93, 0, data["szm_zs_entr_2d"])
-    # assert within_tolerance(0.368, 0, data["szm_sze_2.5d"])
-    # assert within_tolerance(34.2, 0, data["szm_lze_2.5d"])
-    # assert within_tolerance(0.368, 0, data["szm_lgze_2.5d"])
-    # assert within_tolerance(16.2, 0, data["szm_hgze_2.5d"])
-    # assert within_tolerance(0.0295, 0, data["szm_szlge_2.5d"])
-    # assert within_tolerance(9.87, 0, data["szm_szhge_2.5d"])
-    # assert within_tolerance(30.6, 0, data["szm_lzlge_2.5d"])
-    # assert within_tolerance(107, 0, data["szm_lzhge_2.5d"])
-    # assert within_tolerance(5.44, 0, data["szm_glnu_2.5d"])
-    # assert within_tolerance(0.302, 0, data["szm_glnu_norm_2.5d"])
-    # assert within_tolerance(3.44, 0, data["szm_zsnu_2.5d"])
-    # assert within_tolerance(0.191, 0, data["szm_zsnu_norm_2.5d"])
-    # assert within_tolerance(0.243, 0, data["szm_z_perc_2.5d"])
-    # assert within_tolerance(3.92, 0, data["szm_gl_var_2.5d"])
-    # assert within_tolerance(17.3, 0, data["szm_zs_var_2.5d"])
-    # assert within_tolerance(3.08, 0, data["szm_zs_entr_2.5d"])
-    # assert within_tolerance(0.255, 0, data["szm_sze_3d"])
-    # assert within_tolerance(550, 0, data["szm_lze_3d"])
-    # assert within_tolerance(0.253, 0, data["szm_lgze_3d"])
-    # assert within_tolerance(15.6, 0, data["szm_hgze_3d"])
-    # assert within_tolerance(0.0256, 0, data["szm_szlge_3d"])
-    # assert within_tolerance(2.76, 0, data["szm_szhge_3d"])
[…]
