### 2024-07-12 6b10411
First implementation of SZM features.
