+++ mirp/_features/szm_features.py
@@ -98,26 +98,266 @@ class FeatureSZM(FeatureTexture):
         self.table_name = "_".join(table_elements)
 
 
+class FeatureSZMSmallZoneEmphasis(FeatureSZM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "SZM - small zone emphasis"
+        self.abbr_name = "szm_sze"
+        self.ibsi_id = "5QRC"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixSZM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.sj.sj / matrix.sj.j ** 2.0) / matrix.n_s
+
+
+class FeatureSZMLargeZoneEmphasis(FeatureSZM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "SZM - large zone emphasis"
+        self.abbr_name = "szm_lze"
+        self.ibsi_id = "48P8"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixSZM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.sj.sj * matrix.sj.j ** 2.0) / matrix.n_s
+
+
+class FeatureSZMLowGreyLevelZoneEmphasis(FeatureSZM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "SZM - low grey level zone emphasis"
+        self.abbr_name = "szm_lgze"
+        self.ibsi_id = "XMSY"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixSZM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.si.si / matrix.si.i ** 2.0) / matrix.n_s
+
+
+class FeatureSZMHighGreyLevelZoneEmphasis(FeatureSZM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "SZM - high grey level zone emphasis "
+        self.abbr_name = "szm_hgze"
+        self.ibsi_id = "5GN9"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixSZM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return  np.sum(matrix.si.si * matrix.si.i ** 2.0) / matrix.n_s
+
+
+class FeatureSZMSmallZoneLowGreyLevelEmphasis(FeatureSZM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "SZM - small zone low grey level emphasis"
+        self.abbr_name = "szm_szlge"
+        self.ibsi_id = "5RAI"
+        self.ibsi_compliant = True
+
+    @staticmethod
+    def _compute(matrix: MatrixSZM) -> float:
+        if matrix.is_empty():
+            return np.nan
+        return np.sum(matrix.sij.sij / (matrix.sij.i * matrix.sij.j) ** 2.0) / matrix.n_s
+
+
+class FeatureSZMSmallZoneHighGreyLevelEmphasis(FeatureSZM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.name = "SZM - small zone high grey level emphasis"
+        self.abbr_name = "szm_szhge"
+        self.ibsi_id = "HW1V"
+        self.ibsi_compliant = True
+
[…]
