### 2024-07-12 #6413 @emdupre
@openjournals/bcm-eics, just re-pinging here in case this notification slipped through ! Thank you 🙏
