+++ mirp/_features/dzm_features.py
@@ -1,0 +1,403 @@
+from functools import cache
+from typing import Generator
+
+import numpy as np
+
+from mirp._features.dzm_matrix import MatrixDZM
+from mirp._images.generic_image import GenericImage
+from mirp._masks.base_mask import BaseMask
+from mirp._features.histogram import get_discretisation_parameters
+from mirp._features.texture_features import FeatureTexture
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
+
+
+class FeatureDZM(FeatureTexture):
+
+    def __init__(
+            self,
+            **kwargs
+    ):
+        super().__init__(**kwargs)
+
+        # Perform close crop for DZM.
+        self.cropping_distance = 0.0
+
+    def get_matrix(
+            self,
+            image: GenericImage,
+            mask: BaseMask,
+    ) -> list[MatrixDZM]:
+        # First discretise the image.
+        image, mask = self.discretise_image(
+            image=image,
+            mask=mask,
+            discretisation_method=self.discretisation_method,
+            bin_width=self.bin_width,
+            bin_number=self.bin_number,
+            cropping_distance=self.cropping_distance
+        )
+
+        # Then get matrix or matrices
+        matrix_list = self._get_matrix(
+            image=image,
+            mask=mask,
+            spatial_method=self.spatial_method
+        )
+
+        return matrix_list
+
+    @staticmethod
+    @cache
+    def _get_matrix(
+            image: GenericImage,
+            mask: BaseMask,
+            spatial_method: str
+    ) -> list[MatrixDZM]:
+        # Extract data from the mask and image.
+        data = mask.as_pandas_dataframe(
+            image=image,
+            intensity_mask=True,
+            distance_map=True,
+            by_slice=spatial_method in ["2d", "2.5d"]
+        )
+
+        # Instantiate a helper copy of the current class to be able to use class methods without tying the cache to the
+        # instance of the original object from which this method is called.
+        matrix_instance = MatrixDZM(
+            spatial_method=spatial_method
+        )
+
+        # Compute the required matrices.
+        matrix_list = list(matrix_instance.generate(prototype=MatrixDZM, n_slices=image.image_dimension[0]))
+        for matrix in matrix_list:
+            matrix.compute(data=data, image=image, mask=mask)
+
+        # Merge according to the spatial method.
+        matrix_list = matrix_instance.merge(matrix_list, prototype=MatrixDZM)
+
+        # Compute additional values from the individual matrices.
+        for matrix in matrix_list:
+            matrix.set_values_from_matrix()
+        print(f"DZM Matrix being cached for: {spatial_method}.")
+        return matrix_list
+
+    def clear_cache(self):
+        super().clear_cache()
+        self._get_matrix.cache_clear()
+
+    def compute(self, image: GenericImage, mask: BaseMask):
+        # Compute or retrieve matrices from cache.
+        matrices = self.get_matrix(image=image, mask=mask)
+
+        # Compute feature value from matrices, and average over matrices.
+        values = [self._compute(matrix=matrix) for matrix in matrices]
[…]
