### 2024-07-15 be33a7f
WIP on NGTDM features.
