+++ mirp/_features/ngt_matrix.py
@@ -1,0 +1,172 @@
+import copy
+
+import numpy as np
+import pandas as pd
+
+from mirp._features.texture_matrix import Matrix
+
+
+class MatrixNGTDM(Matrix):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        # Occurrence matrix
+        self.pi: pd.DataFrame | None = None
+
+        # Expanded occurrence matrix
+        self.pij: pd.DataFrame | None = None
+
+        # Number of intensity bins (grey levels)
+        self.n_g: int | None = None
+
+        # Number of intensity bins (grey levels) that are represented
+        self.n_p: int | None = None
+
+    def compute(
+            self,
+            data: pd.DataFrame | None = None,
+            image_dimension: tuple[int, int, int] | None = None,
+            **kwargs
+    ):
+        # Check if data actually exists
+        if data is None:
+            return
+
+        # Check if the roi contains any masked voxels. If this is not the case, don't construct the GLRLM.
+        if not np.any(data.roi_int_mask):
+            return
+
+        if image_dimension is None:
+            raise ValueError("image_dimension cannot be None, but may not have been provided in the calling function.")
+
+        # Create local copies of the image table
+        if self.spatial_method == "3d":
+            data = copy.deepcopy(data)
+
+            directions = list(self._generate_neighbour_direction(
+                d=1,
+                metric="chebyshev",
+                complete=True,
+                dim3=True
+            ))
+
+        elif self.spatial_method in ["2d", "2.5d"]:
+            data = copy.deepcopy(data[data.z == self.slice_id])
+            data["index_id"] = np.arange(0, len(data))
+            data["z"] = 0
+            data = data.reset_index(drop=True)
+
+            directions = list(self._generate_neighbour_direction(
+                d=1,
+                metric="chebyshev",
+                complete=True,
+                dim3=False
+            ))
+
+        else:
+            directions = None
+            self._spatial_method_error()
+
+        # Set grey level of voxels outside ROI to NaN
+        data.loc[data.roi_int_mask == False, "g"] = np.nan
+
+        # Initialise sum of grey levels and number of neighbours.
+        data["g_sum"] = 0
+        data["n_nbrs"] = 0
+
+        for direction in directions:
+            # Determine potential valid neighbours.
+            data["to_index"] = self.coord_to_index(
+                x=data.x.values + direction[2],
+                y=data.y.values + direction[1],
+                z=data.z.values + direction[0],
+                dims=image_dimension
+            )
+
+            # Get grey level value from transitions
+            data["to_g"] = self._lookup_intensity(
+                x=data.g.values,
+                index=data.to_index.values
+            )
+
+            # Determine which voxels have valid neighbours
+            valid_index = np.isfinite(data.to_g)
+
+            # Sum grey level and increase neighbour counter
+            data.loc[valid_index, "g_sum"] += data.loc[valid_index, "to_g"]
+            data.loc[valid_index, "n_nbrs"] += 1
+
+        # Calculate average neighbourhood grey level
+        data["g_nbr_avg"] = data.g_sum / data.n_nbrs
+
+        # Work with voxels without a missing grey value and with a valid neighbourhood
+        data = data[np.logical_and(np.isfinite(data.g_nbr_avg), data.roi_int_mask)]
+
+        # Determine contribution to s per voxel
+        data["s_sub"] = np.abs(data.g - data.g_nbr_avg)
+
+        # Drop superfluous columns
[…]

[…]
