### 2024-07-16 4e40d64
Updated NGTDM to function correctly.
