### 2024-07-16 d6124a3
Implemented NGLDM.
