+++ test/ibsi_1_test.py
@@ -532,57 +532,57 @@ def test_ibsi_1_digital_phantom():
     assert within_tolerance(6.54, 0, data["ngt_busyness_3d"])
     assert within_tolerance(13.5, 0, data["ngt_complexity_3d"])
     assert within_tolerance(0.763, 0, data["ngt_strength_3d"])
-    # assert within_tolerance(0.158, 0, data["ngl_lde_d1_a0.0_2d"])
-    # assert within_tolerance(19.2, 0, data["ngl_hde_d1_a0.0_2d"])
-    # assert within_tolerance(0.702, 0, data["ngl_lgce_d1_a0.0_2d"])
-    # assert within_tolerance(7.49, 0, data["ngl_hgce_d1_a0.0_2d"])
-    # assert within_tolerance(0.0473, 0, data["ngl_ldlge_d1_a0.0_2d"])
-    # assert within_tolerance(3.06, 0, data["ngl_ldhge_d1_a0.0_2d"])
-    # assert within_tolerance(17.6, 0, data["ngl_hdlge_d1_a0.0_2d"])
-    # assert within_tolerance(49.5, 0, data["ngl_hdhge_d1_a0.0_2d"])
-    # assert within_tolerance(10.2, 0, data["ngl_glnu_d1_a0.0_2d"])
-    # assert within_tolerance(0.562, 0, data["ngl_glnu_norm_d1_a0.0_2d"])
-    # assert within_tolerance(3.96, 0, data["ngl_dcnu_d1_a0.0_2d"])
-    # assert within_tolerance(0.212, 0, data["ngl_dcnu_norm_d1_a0.0_2d"])
-    # assert within_tolerance(1, 0, data["ngl_dc_perc_d1_a0.0_2d"])
-    # assert within_tolerance(2.7, 0, data["ngl_gl_var_d1_a0.0_2d"])
-    # assert within_tolerance(2.73, 0, data["ngl_dc_var_d1_a0.0_2d"])
-    # assert within_tolerance(2.71, 0, data["ngl_dc_entr_d1_a0.0_2d"])
-    # assert within_tolerance(0.17, 0, data["ngl_dc_energy_d1_a0.0_2d"])
-    # assert within_tolerance(0.159, 0, data["ngl_lde_d1_a0.0_2.5d"])
-    # assert within_tolerance(18.8, 0, data["ngl_hde_d1_a0.0_2.5d"])
-    # assert within_tolerance(0.693, 0, data["ngl_lgce_d1_a0.0_2.5d"])
-    # assert within_tolerance(7.66, 0, data["ngl_hgce_d1_a0.0_2.5d"])
-    # assert within_tolerance(0.0477, 0, data["ngl_ldlge_d1_a0.0_2.5d"])
-    # assert within_tolerance(3.07, 0, data["ngl_ldhge_d1_a0.0_2.5d"])
-    # assert within_tolerance(17.2, 0, data["ngl_hdlge_d1_a0.0_2.5d"])
-    # assert within_tolerance(50.8, 0, data["ngl_hdhge_d1_a0.0_2.5d"])
-    # assert within_tolerance(37.9, 0, data["ngl_glnu_d1_a0.0_2.5d"])
-    # assert within_tolerance(0.512, 0, data["ngl_glnu_norm_d1_a0.0_2.5d"])
-    # assert within_tolerance(12.4, 0, data["ngl_dcnu_d1_a0.0_2.5d"])
-    # assert within_tolerance(0.167, 0, data["ngl_dcnu_norm_d1_a0.0_2.5d"])
[…]
