+++ test/ibsi_1_test.py
@@ -146,29 +146,29 @@ def test_ibsi_1_digital_phantom():
     # assert within_tolerance(0.6, 0, data["stat_qcod"])
     # assert within_tolerance(567, 0, data["stat_energy"])
     # assert within_tolerance(2.77, 0, data["stat_rms"])
-    # assert within_tolerance(2.15, 0, data["ih_mean"])
-    # assert within_tolerance(3.05, 0, data["ih_var"])
-    # assert within_tolerance(1.08, 0, data["ih_skew"])
-    # assert within_tolerance(-0.355, 0, data["ih_kurt"])
-    # assert within_tolerance(1, 0, data["ih_median"])
-    # assert within_tolerance(1, 0, data["ih_min"])
-    # assert within_tolerance(1, 0, data["ih_p10"])
-    # assert within_tolerance(4, 0, data["ih_p90"])
-    # assert within_tolerance(6, 0, data["ih_max"])
-    # assert within_tolerance(1, 0, data["ih_mode"])
-    # assert within_tolerance(3, 0, data["ih_iqr"])
-    # assert within_tolerance(5, 0, data["ih_range"])
-    # assert within_tolerance(1.55, 0, data["ih_mad"])
-    # assert within_tolerance(1.11, 0, data["ih_rmad"])
-    # assert within_tolerance(1.15, 0, data["ih_medad"])
-    # assert within_tolerance(0.812, 0, data["ih_cov"])
-    # assert within_tolerance(0.6, 0, data["ih_qcod"])
-    # assert within_tolerance(1.27, 0, data["ih_entropy"])
-    # assert within_tolerance(0.512, 0, data["ih_uniformity"])
-    # assert within_tolerance(8, 0, data["ih_max_grad"])
-    # assert within_tolerance(3, 0, data["ih_max_grad_g"])
-    # assert within_tolerance(-50, 0, data["ih_min_grad"])
-    # assert within_tolerance(1, 0, data["ih_min_grad_g"])
+    assert within_tolerance(2.15, 0, data["ih_mean"])
+    assert within_tolerance(3.05, 0, data["ih_var"])
+    assert within_tolerance(1.08, 0, data["ih_skew"])
+    assert within_tolerance(-0.355, 0, data["ih_kurt"])
+    assert within_tolerance(1, 0, data["ih_median"])
+    assert within_tolerance(1, 0, data["ih_min"])
+    assert within_tolerance(1, 0, data["ih_p10"])
+    assert within_tolerance(4, 0, data["ih_p90"])
+    assert within_tolerance(6, 0, data["ih_max"])
+    assert within_tolerance(1, 0, data["ih_mode"])
+    assert within_tolerance(3, 0, data["ih_iqr"])
+    assert within_tolerance(5, 0, data["ih_range"])
+    assert within_tolerance(1.55, 0, data["ih_mad"])
+    assert within_tolerance(1.11, 0, data["ih_rmad"])
+    assert within_tolerance(1.15, 0, data["ih_medad"])
+    assert within_tolerance(0.812, 0, data["ih_cov"])
+    assert within_tolerance(0.6, 0, data["ih_qcod"])
+    assert within_tolerance(1.27, 0, data["ih_entropy"])
+    assert within_tolerance(0.512, 0, data["ih_uniformity"])
+    assert within_tolerance(8, 0, data["ih_max_grad"])
+    assert within_tolerance(3, 0, data["ih_max_grad_g"])
+    assert within_tolerance(-50, 0, data["ih_min_grad"])
+    assert within_tolerance(1, 0, data["ih_min_grad_g"])
     # assert within_tolerance(0.324, 0, data["ivh_v10"])
     # assert within_tolerance(0.0946, 0, data["ivh_v90"])
     # assert within_tolerance(5, 0, data["ivh_i10"])
