+++ mirp/_features/ih_features.py
@@ -1,0 +1,556 @@
+from functools import cache
+from typing import Generator
+
+import numpy as np
+import pandas as pd
+
+from mirp._images.generic_image import GenericImage
+from mirp._masks.base_mask import BaseMask
+from mirp._features.histogram import get_discretisation_parameters, HistogramDerivedFeature
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
+
+
+class DataIH(object):
+
+    def __init__(self):
+        # Raw intensity data
+        self.image: np.ndarray | None = None
+
+        # Histogram data
+        self.histogram: pd.DataFrame | None = None
+
+        # Number of voxels
+        self.n_voxels: int | None = None
+
+        # Number of grey levels
+        self.n_g: int | None = None
+
+        # Mean intensity
+        self.mu: float | None = None
+
+        # Intensity standard deviation
+        self.sigma: float | None = None
+
+    def compute(self, image: GenericImage, mask: BaseMask):
+        # Skip processing if input image and/or roi are missing
+        if image is None:
+            raise ValueError(
+                "image cannot be None, but may not have been provided in the calling function."
+            )
+        if mask is None:
+            raise ValueError(
+                "mask cannot be None, but may not have been provided in the calling function."
+            )
+
+        # Check if data actually exists
+        if image.is_empty() or mask.roi_intensity.is_empty_mask():
+            return
+
+        # Convert to dataframe and remove entries outside the mask.
+        image = mask.as_pandas_dataframe(image=image, intensity_mask=True)
+        image = image[image.roi_int_mask == True]
+        if len(image) == 0:
+            return
+
+        # Set image data
+        self.image = image.g.values
+
+        # Number of voxels
+        self.n_voxels = len(image)
+
+        # Number of grey levels
+        self.n_g = mask.intensity_range[1] - mask.intensity_range[0] + 1
+
+        # Define histogram
+        histogram = image.groupby(by="g").size().reset_index(name="n")
+
+        # Append empty grey levels to histogram
+        levels = np.arange(start=0, stop=self.n_g) + 1
+        missing_level = levels[np.logical_not(np.isin(levels, histogram.g))]
+        n_missing = len(missing_level)
+        if n_missing > 0:
+            histogram = pd.concat([
+                histogram,
+                pd.DataFrame({
+                    "g": missing_level,
+                    "n": np.zeros(n_missing)
+                })
+            ], ignore_index=True)
+
+        # Update histogram by sorting grey levels and adding bin probabilities
+        histogram = histogram.sort_values(by="g")
+        histogram["p"] = histogram.n / self.n_voxels
+
+        # Add gradient data.
+        if len(histogram) > 1:
+            histogram["grad"] = np.gradient(histogram.n)
+        else:
+            histogram["grad"] = 0.0
+
+        self.histogram = histogram
+
+        # Mean intensity
+        self.mu = np.sum(histogram.g * histogram.p)
+
+        # Standard deviation
+        self.sigma = np.sqrt(np.sum((histogram.g - self.mu) ** 2.0 * histogram.p))
+
+    def is_empty(self):
+        return self.image is None or self.histogram is None
+
+
+class FeatureIH(HistogramDerivedFeature):
+
+    def __init__(
+            self,
+            **kwargs
+    ):
[…]
