+++ mirp/_features/stat_features.py
@@ -2,28 +2,21 @@ from functools import cache
 from typing import Generator
 
 import numpy as np
-import pandas as pd
 
 from mirp._images.generic_image import GenericImage
 from mirp._masks.base_mask import BaseMask
-from mirp._features.histogram import get_discretisation_parameters, HistogramDerivedFeature
+from mirp._features.base_feature import Feature
 from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
 
 
-class DataIH(object):
+class DataStatistics(object):
 
     def __init__(self):
         # Raw intensity data
         self.image: np.ndarray | None = None
 
-        # Histogram data
-        self.histogram: pd.DataFrame | None = None
-
         # Number of voxels
         self.n_voxels: int | None = None
-
-        # Number of grey levels
-        self.n_g: int | None = None
 
         # Mean intensity
         self.mu: float | None = None
@@ -58,57 +51,23 @@ class DataIH(object):
         # Number of voxels
         self.n_voxels = len(image)
 
-        # Number of grey levels
-        self.n_g = mask.intensity_range[1] - mask.intensity_range[0] + 1
-
-        # Define histogram
-        histogram = image.groupby(by="g").size().reset_index(name="n")
-
-        # Append empty grey levels to histogram
-        levels = np.arange(start=0, stop=self.n_g) + 1
-        missing_level = levels[np.logical_not(np.isin(levels, histogram.g))]
-        n_missing = len(missing_level)
-        if n_missing > 0:
-            histogram = pd.concat([
-                histogram,
-                pd.DataFrame({
-                    "g": missing_level,
-                    "n": np.zeros(n_missing)
-                })
-            ], ignore_index=True)
-
-        # Update histogram by sorting grey levels and adding bin probabilities
-        histogram = histogram.sort_values(by="g")
-        histogram["p"] = histogram.n / self.n_voxels
-
-        # Add gradient data.
-        if len(histogram) > 1:
-            histogram["grad"] = np.gradient(histogram.n)
-        else:
-            histogram["grad"] = 0.0
-
-        self.histogram = histogram
-
         # Mean intensity
-        self.mu = np.sum(histogram.g * histogram.p)
+        self.mu = np.mean(self.image)
 
         # Standard deviation
-        self.sigma = np.sqrt(np.sum((histogram.g - self.mu) ** 2.0 * histogram.p))
+        self.sigma = np.std(self.image, ddof=0)
 
     def is_empty(self):
-        return self.image is None or self.histogram is None
+        return self.image is None
 
 
-class FeatureIH(HistogramDerivedFeature):
+class FeatureStat(Feature):
 
     def __init__(
             self,
             **kwargs
     ):
         super().__init__(**kwargs)
-
-        # Perform close crop for intensity histograms.
-        self.cropping_distance = 0.0
 
     def clear_cache(self):
         super().clear_cache()
@@ -119,23 +78,13 @@ class FeatureIH(HistogramDerivedFeature):
     def _get_data(
             image: GenericImage,
             mask: BaseMask
-    ) -> DataIH:
-        data = DataIH()
+    ) -> DataStatistics:
+        data = DataStatistics()
         data.compute(image=image, mask=mask)
 
         return data
 
     def compute(self, image: GenericImage, mask: BaseMask):
-        # Discretise images.
[…]
