### 2024-07-17 #6413 @Kevin-Mattheus-Moerman
- [ ] European English is used throughout, except for one occurrence of `standardized` which should read `standardised`.
- [ ] Please edit the Zenodo listed version tag to include the v so it will end up reading `version v2.2.4` (I know it seems odd but the thing after the word version there should be the complete version tag as listed for your GitHub release tag, which has the `v`).
