### 2024-07-17 #6413 @alexzwanenburg
I made changed `standardized` to `standardised` in the paper, and updated the Zenodo listed version to `v2.2.4`.
