+++ test/ibsi_1_test.py
@@ -128,24 +128,24 @@ def test_ibsi_1_digital_phantom():
     # assert within_tolerance(0.974, 0.006, data["morph_geary_c"])
     # assert within_tolerance(2.6, 0, data["loc_peak_loc"])
     # assert within_tolerance(3.1, 0, data["loc_peak_glob"])
-    # assert within_tolerance(2.15, 0, data["stat_mean"])
-    # assert within_tolerance(3.05, 0, data["stat_var"])
-    # assert within_tolerance(1.08, 0, data["stat_skew"])
-    # assert within_tolerance(-0.355, 0, data["stat_kurt"])
-    # assert within_tolerance(1, 0, data["stat_median"])
-    # assert within_tolerance(1, 0, data["stat_min"])
-    # assert within_tolerance(1, 0, data["stat_p10"])
-    # assert within_tolerance(4, 0, data["stat_p90"])
-    # assert within_tolerance(6, 0, data["stat_max"])
-    # assert within_tolerance(3, 0, data["stat_iqr"])
-    # assert within_tolerance(5, 0, data["stat_range"])
-    # assert within_tolerance(1.55, 0, data["stat_mad"])
-    # assert within_tolerance(1.11, 0, data["stat_rmad"])
-    # assert within_tolerance(1.15, 0, data["stat_medad"])
-    # assert within_tolerance(0.812, 0, data["stat_cov"])
-    # assert within_tolerance(0.6, 0, data["stat_qcod"])
-    # assert within_tolerance(567, 0, data["stat_energy"])
-    # assert within_tolerance(2.77, 0, data["stat_rms"])
+    assert within_tolerance(2.15, 0, data["stat_mean"])
+    assert within_tolerance(3.05, 0, data["stat_var"])
+    assert within_tolerance(1.08, 0, data["stat_skew"])
+    assert within_tolerance(-0.355, 0, data["stat_kurt"])
+    assert within_tolerance(1, 0, data["stat_median"])
+    assert within_tolerance(1, 0, data["stat_min"])
+    assert within_tolerance(1, 0, data["stat_p10"])
+    assert within_tolerance(4, 0, data["stat_p90"])
+    assert within_tolerance(6, 0, data["stat_max"])
+    assert within_tolerance(3, 0, data["stat_iqr"])
+    assert within_tolerance(5, 0, data["stat_range"])
+    assert within_tolerance(1.55, 0, data["stat_mad"])
+    assert within_tolerance(1.11, 0, data["stat_rmad"])
+    assert within_tolerance(1.15, 0, data["stat_medad"])
+    assert within_tolerance(0.812, 0, data["stat_cov"])
+    assert within_tolerance(0.6, 0, data["stat_qcod"])
+    assert within_tolerance(567, 0, data["stat_energy"])
+    assert within_tolerance(2.77, 0, data["stat_rms"])
     assert within_tolerance(2.15, 0, data["ih_mean"])
     assert within_tolerance(3.05, 0, data["ih_var"])
     assert within_tolerance(1.08, 0, data["ih_skew"])
