+++ mirp/_features/stat_features.py
@@ -163,8 +163,8 @@ class FeatureStatMedian(FeatureStat):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "Statistics - median"
-        self.abbr_name = "ih_median"
-        self.ibsi_id = "WIFQ"
+        self.abbr_name = "stat_median"
+        self.ibsi_id = "Y12H"
         self.ibsi_compliant = True
 
     @staticmethod
@@ -176,8 +176,8 @@ class FeatureStatMinimum(FeatureStat):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "Statistics - minimum"
-        self.abbr_name = "ih_min"
-        self.ibsi_id = "1PR8"
+        self.abbr_name = "stat_min"
+        self.ibsi_id = "1GSF"
         self.ibsi_compliant = True
 
     @staticmethod
@@ -196,28 +196,18 @@ class FeatureStatPercentile(FeatureStat):
         # Set IBSI identifier. Only the 10th and 90th percentile are explicitly identified.
         ibsi_id = ""
         if percentile == 10.0:
-            ibsi_id = "GPMT"
+            ibsi_id = "QG58"
         elif percentile == 90.0:
-            ibsi_id = "OZ0C"
+            ibsi_id = "8DWT"
         self.ibsi_id = ibsi_id
 
         if percentile.is_integer():
             self.name = f"Statistics - {int(percentile)}th percentile"
-            self.abbr_name = "ih_p" + str(int(percentile))
+            self.abbr_name = "stat_p" + str(int(percentile))
 
         self.ibsi_compliant = True
 
     def compute(self, image: GenericImage, mask: BaseMask):
-        # Discretise images.
-        image, mask = self.discretise_image(
-            image=image,
-            mask=mask,
-            discretisation_method=self.discretisation_method,
-            bin_width=self.bin_width,
-            bin_number=self.bin_number,
-            cropping_distance=self.cropping_distance
-        )
-
         # Get data.
         data = self._get_data(image=image, mask=mask)
 
@@ -232,8 +222,8 @@ class FeatureStatMaximum(FeatureStat):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "Statistics - maximum"
-        self.abbr_name = "ih_max"
-        self.ibsi_id = "3NCY"
+        self.abbr_name = "stat_max"
+        self.ibsi_id = "84IY"
         self.ibsi_compliant = True
 
     @staticmethod
@@ -241,26 +231,12 @@ class FeatureStatMaximum(FeatureStat):
         return np.max(data.image)
 
 
-class FeatureStatMode(FeatureStat):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.name = "Statistics - mode"
-        self.abbr_name = "ih_mode"
-        self.ibsi_id = "AMMC"
-        self.ibsi_compliant = True
-
-    @staticmethod
-    def _compute(data: DataStatistics) -> float:
-        mode_g = data.histogram.loc[data.histogram.n == np.max(data.histogram.n)].g.values
-        return mode_g[np.argmin(np.abs(mode_g - data.mu))]
-
-
 class FeatureStatInterQuartileRange(FeatureStat):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.name = "Statistics - interquartile range"
-        self.abbr_name = "ih_iqr"
-        self.ibsi_id = "WR0O"
+        self.abbr_name = "stat_iqr"
+        self.ibsi_id = "SALO"
[…]
