+++ mirp/_features/morph_3d_features.py
@@ -1,0 +1,224 @@
+from functools import cache
+from typing import Generator
+
+import numpy as np
+
+from mirp._images.generic_image import GenericImage
+from mirp._masks.base_mask import BaseMask
+from mirp._features.base_feature import Feature
+from mirp.settings.feature_parameters import FeatureExtractionSettingsClass
+
+
+class Feature3DMorph(Feature):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def create_table_name(self):
+        table_elements = self._get_base_table_name_element()
+        self.table_name = "_".join(table_elements)
+
+
+class Data3DMesh(object):
+
+    def __init__(self):
+
+        # Mesh faces
+        self.mesh_faces: np.ndarray | None = None
+
+        # Mesh vertices
+        self.mesh_vertices: np.ndarray | None = None
+
+        # Volume
+        self.volume: float | None = None
+
+        # Area
+        self.area: float | None = None
+
+    def compute(self, image: GenericImage, mask: BaseMask):
+        # Skip processing if input image and/or roi are missing
+        if image is None:
+            raise ValueError(
+                "image cannot be None, but may not have been provided in the calling function."
+            )
+        if mask is None:
+            raise ValueError(
+                "mask cannot be None, but may not have been provided in the calling function."
+            )
+
+        # Check if data actually exists
+        if image.is_empty() or mask.roi_morphology.is_empty_mask():
+            return
+
+        from skimage.measure import marching_cubes
+
+        # Get ROI and pad with empty voxels
+        morphology_mask = np.pad(
+            mask.roi_morphology.get_voxel_grid(),
+            pad_width=1,
+            mode="constant",
+            constant_values=0.0
+        )
+
+        # Use marching cubes to generate a mesh grid for the ROI
+        vertices, faces, _, _ = marching_cubes(
+            volume=morphology_mask,
+            level=0.5,
+            spacing=mask.roi_morphology.image_spacing
+        )
+
+        self.mesh_vertices = vertices
+        self.mesh_faces = faces
+
+        # Get vertices for each face
+        vert_a = vertices[faces[:, 0], :]
+        vert_b = vertices[faces[:, 1], :]
+        vert_c = vertices[faces[:, 2], :]
+
+        # noinspection PyUnreachableCode
+        self.volume = np.abs(np.sum(
+            1.0 / 6.0 * np.einsum("ij,ij->i", vert_a, np.cross(vert_b, vert_c, 1, 1))
+        ))
+
+        # noinspection PyUnreachableCode
+        self.area = np.sum(np.sum(np.cross(vert_a, vert_b) ** 2.0, axis=1) ** 0.5) / 2.0
+
+    def is_empty(self):
+        return self.volume is None
+
+
+class Feature3DMesh(Feature3DMorph):
+
+    def __init__(
+            self,
+            **kwargs
+    ):
+        super().__init__(**kwargs)
+
+    def clear_cache(self):
+        super().clear_cache()
+        self._get_data.cache_clear()
+
+    @staticmethod
+    @cache
+    def _get_data(
+            image: GenericImage,
+            mask: BaseMask
+    ) -> Data3DMesh:
+        data = Data3DMesh()
+        data.compute(image=image, mask=mask)
+
+        return data
+
+    @staticmethod
+    def _compute(data: Data3DMesh, image: GenericImage | None = None, mask: BaseMask | None = None):
+        raise NotImplementedError("Implement _compute for feature-specific computation.")
+
[…]
